mmtk::memory_manager

Function alloc_slow

pub fn alloc_slow<VM: VMBinding>(
    mutator: &mut Mutator<VM>,
    size: usize,
    align: usize,
    offset: usize,
    semantics: AllocationSemantics,
) -> Address

Invoke the allocation slow path of alloc. Like alloc, this function may trigger GC and call crate::vm::Collection::block_for_gc or crate::vm::Collection::out_of_memory. The caller needs to be aware of that.

Notes: This is only intended for use when a binding implements the fastpath on the binding side. When the binding handles fast path allocation and the fast path fails, it can use this method for slow path allocation. Calling before exhausting fast path allocaiton buffer will lead to bad performance.

Arguments:

mutator: The mutator to perform this allocation request.

size: The number of bytes required for the object.

align: Required alignment for the object.

offset: Offset associated with the alignment.

semantics: The allocation semantic required for the allocation.