Function mmtk::memory_manager::gc_init

pub fn gc_init<VM: VMBinding>(mmtk: &'static mut MMTK<VM>, heap_size: usize)

Initialize an MMTk instance. A VM should call this method after creating an MMTK instance but before using any of the methods provided in MMTk.

Arguments:

mmtk: A reference to an MMTk instance to initialize.

heap_size: The heap size for the MMTk instance in bytes.