Function mmtk::memory_manager::get_finalized_object

pub fn get_finalized_object<VM: VMBinding>(
    mmtk: &'static MMTK<VM>
) -> Option<<VM::VMReferenceGlue as ReferenceGlue<VM>>::FinalizableType>

Get an object that is ready for finalization. After each GC, if any registered object is not alive, this call will return one of the objects. MMTk will retain the liveness of those objects until they are popped through this call. Once an object is popped, it is the responsibility of the VM to make sure they are properly finalized before reclaimed by the GC. This call is non-blocking, and will return None if no object is ready for finalization.

Arguments:

mmtk: A reference to an MMTk instance.