mmtk::memory_manager
Function get_finalizers_for

pub fn get_finalizers_for<VM: VMBinding>(
    mmtk: &'static MMTK<VM>,
    object: ObjectReference,
) -> Vec<<VM::VMReferenceGlue as ReferenceGlue<VM>>::FinalizableType>

Pop finalizers that were registered and associated with a certain object. The returned objects may or may not be ready for finalization. This is useful for some VMs that may manually execute finalize method for an object.

Arguments:

mmtk: A reference to an MMTk instance.

object: the given object that MMTk will pop its finalizers