mmtk::memory_manager
Function process_bulk

pub fn process_bulk(builder: &mut MMTKBuilder, options: &str) -> bool

Process multiple MMTk run-time options. Returns true if all the options are processed successfully.

Arguments:

mmtk: A reference to an MMTk instance.

options: a string that is key value pairs separated by white spaces, e.g. “threads=1 stress_factor=4096”