mmtk

Module mmtk

MMTk instance.

Structs

MMAPPER

A global Mmapper for mmaping and protection of virtual memory.

MMTK

An MMTk instance. MMTk allows multiple instances to run independently, and each instance gives users a separate heap. Note that multi-instances is not fully supported yet

MMTKBuilder

MMTk builder. This is used to set options and other settings before actually creating an MMTk instance.

VM_MAP

A global VMMap that manages the mapping of spaces to virtual memory ranges.

Statics

SFT_MAP

Functions

mmtk_debug_print_object

A non-mangled function to print object information for debugging purposes. This function can be directly called from a debugger.