Module mmtk::plan::generational

Modules

copying

Generational copying (GenCopy) Plan: generational copying

gc_work

global

immix

Generational immix (GenImmix) Plan: generational immix

Structs

ALLOCATOR_MAPPING

Constants

ACTIVE_BARRIER

Barrier overhead measurement:

FULL_NURSERY_GC

Full heap collection as nursery GC.

GEN_CONSTRAINTS

Constraints for generational plans. Each generational plan should overwrite based on this constant.

RESERVED_ALLOCATORS

Functions

create_gen_space_mapping

new_generational_global_metadata_specs

Create global side metadata specs for generational plans. This will call SideMetadataContext::new_global_specs(). So if a plan calls this, it should not call SideMetadataContext::new_global_specs() again.