mmtk::plan::generational
Module gc_work

Structs

GenNurseryProcessEdges

Process edges for a nursery GC. This type is provided if a generational plan does not use crate::scheduler::gc_work::SFTProcessEdges. If a plan uses SFTProcessEdges, it does not need to use this type.

ProcessModBuf

The modbuf contains a list of objects in mature space(s) that may contain pointers to the nursery space. This work packet scans the recorded objects and forwards pointers if necessary.

ProcessRegionModBuf

The array-copy modbuf contains a list of array slices in mature space(s) that may contain pointers to the nursery space. This work packet forwards and updates each entry in the recorded slices.