Module mmtk::plan::generational::global

Structs

CommonGenPlan

Common implementation for generational plans. Each generational plan should include this type, and forward calls to it where possible.

Traits

GenerationalPlan

This trait includes methods that are specific to generational plans. This trait needs to be object safe.

GenerationalPlanExt

This trait is the extension trait for GenerationalPlan (see Rust’s extension trait pattern). Generally any method should be put to GenerationalPlan if possible while keeping GenerationalPlan object safe. In this case, generic methods will be put to this extension trait.

Functions

is_nursery_gc

Is current GC only collecting objects allocated since last GC? This method can be called with any plan (generational or not). For non generational plans, it will always return false.