Struct mmtk::plan::generational::global::Gen

pub struct Gen<VM: VMBinding> {
    pub nursery: CopySpace<VM>,
    pub common: CommonPlan<VM>,
    pub gc_full_heap: AtomicBool,
    pub next_gc_full_heap: AtomicBool,
}

Common implementation for generational plans. Each generational plan should include this type, and forward calls to it where possible.

Fields
nursery: CopySpace<VM>

The nursery space. Its type depends on the actual plan.

common: CommonPlan<VM>

The common plan.

gc_full_heap: AtomicBool

Is this GC full heap?

next_gc_full_heap: AtomicBool

Is next GC full heap?

Implementations

impl<VM: VMBinding> Gen<VM>

pub fn new(
heap: HeapMeta,
global_metadata_specs: Vec<SideMetadataSpec>,
constraints: &'static PlanConstraints,
vm_map: &'static Map64,
mmapper: &'static FragmentedMapper,
options: Arc<UnsafeOptionsWrapper>
) -> Self

pub fn verify_side_metadata_sanity(&self, sanity: &mut SideMetadataSanity)

Verify side metadata specs used in the spaces in Gen.

pub fn gc_init(
&mut self,
heap_size: usize,
vm_map: &'static Map64,
scheduler: &Arc<GCWorkScheduler<VM>>
)

Initialize Gen. This should be called by the gc_init() API call.

pub fn prepare(&mut self, tls: VMWorkerThread)

Prepare Gen. This should be called by a single thread in GC prepare work.

pub fn release(&mut self, tls: VMWorkerThread)

Release Gen. This should be called by a single thread in GC release work.

pub fn collection_required<P: Plan>(
&self,
plan: &P,
space_full: bool,
space: &dyn Space<VM>
) -> bool

Check if we need a GC based on the nursery space usage. This method may mark the following GC as a full heap GC.

pub fn force_full_heap_collection(&self)

pub fn last_collection_full_heap(&self) -> bool

pub fn request_full_heap_collection(
&self,
total_pages: usize,
reserved_pages: usize
) -> bool

Check if we should do a full heap GC. It returns true if we should have a full heap GC. It also sets gc_full_heap based on the result.

pub fn trace_object_full_heap<T: TransitiveClosure, C: CopyContext + GCWorkerLocal>(
&self,
trace: &mut T,
object: ObjectReference,
copy_context: &mut C
) -> ObjectReference

Trace objects for spaces in generational and common plans for a full heap GC.

pub fn trace_object_nursery<T: TransitiveClosure, C: CopyContext + GCWorkerLocal>(
&self,
trace: &mut T,
object: ObjectReference,
copy_context: &mut C
) -> ObjectReference

Trace objects for spaces in generational and common plans for a nursery GC.

pub fn is_current_gc_nursery(&self) -> bool

Is the current GC a nursery GC?

pub fn should_next_gc_be_full_heap(plan: &dyn Plan<VM = VM>) -> bool

Check a plan to see if the next GC should be a full heap GC.

pub fn set_next_gc_full_heap(&self, next_gc_full_heap: bool)

Set next_gc_full_heap to the given value.

pub fn get_collection_reserve(&self) -> usize

Get pages reserved for the collection by a generational plan. A generational plan should add their own reservatioin with the value returned by this method.

pub fn get_pages_used(&self) -> usize

Get pages used by a generational plan. A generational plan should add their own used pages with the value returned by this method.