mmtk::plan::generational::global

Trait GenerationalPlanExt

pub trait GenerationalPlanExt<VM: VMBinding>: GenerationalPlan<VM = VM> {
    // Required method
    fn trace_object_nursery<Q: ObjectQueue, const KIND: u8>(
        &self,
        queue: &mut Q,
        object: ObjectReference,
        worker: &mut GCWorker<VM>,
    ) -> ObjectReference;
}

This trait is the extension trait for GenerationalPlan (see Rust’s extension trait pattern). Generally any method should be put to GenerationalPlan if possible while keeping GenerationalPlan object safe. In this case, generic methods will be put to this extension trait.

Required Methods

fn trace_object_nursery<Q: ObjectQueue, const KIND: u8>( &self, queue: &mut Q, object: ObjectReference, worker: &mut GCWorker<VM>, ) -> ObjectReference

Trace an object in nursery collection. If the object is in nursery, we should call trace_object on the space. Otherwise, we can just return the object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.