Struct mmtk::plan::generational::immix::global::GenImmix

pub struct GenImmix<VM: VMBinding> {
    pub gen: CommonGenPlan<VM>,
    pub immix_space: ImmixSpace<VM>,
    pub last_gc_was_defrag: AtomicBool,
    pub last_gc_was_full_heap: AtomicBool,
}

Generational immix. This implements the functionality of a two-generation copying collector where the higher generation is an immix space. See the PLDI’08 paper by Blackburn and McKinley for a description of the algorithm: http://doi.acm.org/10.1145/1375581.1375586.

Fields
gen: CommonGenPlan<VM>

Generational plan, which includes a nursery space and operations related with nursery.

immix_space: ImmixSpace<VM>

An immix space as the mature space.

last_gc_was_defrag: AtomicBool

Whether the last GC was a defrag GC for the immix space.

last_gc_was_full_heap: AtomicBool

Whether the last GC was a full heap GC

Implementations

impl<VM: VMBinding> GenImmix<VM>

pub fn new(args: CreateGeneralPlanArgs<VM>) -> Self

fn requires_full_heap_collection(&self) -> bool