Struct mmtk::plan::global::BasePlan

pub struct BasePlan<VM: VMBinding> {

    pub initialized: AtomicBool,
    pub trigger_gc_when_heap_is_full: AtomicBool,
    pub gc_status: Mutex<GcStatus>,
    pub last_stress_pages: AtomicUsize,
    pub emergency_collection: AtomicBool,
    pub user_triggered_collection: AtomicBool,
    pub internal_triggered_collection: AtomicBool,
    pub last_internal_triggered_collection: AtomicBool,
    pub allocation_success: AtomicBool,
    pub max_collection_attempts: AtomicUsize,
    pub cur_collection_attempts: AtomicUsize,
    pub gc_requester: Arc<GCRequester<VM>>,
    pub stats: Stats,
    mmapper: &'static FragmentedMapper,
    pub vm_map: &'static Map64,
    pub options: Arc<UnsafeOptionsWrapper>,
    pub heap: HeapMeta,
    pub inside_sanity: AtomicBool,
    scanned_stacks: AtomicUsize,
    stacks_prepared: AtomicBool,
    pub mutator_iterator_lock: Mutex<()>,
    allocation_bytes: AtomicUsize,
    pub analysis_manager: AnalysisManager<VM>,
    pub code_space: ImmortalSpace<VM>,
    pub code_lo_space: ImmortalSpace<VM>,
    pub ro_space: ImmortalSpace<VM>,
    pub vm_space: ImmortalSpace<VM>,

}

BasePlan should contain all plan-related state and functions that are fundamental to all plans. These include VM-specific (but not plan-specific) features such as a code space or vm space, which are fundamental to all plans for a given VM. Features that are common to many (but not intrinsically all) plans should instead be included in CommonPlan.

Fields
initialized: AtomicBool

Whether MMTk is now ready for collection. This is set to true when initialize_collection() is called.

trigger_gc_when_heap_is_full: AtomicBool

Should we trigger a GC when the heap is full? It seems this should always be true. However, we allow bindings to temporarily disable GC, at which point, we do not trigger GC even if the heap is full.

gc_status: Mutex<GcStatus>last_stress_pages: AtomicUsizeemergency_collection: AtomicBooluser_triggered_collection: AtomicBoolinternal_triggered_collection: AtomicBoollast_internal_triggered_collection: AtomicBoolallocation_success: AtomicBoolmax_collection_attempts: AtomicUsizecur_collection_attempts: AtomicUsizegc_requester: Arc<GCRequester<VM>>stats: Statsmmapper: &'static FragmentedMappervm_map: &'static Map64options: Arc<UnsafeOptionsWrapper>heap: HeapMetainside_sanity: AtomicBoolscanned_stacks: AtomicUsize

A counter for per-mutator stack scanning

stacks_prepared: AtomicBool

Have we scanned all the stacks?

mutator_iterator_lock: Mutex<()>allocation_bytes: AtomicUsizeanalysis_manager: AnalysisManager<VM>code_space: ImmortalSpace<VM>code_lo_space: ImmortalSpace<VM>ro_space: ImmortalSpace<VM>vm_space: ImmortalSpace<VM>
Implementations

impl<VM: VMBinding> BasePlan<VM>

pub fn new(
vm_map: &'static Map64,
mmapper: &'static FragmentedMapper,
options: Arc<UnsafeOptionsWrapper>,
heap: HeapMeta,
constraints: &'static PlanConstraints,
global_side_metadata_specs: Vec<SideMetadataSpec>
) -> BasePlan<VM>

pub fn gc_init(&mut self, heap_size: usize, vm_map: &'static Map64)

pub fn handle_user_collection_request(&self, tls: VMMutatorThread, force: bool)

The application code has requested a collection.

pub fn trigger_internal_collection_request(&self)

MMTK has requested stop-the-world activity (e.g., stw within a concurrent gc).

pub fn reset_collection_trigger(&self)

Reset collection state information.

pub fn get_used_pages(&self) -> usize

pub fn trace_object<T: TransitiveClosure>(
&self,
_trace: &mut T,
_object: ObjectReference
) -> ObjectReference

pub fn prepare(&mut self, _tls: VMWorkerThread, _full_heap: bool)

pub fn release(&mut self, _tls: VMWorkerThread, _full_heap: bool)

pub fn set_collection_kind<P: Plan>(&self, plan: &P)

pub fn set_gc_status(&self, s: GcStatus)

pub fn stacks_prepared(&self) -> bool

Are the stacks scanned?

pub fn prepare_for_stack_scanning(&self)

Prepare for stack scanning. This is usually used with inform_stack_scanned(). This should be called before doing stack scanning.

pub fn inform_stack_scanned(&self, n_mutators: usize) -> bool

Inform that 1 stack has been scanned. The argument n_mutators indicates the total stacks we should scan. This method returns true if the number of scanned stacks equals the total mutator count. Otherwise it returns false. This method is thread safe and we guarantee only one thread will return true.

pub fn gc_in_progress(&self) -> bool

pub fn gc_in_progress_proper(&self) -> bool

fn determine_collection_attempts(&self) -> usize

pub fn is_user_triggered_collection(&self) -> bool

Return true if this collection was triggered by application code.

pub fn is_internal_triggered_collection(&self) -> bool

Return true if this collection was triggered internally.

pub fn increase_allocation_bytes_by(&self, size: usize) -> usize

Increase the allocation bytes and return the current allocation bytes after increasing

pub fn is_stress_test_gc_enabled(&self) -> bool

Check if the options are set for stress GC. If either stress_factor or analysis_factor is set, we should do stress GC.

pub fn is_precise_stress(&self) -> bool

Check if we should do precise stress test. If so, we need to check for stress GCs for every allocation. Otherwise, we only check in the allocation slow path.

pub fn should_do_stress_gc(&self) -> bool

Check if we should do a stress GC now. If GC is initialized and the allocation bytes exceeds the stress factor, we should do a stress GC.

pub(super) fn collection_required<P: Plan>(
&self,
plan: &P,
space_full: bool,
_space: &dyn Space<VM>
) -> bool

pub(crate) fn verify_side_metadata_sanity(
&self,
side_metadata_sanity_checker: &mut SideMetadataSanity
)