Struct mmtk::plan::global::CommonPlan

pub struct CommonPlan<VM: VMBinding> {
    pub immortal: ImmortalSpace<VM>,
    pub los: LargeObjectSpace<VM>,
    pub base: BasePlan<VM>,
}

CommonPlan is for representing state and features used by many plans, but that are not fundamental to all plans. Examples include the Large Object Space and an Immortal space. Features that are fundamental to all plans must be included in BasePlan.

Fields
immortal: ImmortalSpace<VM>los: LargeObjectSpace<VM>base: BasePlan<VM>
Implementations

impl<VM: VMBinding> CommonPlan<VM>

pub fn new(
vm_map: &'static Map64,
mmapper: &'static FragmentedMapper,
options: Arc<UnsafeOptionsWrapper>,
heap: HeapMeta,
constraints: &'static PlanConstraints,
global_side_metadata_specs: Vec<SideMetadataSpec>
) -> CommonPlan<VM>

pub fn gc_init(
&mut self,
heap_size: usize,
vm_map: &'static Map64,
scheduler: &Arc<GCWorkScheduler<VM>>
)

pub fn get_pages_used(&self) -> usize

pub fn trace_object<T: TransitiveClosure, C: CopyContext>(
&self,
trace: &mut T,
object: ObjectReference
) -> ObjectReference

pub fn prepare(&mut self, tls: VMWorkerThread, full_heap: bool)

pub fn release(&mut self, tls: VMWorkerThread, full_heap: bool)

pub fn stacks_prepared(&self) -> bool

pub fn get_immortal(&self) -> &ImmortalSpace<VM>

pub fn get_los(&self) -> &LargeObjectSpace<VM>

pub(crate) fn verify_side_metadata_sanity(
&self,
side_metadata_sanity_checker: &mut SideMetadataSanity
)