mmtk::plan::lxr::gc_work::rc

Struct ProcessIncs

pub struct ProcessIncs<VM: VMBinding, const KIND: u8> {
    incs: Vec<VM::VMSlot>,
    new_incs: VectorQueue<VM::VMSlot>,
    new_incs_count: u32,
    pause: Pause,
    in_cm: bool,
    no_evac: bool,
    pub root_kind: Option<RootKind>,
    depth: u32,
    lxr: &'static LXR<VM>,
    rc: RefCountHelper<VM>,
    survival_ratio_predictor_local: SurvivalRatioPredictorLocal,
}

Fields
incs: Vec<VM::VMSlot>

Increments to process

new_incs: VectorQueue<VM::VMSlot>

Recursively generated new increments

new_incs_count: u32pause: Pausein_cm: boolno_evac: boolroot_kind: Option<RootKind>depth: u32lxr: &'static LXR<VM>rc: RefCountHelper<VM>survival_ratio_predictor_local: SurvivalRatioPredictorLocal
Implementations

impl<VM: VMBinding, const KIND: u8> ProcessIncs<VM, KIND>

const CAPACITY: usize = 1_024usize

const UNLOG_BITS: SideMetadataSpec

fn __default(lxr: &'static LXR<VM>) -> Self

fn add_new_slot(&mut self, worker: &mut GCWorker<VM>, s: VM::VMSlot)

pub fn new(incs: Vec<VM::VMSlot>, lxr: &'static LXR<VM>) -> Self

fn promote( &mut self, worker: &mut GCWorker<VM>, o: ObjectReference, copied: bool, los: bool, depth: u32, )

fn record_mature_evac_remset2( &mut self, slot_in_defrag: bool, s: VM::VMSlot, o: ObjectReference, )

fn record_mature_evac_remset(&mut self, s: VM::VMSlot, o: ObjectReference)

fn scan_nursery_object( &mut self, worker: &mut GCWorker<VM>, o: ObjectReference, los: bool, in_place_promotion: bool, _depth: u32, size: usize, )

fn flush(&mut self, worker: &mut GCWorker<VM>)

fn inc(&self, o: ObjectReference) -> bool

Return true if the object’s ref count is incremented and the count was zero before the increment

fn dont_evacuate(&self, o: ObjectReference, los: bool) -> bool

fn process_inc_and_evacuate( &mut self, worker: &mut GCWorker<VM>, o: ObjectReference, depth: u32, ) -> ObjectReference

fn unlog_and_load_rc_object<const K: u8>( &mut self, s: VM::VMSlot, ) -> Option<ObjectReference>

Return None if the increment of the slot should be delayed

fn process_slot<const K: u8>( &mut self, worker: &mut GCWorker<VM>, s: VM::VMSlot, depth: u32, add_root_to_remset: bool, ) -> Option<ObjectReference>

fn process_incs<const K: u8>( &mut self, worker: &mut GCWorker<VM>, incs: AddressBuffer<'_, VM::VMSlot>, depth: u32, add_root_to_remset: bool, ) -> Option<Vec<ObjectReference>>