mmtk::plan::mutator_context
Function create_space_mapping

pub(crate) fn create_space_mapping<VM: VMBinding>(
    reserved: ReservedAllocators,
    include_common_plan: bool,
    plan: &'static dyn Plan<VM = VM>,
) -> Vec<(AllocatorSelector, &'static dyn Space<VM>)>

Create a space mapping for spaces in Common/BasePlan for a plan. A plan should reserve its own allocators.

Arguments

reserved: the number of reserved allocators for the plan specific policies.

include_common_plan: whether the plan uses common plan. If a plan uses CommonPlan, we will initialize allocator mapping for spaces in CommonPlan.

plan: the reference to the plan.