Struct mmtk::plan::mutator_context::Mutator

#[repr(C)]
pub struct Mutator<VM: VMBinding> {
    pub(crate) allocators: Allocators<VM>,
    pub barrier: Box<dyn Barrier<VM>>,
    pub mutator_tls: VMMutatorThread,
    pub(crate) plan: &'static dyn Plan<VM = VM>,
    pub(crate) config: MutatorConfig<VM>,
}

A mutator is a per-thread data structure that manages allocations and barriers. It is usually highly coupled with the language VM. It is recommended for MMTk users 1) to have a mutator struct of the same layout in the thread local storage that can be accessed efficiently, and 2) to implement fastpath allocation and barriers for the mutator in the VM side.

Fields
allocators: Allocators<VM>barrier: Box<dyn Barrier<VM>>

Holds some thread-local states for the barrier.

mutator_tls: VMMutatorThread

The mutator thread that is bound with this Mutator struct.

plan: &'static dyn Plan<VM = VM>config: MutatorConfig<VM>
Implementations

impl<VM: VMBinding> Mutator<VM>

fn get_all_allocator_selectors(&self) -> Vec<AllocatorSelector>

Get all the valid allocator selector (no duplicate)

pub fn on_destroy(&mut self)

Inform each allocator about destroying. Call allocator-specific on destroy methods.

pub unsafe fn allocator( &self, selector: AllocatorSelector ) -> &dyn Allocator<VM>

Get the allocator for the selector.

Safety

The selector needs to be valid, and points to an allocator that has been initialized. crate::memory_manager::get_allocator_mapping can be used to get a selector.

pub unsafe fn allocator_mut( &mut self, selector: AllocatorSelector ) -> &mut dyn Allocator<VM>

Get the mutable allocator for the selector.

Safety

The selector needs to be valid, and points to an allocator that has been initialized. crate::memory_manager::get_allocator_mapping can be used to get a selector.

pub unsafe fn allocator_impl<T: Allocator<VM>>( &self, selector: AllocatorSelector ) -> &T

Get the allocator of a concrete type for the selector.

Safety

The selector needs to be valid, and points to an allocator that has been initialized. crate::memory_manager::get_allocator_mapping can be used to get a selector.

pub unsafe fn allocator_impl_mut<T: Allocator<VM>>( &mut self, selector: AllocatorSelector ) -> &mut T

Get the mutable allocator of a concrete type for the selector.

Safety

The selector needs to be valid, and points to an allocator that has been initialized. crate::memory_manager::get_allocator_mapping can be used to get a selector.

pub fn get_allocator_base_offset(selector: AllocatorSelector) -> usize

Return the base offset from a mutator pointer to the allocator specified by the selector.