mmtk::plan::mutator_context
Struct ReservedAllocators

pub(crate) struct ReservedAllocators {
    pub n_bump_pointer: u8,
    pub n_large_object: u8,
    pub n_malloc: u8,
    pub n_immix: u8,
    pub n_mark_compact: u8,
    pub n_free_list: u8,
}

This is used for plans to indicate the number of allocators reserved for the plan. This is used as a parameter for creating allocator/space mapping. A plan is required to reserve the first few allocators. For example, if n_bump_pointer is 1, it means the first bump pointer allocator will be reserved for the plan (and the plan should initialize its mapping itself), and the spaces in common/base plan will use the following bump pointer allocators.

Fields
n_bump_pointer: u8n_large_object: u8n_malloc: u8n_immix: u8n_mark_compact: u8n_free_list: u8
Implementations

impl ReservedAllocators

pub const DEFAULT: Self = _

fn validate(&self)

check if the number of each allocator is okay. Panics if any allocator exceeds the max number.

fn add_bump_pointer_allocator(&mut self) -> AllocatorSelector

fn add_large_object_allocator(&mut self) -> AllocatorSelector

fn add_malloc_allocator(&mut self) -> AllocatorSelector

fn add_immix_allocator(&mut self) -> AllocatorSelector

fn add_mark_compact_allocator(&mut self) -> AllocatorSelector

fn add_free_list_allocator(&mut self) -> AllocatorSelector