Struct mmtk::plan::pageprotect::gc_work::PPProcessEdges

pub struct PPProcessEdges<VM: VMBinding> {
    plan: &'static PageProtect<VM>,
    base: ProcessEdgesBase<PPProcessEdges<VM>>,
}

Edge scanning work packet.

Fields
plan: &'static PageProtect<VM>

Use a static ref to the specific plan to avoid overhead from dynamic dispatch or downcast for each traced object.

base: ProcessEdgesBase<PPProcessEdges<VM>>