Struct mmtk::plan::Mutator

#[repr(C)]
pub struct Mutator<VM: VMBinding> {
    pub allocators: Allocators<VM>,
    pub barrier: Box<dyn Barrier<VM>>,
    pub mutator_tls: VMMutatorThread,
    pub plan: &'static dyn Plan<VM = VM>,
    pub config: MutatorConfig<VM>,
}

A mutator is a per-thread data structure that manages allocations and barriers. It is usually highly coupled with the language VM. It is recommended for MMTk users 1) to have a mutator struct of the same layout in the thread local storage that can be accessed efficiently, and 2) to implement fastpath allocation and barriers for the mutator in the VM side.

Fields
allocators: Allocators<VM>barrier: Box<dyn Barrier<VM>>mutator_tls: VMMutatorThread

The mutator thread that is bound with this Mutator struct.

plan: &'static dyn Plan<VM = VM>config: MutatorConfig<VM>
Implementations

impl<VM: VMBinding> Mutator<VM>

fn get_all_allocator_selectors(&self) -> Vec<AllocatorSelector>

Get all the valid allocator selector (no duplicate)

pub fn on_destroy(&mut self)

Inform each allocator about destroying. Call allocator-specific on destroy methods.