Struct mmtk::plan::ObjectsClosure

pub struct ObjectsClosure<'a, E: ProcessEdgesWork> {
    buffer: VectorQueue<<<E as ProcessEdgesWork>::VM as VMBinding>::VMEdge>,
    pub(crate) worker: &'a mut GCWorker<E::VM>,
    bucket: WorkBucketStage,
}

A transitive closure visitor to collect the edges from objects. It maintains a buffer for the edges, and flushes edges to a new work packet if the buffer is full or if the type gets dropped.

Fields
buffer: VectorQueue<<<E as ProcessEdgesWork>::VM as VMBinding>::VMEdge>worker: &'a mut GCWorker<E::VM>bucket: WorkBucketStage
Implementations

impl<'a, E: ProcessEdgesWork> ObjectsClosure<'a, E>

pub fn new(worker: &'a mut GCWorker<E::VM>, bucket: WorkBucketStage) -> Self

Create an ObjectsClosure.

Arguments:

worker: the current worker. The objects closure should not leave the context of this worker.

bucket: new work generated will be push ed to the bucket.

fn flush(&mut self)