Struct mmtk::plan::VectorQueue

pub struct VectorQueue<T> {
    buffer: Vec<T>,
}

An implementation of ObjectQueue using a Vec.

This can also be used as a buffer. For example, the mark stack or the write barrier mod-buffer.

Fields
buffer: Vec<T>

Enqueued nodes.

Implementations

impl<T> VectorQueue<T>

const CAPACITY: usize = 4_096usize

Reserve a capacity of this on first enqueue to avoid frequent resizing.

pub fn new() -> Self

Create an empty VectorObjectQueue.

pub fn is_empty(&self) -> bool

Return true if the queue is empty.

pub fn take(&mut self) -> Vec<T>

Return the contents of the underlying vector. It will empty the queue.

pub fn into_vec(self) -> Vec<T>

Consume this VectorObjectQueue and return its underlying vector.

pub fn is_full(&self) -> bool

Check if the buffer size reaches CAPACITY.

pub fn push(&mut self, v: T)