mmtk::plan::tracing

Module gc_work

Modules

closure 🔒

root 🔒

weakref 🔒

Structs

DefaultObjectTracer 🔒

This implementation of ObjectTracer queues newly visited objects and create the ProcessNodes work packets to scan and trace objects.

DefaultObjectTracerContext 🔒

This implementation of ObjectTracerContext creates the DefaultObjectTracer to expand the transitive closure during a stop-the-world tracing GC or the final mark pause of a concurrent GC. It is used during object scanning as well as weak reference processing.