Module mmtk::policy::immix

Re-exports

pub use immixspace::*;

Modules

block

defrag

immixspace

line

Macros

validate 🔒

Constants

BLOCK_ONLY

Mark/sweep memory for block-level only

DEFRAG

Do we allow Immix to do defragmentation?

DEFRAG_EVERY_BLOCK

Mark every allocated block as defragmentation source before GC. (for debugging) Set both this and STRESS_DEFRAG to true to make Immix move as many objects as possible.

MARK_LINE_AT_SCAN_TIME

Mark lines when scanning objects. Otherwise, do it at mark time.

MAX_IMMIX_OBJECT_SIZE

The max object size for immix: half of a block

NEVER_MOVE_OBJECTS

In some cases/settings, Immix may never move objects. Currently we only have two cases where we move objects: 1. defrag, 2. nursery copy. If we do neither, we will not move objects. If we have other reasons to move objects, we need to add them here.

PREFER_COPY_ON_NURSERY_GC

If Immix is used as a nursery space, do we prefer copy?

STRESS_DEFRAG

Make every GC a defragment GC. (for debugging)

Functions

validate_features 🔒