mmtk::policy::immix

Module immixspace

Structs

ClearVOBitsAfterPrepare 🔒

A work packet to clear VO bit metadata after Prepare.

FlushPageResource 🔒

Count number of remaining work pacets, and flush page resource if all packets are finished.

ImmixCopyContext

Normal immix copy context. It has one copying Immix allocator. Most immix plans use this copy context.

ImmixHybridCopyContext

Hybrid Immix copy context. It includes two different immix allocators. One with copy = true is used for defrag GCs, and the other is used for other purposes (such as promoting objects from nursery to Immix mature space). This is used by generational immix.

ImmixSpace

ImmixSpaceArgs

Some arguments for Immix Space.

PrepareBlockState

A work packet to prepare each block for a major GC. Performs the action on a range of chunks.

SweepChunk 🔒

Chunk sweeping work packet.

Enums

VOBitsClearingScope 🔒

Constants

TRACE_KIND_DEFRAG 🔒

TRACE_KIND_FAST 🔒