mmtk::policy::lockfreeimmortalspace
Struct LockFreeImmortalSpace

pub struct LockFreeImmortalSpace<VM: VMBinding> {
    name: &'static str,
    cursor: Atomic<Address>,
    limit: Address,
    start: Address,
    total_bytes: usize,
    slow_path_zeroing: bool,
    metadata: SideMetadataContext,
    gc_trigger: Arc<GCTrigger<VM>>,
}

This type implements a lock free version of the immortal collection policy. This is close to the OpenJDK’s epsilon GC. Different from the normal ImmortalSpace, this version should only be used by NoGC plan, and it now uses the whole heap range.

Fields
name: &'static strcursor: Atomic<Address>

Heap range start

limit: Address

Heap range end

start: Address

start of this space

total_bytes: usize

Total bytes for the space

slow_path_zeroing: bool

Zero memory after slow-path allocation

metadata: SideMetadataContextgc_trigger: Arc<GCTrigger<VM>>
Implementations

impl<VM: VMBinding> LockFreeImmortalSpace<VM>

pub fn new(args: PlanCreateSpaceArgs<'_, VM>) -> Self