Struct mmtk::policy::mallocspace::global::MallocSpace

pub struct MallocSpace<VM: VMBinding> {
    phantom: PhantomData<VM>,
    active_bytes: AtomicUsize,
    pub chunk_addr_min: AtomicUsize,
    pub chunk_addr_max: AtomicUsize,
    metadata: SideMetadataContext,
    active_mem: Mutex<HashMap<Address, usize>>,
    pub total_work_packets: AtomicU32,
    pub completed_work_packets: AtomicU32,
    pub work_live_bytes: AtomicUsize,
}

Fields
phantom: PhantomData<VM>active_bytes: AtomicUsizechunk_addr_min: AtomicUsizechunk_addr_max: AtomicUsizemetadata: SideMetadataContextactive_mem: Mutex<HashMap<Address, usize>>total_work_packets: AtomicU32completed_work_packets: AtomicU32work_live_bytes: AtomicUsize
Implementations

impl<VM: VMBinding> MallocSpace<VM>

pub fn new(global_side_metadata_specs: Vec<SideMetadataSpec>) -> Self

pub fn alloc(&self, tls: VMThread, size: usize) -> Address

pub fn free(&self, addr: Address, bytes: usize)

pub fn trace_object<T: TransitiveClosure>(
&self,
trace: &mut T,
object: ObjectReference
) -> ObjectReference

fn map_metadata_and_update_bound(&self, chunk_start: Address)

pub fn sweep_chunk(&self, chunk_start: Address)

fn sweep_chunk_mark_on_side(
&self,
chunk_start: Address,
mark_bit_spec: SideMetadataSpec
)

This function is called when the mark bits sit on the side metadata. This has been optimized with the use of bulk loading and bulk zeroing of metadata.

This function uses non-atomic accesses to side metadata (although these non-atomic accesses should not have race conditions associated with them) as well as calls libc functions (malloc_usable_size(), free())

fn sweep_chunk_mark_in_header(&self, chunk_start: Address)

This sweep function is called when the mark bit sits in the object header

This function uses non-atomic accesses to side metadata (although these non-atomic accesses should not have race conditions associated with them) as well as calls libc functions (malloc_usable_size(), free())