Struct mmtk::policy::marksweepspace::malloc_ms::global::MallocSpace

pub struct MallocSpace<VM: VMBinding> {
    phantom: PhantomData<VM>,
    active_bytes: AtomicUsize,
    active_pages: AtomicUsize,
    pub chunk_addr_min: Atomic<Address>,
    pub chunk_addr_max: Atomic<Address>,
    metadata: SideMetadataContext,
    scheduler: Arc<GCWorkScheduler<VM>>,
    gc_trigger: Arc<GCTrigger<VM>>,
    active_mem: Mutex<HashMap<Address, usize>>,
    pub total_work_packets: AtomicU32,
    pub completed_work_packets: AtomicU32,
    pub work_live_bytes: AtomicUsize,
}

This space uses malloc to get new memory, and performs mark-sweep for the memory.

Fields
phantom: PhantomData<VM>active_bytes: AtomicUsizeactive_pages: AtomicUsizechunk_addr_min: Atomic<Address>chunk_addr_max: Atomic<Address>metadata: SideMetadataContextscheduler: Arc<GCWorkScheduler<VM>>

Work packet scheduler

gc_trigger: Arc<GCTrigger<VM>>active_mem: Mutex<HashMap<Address, usize>>total_work_packets: AtomicU32completed_work_packets: AtomicU32work_live_bytes: AtomicUsize
Implementations

impl<VM: VMBinding> MallocSpace<VM>

pub fn extend_global_side_metadata_specs(specs: &mut Vec<SideMetadataSpec>)

pub fn new(args: PlanCreateSpaceArgs<'_, VM>) -> Self

fn set_page_mark(&self, start: Address, size: usize)

Set multiple pages, starting from the given address, for the given size, and increase the active page count if we set any page mark in the region. This is a thread-safe method, and can be used during mutator phase when mutators may access the same page. Performance-wise, this method may impose overhead, as we are doing a compare-exchange for every page in the range.

unsafe fn unset_page_mark(&self, start: Address, size: usize)

Unset multiple pages, starting from the given address, for the given size, and decrease the active page count if we unset any page mark in the region

Safety

We need to ensure that only one GC thread is accessing the range.

pub fn alloc( &self, tls: VMThread, size: usize, align: usize, offset: usize ) -> Address

pub fn free(&self, addr: Address)

fn free_internal(&self, addr: Address, bytes: usize, offset_malloc_bit: bool)

pub fn trace_object<Q: ObjectQueue>( &self, queue: &mut Q, object: ObjectReference ) -> ObjectReference

fn map_metadata_and_update_bound(&self, addr: Address, size: usize)

pub fn prepare(&mut self)

pub fn release(&mut self)

pub fn end_of_gc(&mut self)

pub fn sweep_chunk(&self, chunk_start: Address)

fn get_malloc_addr_size(object: ObjectReference) -> (Address, bool, usize)

Given an object in MallocSpace, return its malloc address, whether it is an offset malloc, and malloc size

fn clean_up_empty_chunk(&self, chunk_start: Address)

Clean up for an empty chunk

fn sweep_object( &self, object: ObjectReference, empty_page_start: &mut Address ) -> bool

Sweep an object if it is dead, and unset page marks for empty pages before this object. Return true if the object is swept.

fn debug_sweep_chunk_done(&self, live_bytes_in_the_chunk: usize)

Used when each chunk is done. Only called in debug build.

fn sweep_chunk_mark_on_side( &self, chunk_start: Address, mark_bit_spec: SideMetadataSpec )

This function is called when the mark bits sit on the side metadata. This has been optimized with the use of bulk loading and bulk zeroing of metadata.

This function uses non-atomic accesses to side metadata (although these non-atomic accesses should not have race conditions associated with them) as well as calls libc functions (malloc_usable_size(), free())

fn sweep_chunk_mark_in_header(&self, chunk_start: Address)

This sweep function is called when the mark bit sits in the object header

This function uses non-atomic accesses to side metadata (although these non-atomic accesses should not have race conditions associated with them) as well as calls libc functions (malloc_usable_size(), free())

fn sweep_each_object_in_chunk(&self, chunk_start: Address)