mmtk::policy::marksweepspace::malloc_ms

Module metadata

Constants

ACTIVE_PAGE_METADATA_SPEC 🔒

Metadata spec for the active page byte

OFFSET_MALLOC_METADATA_SPEC 🔒

Functions

compare_exchange_set_page_mark 🔒

Set the page mark from 0 to 1. Return true if we set it successfully in this call.

is_marked

is_marked_unsafe⚠

is_offset_malloc 🔒

Is this allocation an offset malloc? The argument address should be the allocation address (object start)

is_page_marked 🔒

is_page_marked_unsafe 🔒 ⚠

load128 🔒 ⚠

Load u128 bits of side metadata

set_mark_bit

set_offset_malloc_bit 🔒

Set the offset bit for the allocation. The argument address should be the allocation address (object start)

set_page_mark 🔒

set_vo_bit

unset_mark_bit⚠

unset_offset_malloc_bit_unsafe 🔒 ⚠

Unset the offset bit for the allocation. The argument address should be the allocation address (object start)

unset_page_mark_unsafe 🔒 ⚠

unset_vo_bit

unset_vo_bit_unsafe⚠