mmtk::policy::marksweepspace::malloc_ms::metadata

Function load128

pub(super) unsafe fn load128(
    metadata_spec: &SideMetadataSpec,
    data_addr: Address,
) -> u128

Load u128 bits of side metadata

Safety

unsafe as it can segfault if one tries to read outside the bounds of the mapped side metadata