mmtk::policy::marksweepspace::native_ms

Module block_list

Structs

BlockList

List of blocks owned by the allocator

BlockListIterator

Constants

MAX_BIN 🔒

The largest valid bin.

MAX_BIN_SIZE 🔒

The object size for the last bin. We should not try allocate objects larger than this with the allocator.

MI_BIN_FULL 🔒

Number of bins in BlockLists. Reserve bin0 as an empty bin.

MI_INTPTR_BITS 🔒

pointer size in bits

MI_INTPTR_SHIFT 🔒

Log2 of pointer size

MI_INTPTR_SIZE 🔒

pointer size in bytes

MI_LARGE_OBJ_SIZE_MAX 🔒

Largest object size allowed with our mimalloc implementation, in bytes

MI_LARGE_OBJ_WSIZE_MAX 🔒

Largest object size in words

Functions

mi_bin

mi_bin_from_size 🔒

mi_wsize_from_size 🔒

Align a byte size to a size in machine words i.e. byte size == wsize*sizeof(void*) adapted from _mi_wsize_from_size in mimalloc

new_empty_block_lists 🔒

Create an empty set of block lists of different size classes (bins)

pages_used_by_blocklists 🔒

Returns how many pages the block lists uses.

Type Aliases

BlockLists

All the bins for the block lists