mmtk::policy::marksweepspace::native_ms

Module global

Structs

AbandonedBlockLists

MarkSweepSpace

A mark sweep space.

PrepareChunkMap 🔒

RecycleBlocks 🔒

ReleaseMarkSweepSpace 🔒

SweepChunk 🔒

Chunk sweeping work packet. Only used by eager sweeping to sweep marked blocks after unmarked blocks have been released.

Enums

BlockAcquireResult

The result for MarkSweepSpace.acquire_block(). MarkSweepSpace will attempt to allocate from abandoned blocks first. If none found, it will get a new block from the page resource.

Constants

MAX_OBJECT_SIZE