Struct mmtk::policy::sft_map::dense_chunk_map::SFTDenseChunkMap

pub struct SFTDenseChunkMap {
    sft: Vec<SFTRefStorage>,
    index_map: HashMap<String, usize>,
}

SFTDenseChunkMap is a small table. It has one entry for each space in the table, and use side metadata to record the index for each chunk. This works for both 32 bits and 64 bits. However, its performance is expected to be suboptimal, compared to the sparse chunk map on 32 bits, and the space map on 64 bits. So usually we do not use this implementation. However, it provides some flexibility so we can set SFT at chunk basis for 64bits for decent performance. For example, when we use library malloc for mark sweep, we have no control of where the library malloc may allocate into, so we cannot use the space map. And using a sparse chunk map will be costly in terms of memory. In this case, the dense chunk map is a good solution.

Fields
sft: Vec<SFTRefStorage>

The dense table, one entry per space. We use side metadata to store the space index for each chunk. 0 is EMPTY_SPACE_SFT.

index_map: HashMap<String, usize>

A map from space name (assuming they are unique) to their index. We use this to know whether we have pushed &dyn SFT for a space, and to know its index.

Implementations

impl SFTDenseChunkMap

const EMPTY_SFT_INDEX: u8 = 0u8

Empty space is at index 0

pub fn new() -> Self

pub fn addr_to_index(addr: Address) -> u8