Function mmtk::policy::space::print_vm_map

pub(crate) fn print_vm_map<VM: VMBinding>(
    space: &dyn Space<VM>,
    out: &mut impl Write
) -> Result<(), Error>

Print the VM map for a space. Space needs to be object-safe, so it cannot have methods that use extra generic type paramters. So this method is placed outside the Space trait. This method can be invoked on a &dyn Space (space.as_space() will return &dyn Space).