Trait mmtk::policy::space::SFT

pub trait SFT {
    fn name(&self) -> &str;

    fn is_live(&self, object: ObjectReference) -> bool;

    fn is_movable(&self) -> bool;

    fn initialize_header(&self, object: ObjectReference, alloc: bool);
}

Space Function Table (SFT).

This trait captures functions that reflect space-specific per-object semantics. These functions are implemented for each object via a special space-based dynamic dispatch mechanism where the semantics are not determined by the object's type, but rather, are determined by the space that the object is in.

The underlying mechanism exploits the fact that spaces use the address space at an MMTk chunk granularity with the consequence that each chunk maps to exactluy one space, so knowing the chunk for an object reveals its space. The dispatch then works by performing simple address arithmetic on the object reference to find a chunk index which is used to index a table which returns the space. The relevant function is then dispatched against that space object.

We use the SFT trait to simplify typing for Rust, so our table is a table of SFT rather than Space.

Required methods

fn name(&self) -> &str

fn is_live(&self, object: ObjectReference) -> bool

fn is_movable(&self) -> bool

fn initialize_header(&self, object: ObjectReference, alloc: bool)

Loading content...