Trait mmtk::policy::space::Space

pub trait Space<VM: VMBinding>: 'static + SFT + Sync + Downcast {
    pub fn as_space(&self) -> &dyn Space<VM>;

    pub fn as_sft(&self) -> &(dyn SFT + Sync + 'static);

    pub fn get_page_resource(&self) -> &dyn PageResource<VM>;

    pub fn init(&mut self, vm_map: &'static VMMap);

    pub fn common(&self) -> &CommonSpace<VM>;

    pub fn release_multiple_pages(&mut self, start: Address);

    pub fn acquire(&self, tls: OpaquePointer, pages: usize) -> Address { ... }

    pub fn address_in_space(&self, start: Address) -> bool { ... }

    pub fn in_space(&self, object: ObjectReference) -> bool { ... }

    pub fn grow_space(&self, start: Address, bytes: usize, new_chunk: bool) { ... }

    pub fn ensure_mapped(&self) { ... }

    pub fn reserved_pages(&self) -> usize { ... }

    pub fn get_name(&self) -> &'static str { ... }

    pub fn print_vm_map(&self) { ... }

    pub fn local_side_metadata_specs(&self) -> &[SideMetadataSpec] { ... }
}

Required methods

pub fn as_space(&self) -> &dyn Space<VM>

pub fn as_sft(&self) -> &(dyn SFT + Sync + 'static)

pub fn get_page_resource(&self) -> &dyn PageResource<VM>

pub fn init(&mut self, vm_map: &'static VMMap)

pub fn common(&self) -> &CommonSpace<VM>

pub fn release_multiple_pages(&mut self, start: Address)

Loading content...
Provided methods

pub fn acquire(&self, tls: OpaquePointer, pages: usize) -> Address

pub fn address_in_space(&self, start: Address) -> bool

pub fn in_space(&self, object: ObjectReference) -> bool

pub fn grow_space(&self, start: Address, bytes: usize, new_chunk: bool)

This is called after we get result from page resources. The space may tap into the hook to monitor heap growth. The call is made from within the page resources' critical region, immediately before yielding the lock.

@param start The start of the newly allocated space @param bytes The size of the newly allocated space @param new_chunk {@code true} if the new space encroached upon or started a new chunk or chunks.

pub fn ensure_mapped(&self)

Ensure this space is marked as mapped -- used when the space is already mapped (e.g. for a vm image which is externally mmapped.)

pub fn reserved_pages(&self) -> usize

pub fn get_name(&self) -> &'static str

pub fn print_vm_map(&self)

pub fn local_side_metadata_specs(&self) -> &[SideMetadataSpec]

Loading content...
Implementations

impl<VM> dyn Space<VM> where
VM: Any + 'static,
VM: VMBinding,

pub fn is<__T: Space<VM>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Space<VM>>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: Space<VM>>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: Space<VM>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: Space<VM>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.