Trait mmtk::scheduler::context::WorkerLocal

pub trait WorkerLocal {
    fn init(&mut self, _tls: VMWorkerThread) { ... }
}

Thread-local data for each worker thread.

For mmtk, each gc can define their own worker-local data, to contain their required copy allocators and other stuffs.

Provided methods

fn init(&mut self, _tls: VMWorkerThread)

Loading content...