Enum mmtk::scheduler::WorkBucketStage

pub enum WorkBucketStage {

    Unconstrained,
    Prepare,
    ClearVOBits,
    TPinningClosure,
    PinningRootsTrace,
    Closure,
    SoftRefClosure,
    WeakRefClosure,
    FinalRefClosure,
    PhantomRefClosure,
    VMRefClosure,
    CalculateForwarding,
    SecondRoots,
    RefForwarding,
    FinalizableForwarding,
    VMRefForwarding,
    Compact,
    Release,
    Final,

}

Variants

Unconstrained

This bucket is always open.

Prepare

Preparation work. Plans, spaces, GC workers, mutators, etc. should be prepared for GC at this stage.

ClearVOBits

Clear the VO bit metadata. Mainly used by ImmixSpace.

TPinningClosure

Compute the transtive closure starting from transitively pinning (TP) roots following only strong references. No objects in this closure are allow to move.

PinningRootsTrace

Trace (non-transitively) pinning roots. Objects pointed by those roots must not move, but their children may. To ensure correctness, these must be processed after TPinningClosure

Closure

Compute the transtive closure following only strong references.

SoftRefClosure

Handle Java-style soft references, and potentially expand the transitive closure.

WeakRefClosure

Handle Java-style weak references.

FinalRefClosure

Resurrect Java-style finalizable objects, and potentially expand the transitive closure.

PhantomRefClosure

Handle Java-style phantom references.

VMRefClosure

Let the VM handle VM-specific weak data structures, including weak references, weak collections, table of finalizable objects, ephemerons, etc. Potentially expand the transitive closure.

NOTE: This stage is intended to replace the Java-specific weak reference handling stages above.

CalculateForwarding

Compute the forwarding addresses of objects (mark-compact-only).

SecondRoots

Scan roots again to initiate another transitive closure to update roots and reference after computing the forwarding addresses (mark-compact-only).

RefForwarding

Update Java-style weak references after computing forwarding addresses (mark-compact-only).

NOTE: This stage should be updated to adapt to the VM-side reference handling. It shall be kept after removing {Soft,Weak,Final,Phantom}RefClosure.

FinalizableForwarding

Update the list of Java-style finalization cadidates and finalizable objects after computing forwarding addresses (mark-compact-only).

VMRefForwarding

Let the VM handle the forwarding of reference fields in any VM-specific weak data structures, including weak references, weak collections, table of finalizable objects, ephemerons, etc., after computing forwarding addresses (mark-compact-only).

NOTE: This stage is intended to replace Java-specific forwarding phases above.

Compact

Compact objects (mark-compact-only).

Release

Work packets that should be done just before GC shall go here. This includes releasing resources and setting states in plans, spaces, GC workers, mutators, etc.

Final

Resume mutators and end GC.

Implementations

impl WorkBucketStage

pub fn first_stw_stage() -> Self