mmtk::scheduler

Module gc_work

Structs

DefaultObjectTracer 🔒

This implementation of ObjectTracer queues newly visited objects and create the ProcessNodes work packets to scan and trace objects.

DefaultObjectTracerContext 🔒

This implementation of ObjectTracerContext creates the DefaultObjectTracer to expand the transitive closure during a stop-the-world tracing GC or the final mark pause of a concurrent GC. It is used during object scanning as well as weak reference processing.

DefaultRootsWorkFactory 🔒

An implementation of RootsWorkFactory for stop-the-world tracing GC, i.e. finding the transitive closure from roots, with all mutators stopped.

Prepare

The global GC Preparation Work This work packet invokes prepare() for the plan (which will invoke prepare() for each space), and pushes work packets for preparing mutators and collectors. We should only have one such work packet per GC, before any actual GC work starts. We assume this work packet is the only running work packet that accesses plan, and there should be no other concurrent work packet that accesses plan (read or write). Otherwise, there may be a race condition.

PrepareCollector

The collector GC Preparation Work

PrepareMutator

The mutator GC Preparation Work

ProcessNodes

A work packet for scanning objects and optionally do node-enqueuing tracing during a stop-the-world tracing GC and the final mark pause of a concurrent GC.

ProcessPinningRoots 🔒

This work packet processes pinning roots during stop-the-world tracing GC.

ProcessSlots

A work packet for processing slots during a stop-the-world tracing GC and the final mark pause of a concurrent GC.

Release

The global GC release Work This work packet invokes release() for the plan (which will invoke release() for each space), and pushes work packets for releasing mutators and collectors. We should only have one such work packet per GC, after all actual GC work ends. We assume this work packet is the only running work packet that accesses plan, and there should be no other concurrent work packet that accesses plan (read or write). Otherwise, there may be a race condition.

ReleaseCollector

The collector release Work

ReleaseMutator

The mutator release Work

ScanMutatorRoots

ScanVMSpecificRoots

ScheduleCollection

StopMutators

Stop all mutators

VMForwardWeakRefs

Delegate to the VM binding for forwarding weak references.

VMPostForwarding

This work packet calls Collection::post_forwarding.

VMProcessWeakRefs

Delegate to the VM binding for weak reference processing.

Enums

RootsKind 🔒

For USDT tracepoints for roots. Keep in sync with tools/tracing/timeline/visualize.py.