Struct mmtk::scheduler::gc_work::ScanObjects

pub struct ScanObjects<Edges: ProcessEdgesWork> {
    buffer: Vec<ObjectReference>,
    concurrent: bool,
    phantom: PhantomData<Edges>,
    bucket: WorkBucketStage,
}

Scan objects and enqueue the slots of the objects. For objects that do not support slot-enqueuing, this work packet also traces their outgoing edges directly.

This work packet does not execute policy-specific post-scanning hooks (it won’t call post_scan_object() in policy::gc_work::PolicyTraceObject). It should be used only for policies that do not perform policy-specific actions when scanning an object.

Fields
buffer: Vec<ObjectReference>concurrent: boolphantom: PhantomData<Edges>bucket: WorkBucketStage
Implementations

impl<Edges: ProcessEdgesWork> ScanObjects<Edges>

pub fn new( buffer: Vec<ObjectReference>, concurrent: bool, bucket: WorkBucketStage ) -> Self