Trait mmtk::scheduler::gc_work::ProcessEdgesWork

pub trait ProcessEdgesWork: Send + 'static + Sized + DerefMut + Deref<Target = ProcessEdgesBase<Self>> {
    type VM: VMBinding;

    const CAPACITY: usize;
    const OVERWRITE_REFERENCE: bool;
    const SCAN_OBJECTS_IMMEDIATELY: bool;

    fn new(
        edges: Vec<Address>, 
        roots: bool, 
        mmtk: &'static MMTK<Self::VM>
    ) -> Self;

    fn trace_object(&mut self, object: ObjectReference) -> ObjectReference;

    fn cache_roots_for_sanity_gc(&mut self) { ... }

    fn process_node(&mut self, object: ObjectReference) { ... }

    fn new_scan_work(&mut self, work_packet: impl GCWork<Self::VM>) { ... }

    fn flush(&mut self) { ... }

    fn process_edge(&mut self, slot: Address) { ... }

    fn process_edges(&mut self) { ... }
}

Scan & update a list of object slots

Associated Types

type VM: VMBinding

Associated Constants

const CAPACITY: usize

const OVERWRITE_REFERENCE: bool

const SCAN_OBJECTS_IMMEDIATELY: bool

Required methods

fn new(edges: Vec<Address>, roots: bool, mmtk: &'static MMTK<Self::VM>) -> Self

fn trace_object(&mut self, object: ObjectReference) -> ObjectReference

Provided methods

fn cache_roots_for_sanity_gc(&mut self)

fn process_node(&mut self, object: ObjectReference)

fn new_scan_work(&mut self, work_packet: impl GCWork<Self::VM>)

Create a new scan work packet. If SCAN_OBJECTS_IMMEDIATELY, the work packet will be executed immediately, in this method. Otherwise, the work packet will be added the Closure work bucket and will be dispatched later by the scheduler.

fn flush(&mut self)

Flush the nodes in ProcessEdgesBase, and create a ScanObjects work packet for it. If the node set is empty, this method will simply return with no work packet created.

fn process_edge(&mut self, slot: Address)

fn process_edges(&mut self)