Struct mmtk::scheduler::scheduler::GCWorkScheduler

pub struct GCWorkScheduler<VM: VMBinding> {
    pub work_buckets: EnumMap<WorkBucketStage, WorkBucket<VM>>,
    pub coordinator_work: WorkBucket<VM>,
    pub workers_shared: Vec<Arc<GCWorkerShared<VM>>>,
    coordinator_worker_shared: Arc<GCWorkerShared<VM>>,
    pub worker_monitor: Arc<(Mutex<()>, Condvar)>,
    closure_end: Mutex<Option<Box<dyn Send + Fn() -> bool>>>,
}

The shared data structure for distributing work packets between worker threads and the coordinator thread.

Fields
work_buckets: EnumMap<WorkBucketStage, WorkBucket<VM>>

Work buckets for worker threads

coordinator_work: WorkBucket<VM>

Work for the coordinator thread

workers_shared: Vec<Arc<GCWorkerShared<VM>>>

The shared parts of GC workers

coordinator_worker_shared: Arc<GCWorkerShared<VM>>

The shared part of the GC worker object of the controller thread

worker_monitor: Arc<(Mutex<()>, Condvar)>

Condition Variable for worker synchronization

closure_end: Mutex<Option<Box<dyn Send + Fn() -> bool>>>

A callback to be fired after the Closure bucket is drained. This callback should return true if it adds more work packets to the Closure bucket. WorkBucket::can_open then consult this return value to prevent the GC from proceeding to the next stage, if we still have Closure work to do.

We use this callback to process ephemeron objects. closure_end can re-enable the Closure bucket multiple times to iteratively discover and process more ephemeron objects.

Implementations

impl<VM: VMBinding> GCWorkScheduler<VM>

pub fn new(num_workers: usize) -> Arc<Self>

pub fn num_workers(&self) -> usize

pub fn all_workers_parked(&self) -> bool

pub fn spawn_gc_threads(
self: &Arc<Self>,
mmtk: &'static MMTK<VM>,
tls: VMThread
)

Create GC threads, including the controller thread and all workers.

pub fn schedule_common_work<C: GCWorkContext<VM = VM> + 'static>(
&self,
plan: &'static C::PlanType
)

Schedule all the common work packets

fn are_buckets_drained(&self, buckets: &[WorkBucketStage]) -> bool

pub fn on_closure_end(&self, f: Box<dyn Send + Fn() -> bool>)

pub fn all_buckets_empty(&self) -> bool

pub fn update_buckets(&self)

Open buckets if their conditions are met

pub fn deactivate_all(&self)

pub fn reset_state(&self)

pub fn debug_assert_all_buckets_deactivated(&self)

pub fn add_coordinator_work(
&self,
work: impl CoordinatorWork<VM>,
worker: &GCWorker<VM>
)

fn pop_scheduable_work(
&self,
worker: &GCWorker<VM>
) -> Option<(Box<dyn GCWork<VM>>, bool)>

pub fn poll(&self, worker: &GCWorker<VM>) -> Box<dyn GCWork<VM>>

Get a scheduable work. Called by workers

fn poll_slow(&self, worker: &GCWorker<VM>) -> Box<dyn GCWork<VM>>

pub fn enable_stat(&self)

pub fn statistics(&self) -> HashMap<String, String>

pub fn notify_mutators_paused(&self, mmtk: &'static MMTK<VM>)