Trait mmtk::scheduler::CoordinatorWork

pub trait CoordinatorWork<C: Context>: 'static + Send + Sync + Work<C> { }

A special kind of work that will execute on the coorddinator (i.e. controller) thread

The coorddinator thread holds the global monitor lock when executing CoordinatorWorks. So, directly adding new work to any buckets will cause dead lock. For this case, use WorkBucket::add_with_priority_unsync instead.