Trait mmtk::scheduler::work::GCWork

pub trait GCWork<VM: VMBinding>: 'static + Send {
    // Required method
    fn do_work(&mut self, worker: &mut GCWorker<VM>, mmtk: &'static MMTK<VM>);

    // Provided methods
    fn do_work_with_stat(
        &mut self,
        worker: &mut GCWorker<VM>,
        mmtk: &'static MMTK<VM>
    ) { ... }
    fn get_type_name(&self) -> &'static str { ... }
}

Required Methods

fn do_work(&mut self, worker: &mut GCWorker<VM>, mmtk: &'static MMTK<VM>)

Define the work for this packet. However, this is not supposed to be called directly. Usually do_work_with_stat() should be used.

Provided Methods

fn do_work_with_stat( &mut self, worker: &mut GCWorker<VM>, mmtk: &'static MMTK<VM> )

Do work and collect statistics. This internally calls do_work(). In most cases, this should be called rather than do_work() so that MMTk can correctly collect statistics for the work packets. If the feature “work_packet_stats” is not enabled, this call simply forwards the call to do_work().

fn get_type_name(&self) -> &'static str

Get the compile-time static type name for the work packet.