Struct mmtk::scheduler::work_bucket::WorkBucket

pub struct WorkBucket<C: Context> {
    active: AtomicBool,
    queue: RwLock<BinaryHeap<PrioritizedWork<C>>>,
    monitor: Arc<(Mutex<()>, Condvar)>,
    can_open: Option<Box<dyn Fn() -> bool + Send>>,
}

Fields
active: AtomicBoolqueue: RwLock<BinaryHeap<PrioritizedWork<C>>>

A priority queue

monitor: Arc<(Mutex<()>, Condvar)>can_open: Option<Box<dyn Fn() -> bool + Send>>
Implementations

impl<C: Context> WorkBucket<C>

pub const DEFAULT_PRIORITY: usize

pub fn new(active: bool, monitor: Arc<(Mutex<()>, Condvar)>) -> Self

fn notify_one_worker(&self)

fn notify_all_workers(&self)

pub fn is_activated(&self) -> bool

pub fn activate(&self)

Enable the bucket

pub fn is_empty(&self) -> bool

Test if the bucket is drained

pub fn is_drained(&self) -> bool

pub fn deactivate(&self)

Disable the bucket

pub fn add_with_priority(&self, priority: usize, work: Box<dyn Work<C>>)

Add a work packet to this bucket, with a given priority

pub fn add<W: Work<C>>(&self, work: W)

Add a work packet to this bucket, with a default priority (1000)

pub fn bulk_add_with_priority(
&self,
priority: usize,
work_vec: Vec<Box<dyn Work<C>>>
)

pub fn bulk_add(&self, work_vec: Vec<Box<dyn Work<C>>>)

pub fn poll(&self) -> Option<Box<dyn Work<C>>>

Get a work packet (with the greatest priority) from this bucket

pub fn set_open_condition(&mut self, pred: impl Fn() -> bool + Send + 'static)

pub fn update(&self) -> bool