Trait mmtk::scheduler::work_counter::WorkCounter

pub(super) trait WorkCounter: WorkCounterClone + Debug {
    fn start(&mut self);

    fn stop(&mut self);

    fn name(&self) -> String;

    fn get_base(&self) -> &WorkCounterBase;

    fn get_base_mut(&mut self) -> &mut WorkCounterBase;
}

An abstraction of work counters

Use for trait objects, as we have might have types of work counters for the same work packet and the types are not statically known. The overhead should be negligible compared with the cost of executing a work packet.

Required methods

fn start(&mut self)

Start the counter

fn stop(&mut self)

Stop the counter

fn name(&self) -> String

Name of counter

fn get_base(&self) -> &WorkCounterBase

Return a reference to WorkCounterBase

fn get_base_mut(&mut self) -> &mut WorkCounterBase

Return a mutatable reference to WorkCounterBase

Loading content...