mmtk::scheduler::worker

Struct GCWorker

pub struct GCWorker<VM: VMBinding> {
    pub tls: VMWorkerThread,
    pub ordinal: usize,
    scheduler: Arc<GCWorkScheduler<VM>>,
    copy: GCWorkerCopyContext<VM>,
    pub mmtk: &'static MMTK<VM>,
    pub shared: Arc<GCWorkerShared<VM>>,
    pub local_work_buffer: Worker<Box<dyn GCWork<VM>>>,
}

A GC worker. This part is privately owned by a worker thread.

Fields
tls: VMWorkerThread

The VM-specific thread-local state of the GC thread.

ordinal: usize

The ordinal of the worker, numbered from 0 to the number of workers minus one.

scheduler: Arc<GCWorkScheduler<VM>>

The reference to the scheduler.

copy: GCWorkerCopyContext<VM>

The copy context, used to implement copying GC.

mmtk: &'static MMTK<VM>

The reference to the MMTk instance.

shared: Arc<GCWorkerShared<VM>>

Reference to the shared part of the GC worker. It is used for synchronization.

local_work_buffer: Worker<Box<dyn GCWork<VM>>>

Local work packet queue.

Implementations

impl<VM: VMBinding> GCWorker<VM>

const LOCALLY_CACHED_WORK_PACKETS: usize = 16usize

pub(crate) fn new( mmtk: &'static MMTK<VM>, ordinal: usize, scheduler: Arc<GCWorkScheduler<VM>>, shared: Arc<GCWorkerShared<VM>>, local_work_buffer: Worker<Box<dyn GCWork<VM>>>, ) -> Self

pub fn add_work_prioritized( &mut self, bucket: WorkBucketStage, work: impl GCWork<VM>, )

Add a work packet to the work queue and mark it with a higher priority. If the bucket is open, the packet will be pushed to the local queue, otherwise it will be pushed to the global bucket with a higher priority.

pub fn add_work(&mut self, bucket: WorkBucketStage, work: impl GCWork<VM>)

Add a work packet to the work queue. If the bucket is open, the packet will be pushed to the local queue, otherwise it will be pushed to the global bucket.

pub fn scheduler(&self) -> &GCWorkScheduler<VM>

Get the scheduler. There is only one scheduler per MMTk instance.

pub fn get_copy_context_mut(&mut self) -> &mut GCWorkerCopyContext<VM>

Get a mutable reference of the copy context for this worker.

fn poll(&mut self) -> Result<Box<dyn GCWork<VM>>, WorkerShouldExit>

Poll a ready-to-execute work packet in the following order:

Any packet that should be processed only by this worker.

Poll from the local work queue.

Poll from open global work-buckets

Steal from other workers

pub fn run(self: Box<Self>, tls: VMWorkerThread, mmtk: &'static MMTK<VM>)

Entry point of the worker thread.

This function will resolve thread affinity, if it has been specified by the user.

Each worker will keep polling and executing work packets in a loop. It runs until the worker is requested to exit. Currently a worker may exit after crate::mmtk::MMTK::prepare_to_fork is called.

Arguments:

tls: The VM-specific thread-local storage for this GC worker thread.

mmtk: A reference to an MMTk instance.