Struct mmtk::scheduler::worker::WorkerMonitor

pub(crate) struct WorkerMonitor {
    sync: Mutex<WorkerMonitorSync>,
    work_available: Condvar,
    all_workers_parked: Condvar,
}

Used to synchronize mutually exclusive operations between workers and controller, and also waking up workers when more work packets are available.

Fields
sync: Mutex<WorkerMonitorSync>

The synchronized part.

work_available: Condvar

This is notified when new work is made available for the workers. Particularly, it is notified when

sync.worker_group_state is transitioned to Working because

some workers still have designated work, or

some sentinel work packets are added to their drained buckets, or

some work buckets are opened, or

any work packet is added to any open bucket. Workers wait on this condvar.

all_workers_parked: Condvar

This is notified when all workers parked. The coordinator waits on this condvar.

Implementations

impl WorkerMonitor

pub fn new(worker_count: usize) -> Self

pub fn notify_work_available(&self, all: bool)

Wake up workers when more work packets are made available for workers. This function is called when adding work packets to buckets. This function doesn’t change the work_group_state variable. If workers are in the Sleeping state, use resume_and_wait to resume workers.

pub fn resume_and_wait(&self, all: bool)

Wake up workers and wait until they transition to Sleeping state again. This is called by the coordinator. If all is true, notify all workers; otherwise only notify one worker.

pub fn debug_is_sleeping(&self) -> bool

Test if the worker group is in the Sleeping state.

pub fn park_and_wait<VM: VMBinding>(&self, worker: &GCWorker<VM>)

Park until more work is available. The argument worker indicates this function can only be called by workers.