mmtk::scheduler
Module worker_monitor

This module contains WorkerMonitor and related types. It purposes includes:

allowing workers to park,

letting the last parked worker take action, and

letting workers and mutators notify workers when workers are given things to do.

Structs

WorkerMonitor 🔒

A data structure for synchronizing workers with each other and with mutators.

WorkerMonitorSync 🔒

The synchronized part of WorkerMonitor.

WorkerParker 🔒

This struct counts the number of workers parked and identifies the last parked worker.

Enums

LastParkedResult 🔒

The result type of the on_last_parked call-back in WorkMonitor::park_and_wait. It decides how many workers should wake up after on_last_parked.