mmtk::util
Module address

An abstract of memory address and object reference.

Structs

Address

Address represents an arbitrary address. This is designed to represent address and do address arithmetic mostly in a safe way, and to allow mark some operations as unsafe. This type needs to be zero overhead (memory wise and time wise). The idea is from the paper High-level Low-level Programming (VEE09) and JikesRVM.

ObjectReference

ObjectReference represents address for an object. Compared with Address, operations allowed on ObjectReference are very limited. No address arithmetics are allowed for ObjectReference. The idea is from the paper Demystifying Magic: High-level Low-level Programming (VEE09) and JikesRVM.

Type Aliases

ByteOffset

offset in byte

ByteSize

size in bytes