mmtk::util::alloc

Module allocator

The allocator trait and allocation-related functions.

Structs

AllocationOptions

Allow specifying different behaviors with Allocator::alloc_with_options.

AllocationOptionsHolder 🔒

A wrapper for AllocatorContext to hold a AllocationOptions that can be modified by the same mutator thread.

AllocatorContext

The context an allocator needs to access in order to perform allocation.

Enums

AllocationError

A list of errors that MMTk can encounter during allocation.

Traits

Allocator

A trait which implements allocation routines. Every allocator needs to implements this trait.

Functions

align_allocation

align_allocation_inner

align_allocation_no_fill

assert_allocation_args 🔒

fill_alignment_gap

Fill the specified region with the alignment value.

get_maximum_aligned_size

get_maximum_aligned_size_inner