mmtk::util::alloc

Module allocators

A list of all the allocators, embedded in Mutator

Structs

Allocators

Enums

AllocatorInfo

This type describes allocator information. It is used to generate fast paths for the GC. All offset fields are relative to Mutator.

AllocatorSelector

This type describe an allocator in the crate::Mutator. For some VM bindings, they may need to access this type from native code. This type is equivalent to the following native types: #[repr(C)] struct AllocatorSelector { tag: AllocatorSelectorTag, payload: u8, } #[repr(u8)] enum AllocatorSelectorTag { BumpPointer, LargeObject, … }

Constants

MAX_BUMP_ALLOCATORS 🔒

MAX_FREE_LIST_ALLOCATORS 🔒

MAX_IMMIX_ALLOCATORS 🔒

MAX_LARGE_OBJECT_ALLOCATORS 🔒

MAX_MALLOC_ALLOCATORS 🔒

MAX_MARK_COMPACT_ALLOCATORS 🔒