Struct mmtk::util::alloc::allocators::Allocators

#[repr(C)]
pub struct Allocators<VM: VMBinding> {
    pub bump_pointer: [MaybeUninit<BumpAllocator<VM>>; 6],
    pub large_object: [MaybeUninit<LargeObjectAllocator<VM>>; 2],
    pub malloc: [MaybeUninit<MallocAllocator<VM>>; 1],
    pub immix: [MaybeUninit<ImmixAllocator<VM>>; 1],
    pub free_list: [MaybeUninit<FreeListAllocator<VM>>; 2],
    pub markcompact: [MaybeUninit<MarkCompactAllocator<VM>>; 1],
}

Fields
bump_pointer: [MaybeUninit<BumpAllocator<VM>>; 6]large_object: [MaybeUninit<LargeObjectAllocator<VM>>; 2]malloc: [MaybeUninit<MallocAllocator<VM>>; 1]immix: [MaybeUninit<ImmixAllocator<VM>>; 1]free_list: [MaybeUninit<FreeListAllocator<VM>>; 2]markcompact: [MaybeUninit<MarkCompactAllocator<VM>>; 1]
Implementations

impl<VM: VMBinding> Allocators<VM>

pub unsafe fn get_allocator( &self, selector: AllocatorSelector ) -> &dyn Allocator<VM>

Safety

The selector needs to be valid, and points to an allocator that has been initialized.

pub unsafe fn get_typed_allocator<T: Allocator<VM>>( &self, selector: AllocatorSelector ) -> &T

Safety

The selector needs to be valid, and points to an allocator that has been initialized.

pub unsafe fn get_allocator_mut( &mut self, selector: AllocatorSelector ) -> &mut dyn Allocator<VM>

Safety

The selector needs to be valid, and points to an allocator that has been initialized.

pub unsafe fn get_typed_allocator_mut<T: Allocator<VM>>( &mut self, selector: AllocatorSelector ) -> &mut T

Safety

The selector needs to be valid, and points to an allocator that has been initialized.

pub fn new( mutator_tls: VMMutatorThread, mmtk: &MMTK<VM>, space_mapping: &[(AllocatorSelector, &'static dyn Space<VM>)] ) -> Self