mmtk::util::alloc

Module bumpallocator

Bump pointer allocator

Structs

BumpAllocator

A bump pointer allocator. It keeps a thread local allocation buffer, and bumps a cursor to allocate from the buffer.

BumpPointer

A common fast-path bump-pointer allocator shared across different allocator implementations that use bump-pointer allocation. A BumpPointer is always initialized with cursor = 0, limit = 0, so the first allocation always fails the check of cursor + size < limit and goes to the slowpath. A binding can also take advantage of this design to zero-initialize the a bump pointer.

Constants

BLOCK_MASK 🔒

BLOCK_SIZE 🔒

Size of a bump allocator block. Currently it is set to 32 KB.