Struct mmtk::util::alloc::immix_allocator::ImmixAllocator

#[repr(C)]
pub struct ImmixAllocator<VM: VMBinding> {
    pub tls: VMThread,
    pub bump_pointer: BumpPointer,
    space: &'static ImmixSpace<VM>,
    context: Arc<AllocatorContext<VM>>,
    hot: bool,
    copy: bool,
    pub(super) large_bump_pointer: BumpPointer,
    request_for_large: bool,
    line: Option<Line>,
}

Immix allocator

Fields
tls: VMThread

VMThread associated with this allocator instance

bump_pointer: BumpPointer

The fastpath bump pointer.

space: &'static ImmixSpace<VM>

Space instance associated with this allocator instance.

context: Arc<AllocatorContext<VM>>hot: bool

unused

copy: bool

Is this a copy allocator?

large_bump_pointer: BumpPointer

Bump pointer for large objects

request_for_large: bool

Is the current request for large or small?

line: Option<Line>

Hole-searching cursor

Implementations

impl<VM: VMBinding> ImmixAllocator<VM>

pub(crate) fn reset(&mut self)

impl<VM: VMBinding> ImmixAllocator<VM>

pub(crate) fn new( tls: VMThread, space: Option<&'static dyn Space<VM>>, context: Arc<AllocatorContext<VM>>, copy: bool ) -> Self

pub(crate) fn immix_space(&self) -> &'static ImmixSpace<VM>

fn overflow_alloc( &mut self, size: usize, align: usize, offset: usize ) -> Address

Large-object (larger than a line) bump allocation.

fn alloc_slow_hot( &mut self, size: usize, align: usize, offset: usize ) -> Address

Bump allocate small objects into recyclable lines (i.e. holes).

fn acquire_recyclable_lines( &mut self, size: usize, align: usize, offset: usize ) -> bool

Search for recyclable lines.

fn acquire_recyclable_block(&mut self) -> bool

Get a recyclable block from ImmixSpace.

fn acquire_clean_block( &mut self, size: usize, align: usize, offset: usize ) -> Address

fn require_new_block( &mut self, size: usize, align: usize, offset: usize ) -> bool

Return whether the TLAB has been exhausted and we need to acquire a new block. Assumes that the buffer limits have been restored using ImmixAllocator::restore_limit_for_stress. Note that this function may implicitly change the limits of the allocator.

fn set_limit_for_stress(&mut self)

Set fake limits for the bump allocation for stress tests. The fake limit is the remaining thread local buffer size, which should be always smaller than the bump cursor. This method may be reentrant. We need to check before setting the values.

fn restore_limit_for_stress(&mut self)

Restore the real limits for the bump allocation so we can properly do a thread local allocation. The fake limit is the remaining thread local buffer size, and we restore the actual limit from the size and the cursor. This method may be reentrant. We need to check before setting the values.