Module mmtk::util::alloc::object_ref_guard

Functions to ensure an object reference for an allocation has valid metadata. This module includes functions to make sure the following invariant always holds: for each object we allocate ([cell, cell + bytes)), the metadata for the object reference (object_ref) is always in the range of the allocated memory. Given that we always initialize metadata based on chunks, we simply need to make sure that object_ref is in the same chunk as [cell, cell + bytes). In other words, we avoid allocating an address for which the object reference may be in another chunk.

Note that where an ObjectReference points to is defined by a binding. We only have this problem if an object reference may point to an address that is outside our allocated memory (object_ref >= cell + bytes). We ask a binding to specify ObjectModel::OBJECT_REF_OFFSET_BEYOND_CELL if their object reference may point to an address outside the allocated memory. ObjectModel::OBJECT_REF_OFFSET_BEYOND_CELL should be the max of object_ref - cell.

There are various ways we deal with this.

For allocators that have a thread local buffer, we can adjust the buffer limit to make sure that the last object allocated in the buffer won’t cross chunks.

For allocators that allocate large objects, if the object size is larger than OBJECT_REF_OFFSET_BEYOND_CELL, it is guaranteed that the object reference is within the allocated memory.

For other allocators, we can check if the allocation result violates this invariant.

Functions

adjust_thread_local_buffer_limit

Adjust limit for thread local buffer to make sure that we will not allocate objects whose object reference may be in another chunk.

assert_object_ref_in_cell

Assert that the object reference should always inside the allocation cell

object_ref_may_cross_chunk

Check if the object reference for this allocation may cross and fall into the next chunk.