mmtk::util::analysis
Module obj_size

Macros

new_ctr 🔒

Structs

PerSizeClassObjectCounter

This file implements an analysis routine that counts the number of objects allocated in each size class. Here, a size class ‘sizeX’ is defined as ‘X bytes or lower’. For example, size64 is the size class with objects <= 64 bytes but > 32 bytes which is the previous size class.