Struct mmtk::util::analysis::obj_size::PerSizeClassObjectCounter

pub struct PerSizeClassObjectCounter {
    running: bool,
    size_classes: Mutex<HashMap<String, Arc<Mutex<EventCounter>>>>,
    stats: Arc<Stats>,
}

This file implements an analysis routine that counts the number of objects allocated in each size class. Here, a size class ‘sizeX’ is defined as ‘X bytes or lower’. For example, size64 is the size class with objects <= 64 bytes but > 32 bytes which is the previous size class.

We keep track of the size classes using a HashMap with the key being the name of the size class.

Fields
running: boolsize_classes: Mutex<HashMap<String, Arc<Mutex<EventCounter>>>>stats: Arc<Stats>
Implementations

impl PerSizeClassObjectCounter

pub fn new(running: bool, stats: Arc<Stats>) -> Self

fn size_class(&self, size: usize) -> usize