Trait mmtk::util::analysis::RtAnalysis

pub trait RtAnalysis<VM: VMBinding> {
    pub fn set_running(&mut self, running: bool);

    pub fn alloc_hook(&mut self, _size: usize, _align: usize, _offset: isize) { ... }

    pub fn gc_hook(&mut self, _mmtk: &'static MMTK<VM>) { ... }
}

This trait exposes hooks for developers to implement their own analysis routines.

Most traits would want to hook into the Stats and counters provided by the MMTk framework that are exposed to the Harness.

The arguments for the hooks should be sufficient, however, if one wishes to add other arguments, then they can create an analysis routine specific function and invoke it in its respective place.

Required methods

pub fn set_running(&mut self, running: bool)

Loading content...
Provided methods

pub fn alloc_hook(&mut self, _size: usize, _align: usize, _offset: isize)

pub fn gc_hook(&mut self, _mmtk: &'static MMTK<VM>)

Loading content...