Module mmtk::util::constants

Constants used in MMTk

Modules

card_scanning 🔒

Some card scanning constants ported from Java MMTK. As we haven’t implemented card scanning, these are not used at the moment.

java_specific_constants 🔒

Constants

BITS_IN_ADDRESS

The number of bits in an address

BITS_IN_BYTE

The number of bits in a byte

BITS_IN_PAGE

The number of bits in a page

BITS_IN_WORD

The number of bits in a word

BYTES_IN_ADDRESS

The number of bytes in an address

BYTES_IN_BYTE

The number of bytes in a byte

BYTES_IN_GBYTE

The number of bytes in a gigabyte

BYTES_IN_KBYTE

The number of bytes in a kilobyte

BYTES_IN_MBYTE

The number of bytes in a megabyte

BYTES_IN_PAGE

The number of bytes in a page

BYTES_IN_WORD

The number of bytes in a word

LAZY_SWEEP 🔒

Lazy sweeping - controlled from here because PlanConstraints needs to tell the VM that we need to support linear scan.

LOG_BITS_IN_ADDRESS

log2 of the number of bits in an address

LOG_BITS_IN_BYTE

log2 of the number of bits in a byte

LOG_BITS_IN_PAGE

log2 of the number of bits in a page

LOG_BITS_IN_WORD

log2 of the number of bits in a word

LOG_BYTES_IN_ADDRESS

log2 of the number of bytes in an address

LOG_BYTES_IN_ADDRESS_SPACE

log2 of the number of bytes in the address space

LOG_BYTES_IN_BYTE

log2 of the number of bytes in a byte

LOG_BYTES_IN_GBYTE

log2 of the number of bytes in a gigabyte

LOG_BYTES_IN_KBYTE

log2 of the number of bytes in a kilobyte

LOG_BYTES_IN_MBYTE

log2 of the number of bytes in a megabyte

LOG_BYTES_IN_PAGE

log2 of the number of bytes in a page

LOG_BYTES_IN_WORD

log2 of the number of bytes in a word

LOG_MIN_OBJECT_SIZE

log2 of the minimal object size in bytes.

MIN_OBJECT_SIZE

The minimal object size in bytes