Struct mmtk::util::copy::GCWorkerCopyContext

pub struct GCWorkerCopyContext<VM: VMBinding> {
    pub copy: [MaybeUninit<CopySpaceCopyContext<VM>>; 1],
    pub immix: [MaybeUninit<ImmixCopyContext<VM>>; 2],
    config: CopyConfig<VM>,
}

The thread local struct for each GC worker for copying. Each GC worker should include one instance of this struct for copying operations.

Fields
copy: [MaybeUninit<CopySpaceCopyContext<VM>>; 1]

Copy allocators for CopySpace

immix: [MaybeUninit<ImmixCopyContext<VM>>; 2]

Copy allocators for ImmixSpace

config: CopyConfig<VM>

The config for the plan

Implementations

impl<VM: VMBinding> GCWorkerCopyContext<VM>

pub fn alloc_copy(
&mut self,
original: ObjectReference,
bytes: usize,
align: usize,
offset: isize,
semantics: CopySemantics
) -> Address

Allocate for the object for GC copying.

Arguments:

original: The original object that will be copied.

bytes: The size in bytes for the allocation.

align: The alignment in bytes for the allocation.

offset: The offset in bytes for the allocation.

semantics: The copy semantic for this coying allocation. It determins which copy allocator will be used for the copying.

pub fn post_copy(
&mut self,
object: ObjectReference,
bytes: usize,
semantics: CopySemantics
)

Post allocation after allocating an object.

Arguments:

object: The newly allocated object (the new object after copying).

bytes: The size of the object in bytes.

semantics: The copy semantic used for the copying.

pub fn prepare(&mut self)

Prepare the copying allocators.

pub fn release(&mut self)

Release the copying allocators.

pub fn new(
worker_tls: VMWorkerThread,
plan: &'static dyn Plan<VM = VM>,
config: CopyConfig<VM>
) -> Self

Create a GCWorkerCopyContext based on the configuration for a copying plan.

Arguments:

worker_tls: The worker thread for this copy context.

plan: A reference to the current plan.

config: The configuration for the copy context.

pub fn new_non_copy() -> Self

Create a stub GCWorkerCopyContext for non copying plans.