mmtk::util

Module heap

Heap implementation, including page resource, mmapper, etc.

Modules

accounting ğŸ”’

blockpageresource ğŸ”’

chunk_map ğŸ”’

externalpageresource ğŸ”’

freelistpageresource ğŸ”’

gc_trigger ğŸ”’

heap_meta ğŸ”’

layout ğŸ”’

monotonepageresource ğŸ”’

pageresource ğŸ”’

regionpageresource ğŸ”’

space_descriptor ğŸ”’

vm_layout

The module defines virutal memory layout parameters.

vmrequest ğŸ”’

Structs

SpaceStats

Provides statistics about the space. This is exposed to bindings, as it is used in both crate::plan::Plan and GCTriggerPolicy.

Traits

GCTriggerPolicy

This trait describes a GC trigger policy. A triggering policy have hooks to be informed about GC start/end so they can collect some statistics about GC and allocation. The policy needs to decide the (current) heap limit and decide whether a GC should be performed.