Struct mmtk::util::heap::layout::fragmented_mapper::FragmentedMapper

pub struct FragmentedMapper {
    lock: Mutex<()>,
    free_slab_index: usize,
    free_slabs: Vec<Option<Box<[Atomic<MapState>; 2048]>>>,
    slab_table: Vec<Option<Box<[Atomic<MapState>; 2048]>>>,
    slab_map: Vec<Address>,
}

Fields
lock: Mutex<()>free_slab_index: usizefree_slabs: Vec<Option<Box<[Atomic<MapState>; 2048]>>>slab_table: Vec<Option<Box<[Atomic<MapState>; 2048]>>>slab_map: Vec<Address>
Implementations

impl FragmentedMapper

pub fn new() -> Self

fn new_slab() -> Box<[Atomic<MapState>; 2048]>

fn hash(addr: Address) -> usize

fn slab_table(&self, addr: Address) -> Option<&[Atomic<MapState>; 2048]>

fn get_or_allocate_slab_table(&self, addr: Address) -> &[Atomic<MapState>; 2048]

unsafe fn mut_self(&self) -> &mut Self

fn get_or_optionally_allocate_slab_table(
&mut self,
addr: Address,
allocate: bool
) -> Option<&[Atomic<MapState>; 2048]>

fn slab_table_for(
&self,
_addr: Address,
index: usize
) -> Option<&[Atomic<MapState>; 2048]>

fn commit_free_slab(&mut self, index: usize)

Take a free slab of chunks from the freeSlabs array, and insert it at the correct index in the slabTable. @param index slab table index

fn chunk_index_to_address(base: Address, chunk: usize) -> Address

fn slab_align_down(addr: Address) -> Address

@param addr an address @return the base address of the enclosing slab

fn slab_limit(addr: Address) -> Address

@param addr an address @return the base address of the next slab

fn chunk_index(slab: Address, addr: Address) -> usize

@param slab Address of the slab @param addr Address within a chunk (could be in the next slab) @return The index of the chunk within the slab (could be beyond the end of the slab)