Struct mmtk::util::heap::layout::map32::Map32

pub struct Map32 {
    prev_link: Vec<i32>,
    next_link: Vec<i32>,
    region_map: IntArrayFreeList,
    global_page_map: IntArrayFreeList,
    shared_discontig_fl_count: usize,
    shared_fl_map: Vec<Option<&'static CommonFreeListPageResource>>,
    total_available_discontiguous_chunks: usize,
    finalized: bool,
    sync: Mutex<()>,
    descriptor_map: Vec<SpaceDescriptor>,
    cumulative_committed_pages: AtomicUsize,
}

Fields
prev_link: Vec<i32>next_link: Vec<i32>region_map: IntArrayFreeListglobal_page_map: IntArrayFreeListshared_discontig_fl_count: usizeshared_fl_map: Vec<Option<&'static CommonFreeListPageResource>>total_available_discontiguous_chunks: usizefinalized: boolsync: Mutex<()>descriptor_map: Vec<SpaceDescriptor>cumulative_committed_pages: AtomicUsize
Implementations

impl Map32

pub fn new() -> Self

impl Map32

unsafe fn mut_self(&self) -> &mut Self

Safety

The caller needs to guarantee there is no race condition. Either only one single thread is using this method, or multiple threads are accessing mutally exclusive data (e.g. different indices in arrays). In other cases, use mut_self_with_sync().

fn mut_self_with_sync(&self) -> (MutexGuard<'_, ()>, &mut Self)

fn free_contiguous_chunks_no_lock(&mut self, chunk: i32) -> usize

fn get_discontig_freelist_pr_ordinal(&self) -> usize