mmtk::util::heap::layout::mmapper
Module csm

Modules

byte_map_storage 🔒

two_level_storage 🔒

This module contains TwoLevelStateStorage, an implementation of MapStateStorage that is designed to work well on 64-bit machines. Currently it supports 48-bit address spaces, and many constants and data structures (such as Slab) are larger than i32::MAX. For this reason, this module is only available on 64-bit machines.

Structs

ChunkRange 🔒

A range of whole chunks. Always aligned.

ChunkStateMmapper

A Mmapper implementation based on a logical array of chunk states.

Enums

MapState 🔒

The mmap state of a mmap chunk.

Traits

MapStateStorage 🔒

The back-end storage of ChunkStateMmapper. It is responsible for holding the states of each chunk (eagerly or lazily) and transitioning the states in bulk.

Type Aliases

ChosenMapStateStorage 🔒