Module mmtk::util::heap::layout::vm_layout

The module defines virutal memory layout parameters.

Structs

VMLayout

Runtime-initialized virtual memory constants

Constants

BYTES_IN_CHUNK

Coarsest unit of address space allocation.

CHUNK_MASK

Mask for chunk size.

LOG_BYTES_IN_CHUNK

log_2 of the coarsest unit of address space allocation.

LOG_MMAP_CHUNK_BYTES

log_2 of the granularity at which we map and unmap virtual address space in the heap

MMAP_CHUNK_BYTES

Granularity at which we map and unmap virtual address space in the heap

PAGES_IN_CHUNK

Coarsest unit of address space allocation, in pages

Statics

VM_LAYOUT 🔒

VM_LAYOUT_FETCHED 🔒

Functions

vm_layout

Get the current virtual memory layout in use. If the binding would like to set a custom virtual memory layout (crate::mmtk::MMTKBuilder::set_vm_layout), they should not call this function before they set a custom layout.