mmtk::util::heap::monotonepageresource

Struct MonotonePageResource

pub struct MonotonePageResource<VM: VMBinding> {
    common: CommonPageResource,
    sync: Mutex<MonotonePageResourceSync>,
    _p: PhantomData<VM>,
}

Fields
common: CommonPageResourcesync: Mutex<MonotonePageResourceSync>_p: PhantomData<VM>
Implementations

impl<VM: VMBinding> MonotonePageResource<VM>

pub fn new_contiguous( start: Address, bytes: usize, vm_map: &'static dyn VMMap, ) -> Self

pub fn new_discontiguous(vm_map: &'static dyn VMMap) -> Self

pub fn cursor(&self) -> Address

Get highwater mark of current monotone space.

fn log_chunk_fields(&self, space_descriptor: SpaceDescriptor, site: &str)

fn get_region_start(addr: Address) -> Address

pub unsafe fn reset(&self)

Safety

TODO: I am not sure why this is unsafe.

pub unsafe fn get_current_chunk(&self) -> Address

pub fn reset_cursor(&self, top: Address)

unsafe fn release_pages( &self, guard: &mut MutexGuard<'_, MonotonePageResourceSync>, )

pub fn iterate_allocated_regions( &self, ) -> impl Iterator<Item = (Address, usize)> + '_

Iterate over all contiguous memory regions in this space. For contiguous space, this iterator should yield only once, and returning a contiguous memory region covering the whole space.

fn release_pages_extent(&self, _first: Address, bytes: usize)

fn move_to_next_chunk( &self, guard: &mut MutexGuard<'_, MonotonePageResourceSync>, ) -> bool