Trait mmtk::util::heap::pageresource::PageResource

pub trait PageResource<VM: VMBinding>: 'static {
    pub fn alloc_pages(
        &self, 
        reserved_pages: usize, 
        required_pages: usize, 
        zeroed: bool, 
        tls: OpaquePointer
    ) -> Address;

    pub fn adjust_for_metadata(&self, pages: usize) -> usize;

    pub fn common(&self) -> &CommonPageResource<VM>;

    pub fn common_mut(&mut self) -> &mut CommonPageResource<VM>;

    pub fn get_new_pages(
        &self, 
        reserved_pages: usize, 
        required_pages: usize, 
        zeroed: bool, 
        tls: OpaquePointer
    ) -> Address { ... }

    pub fn reserve_pages(&self, pages: usize) -> usize { ... }

    pub fn clear_request(&self, reserved_pages: usize) { ... }

    pub fn update_zeroing_approach(&self, _nontemporal: bool, concurrent: bool) { ... }

    pub fn skip_concurrent_zeroing(&self) { ... }

    pub fn trigger_concurrent_zeroing(&self) { ... }

    pub fn concurrent_zeroing(&self) { ... }

    pub fn commit_pages(
        &self, 
        reserved_pages: usize, 
        actual_pages: usize, 
        tls: OpaquePointer
    ) { ... }

    pub fn reserved_pages(&self) -> usize { ... }

    pub fn committed_pages(&self) -> usize { ... }

    pub fn bind_space(&mut self, space: &'static dyn Space<VM>) { ... }

    pub fn vm_map(&self) -> &'static VMMap { ... }
}

Required methods

pub fn alloc_pages(
&self,
reserved_pages: usize,
required_pages: usize,
zeroed: bool,
tls: OpaquePointer
) -> Address

pub fn adjust_for_metadata(&self, pages: usize) -> usize

pub fn common(&self) -> &CommonPageResource<VM>

pub fn common_mut(&mut self) -> &mut CommonPageResource<VM>

Loading content...
Provided methods

pub fn get_new_pages(
&self,
reserved_pages: usize,
required_pages: usize,
zeroed: bool,
tls: OpaquePointer
) -> Address

Allocate pages from this resource. Simply bump the cursor, and fail if we hit the sentinel. Return The start of the first page if successful, zero on failure.

pub fn reserve_pages(&self, pages: usize) -> usize

pub fn clear_request(&self, reserved_pages: usize)

pub fn update_zeroing_approach(&self, _nontemporal: bool, concurrent: bool)

pub fn skip_concurrent_zeroing(&self)

pub fn trigger_concurrent_zeroing(&self)

pub fn concurrent_zeroing(&self)

pub fn commit_pages(
&self,
reserved_pages: usize,
actual_pages: usize,
tls: OpaquePointer
)

Commit pages to the page budget. This is called after successfully determining that the request can be satisfied by both the page budget and virtual memory. This simply accounts for the discrepancy between committed and reserved while the request was pending.

This MUST be called by each PageResource during the allocPages, and the caller must hold the lock.

pub fn reserved_pages(&self) -> usize

pub fn committed_pages(&self) -> usize

pub fn bind_space(&mut self, space: &'static dyn Space<VM>)

pub fn vm_map(&self) -> &'static VMMap

Loading content...