Module mmtk::util::malloc

Various malloc implementations (conditionally compiled by features)

Functions

align_alloc

align_offset_alloc

alloc

allocate size bytes, which is aligned to align at offset return the address, and whether it is an offset allocation

calloc⚠

free⚠

get_malloc_usable_size

get malloc usable size of an address is_offset_malloc: whether the address is allocated with some offset

malloc_usable_size⚠

offset_free

free an address that is allocated with some offset

offset_malloc_usable_size

posix_memalign⚠