mmtk::util::memory

Function mmap_noreserve

pub fn mmap_noreserve(
    start: Address,
    size: usize,
    strategy: MmapStrategy,
    anno: &MmapAnnotation<'_>,
) -> Result<()>

mmap with no swap space reserve: This function does not reserve swap space for this mapping, which means there is no guarantee that writes to the mapping can always be successful. In case of out of physical memory, one may get a segfault for writing to the mapping. We can use this to reserve the address range, and then later overwrites the mapping with dzmmap().