Function mmtk::util::metadata::compare_exchange_metadata

pub fn compare_exchange_metadata<VM: VMBinding>(
    metadata_spec: &MetadataSpec, 
    object: ObjectReference, 
    old_val: usize, 
    new_val: usize, 
    mask: Option<usize>, 
    success_order: Ordering, 
    failure_order: Ordering
) -> bool

A function to atomically compare-and-exchange the specified metadata’s content.

Arguments:

metadata_spec: is one of the const MetadataSpec instances from the ObjectModel trait, for the target metadata. Whether the metadata is in-header or on-side is a VM-specific choice.

object: is a reference to the target object.

old_val: is the expected current value of the metadata.

new_val: is the new metadata value to be stored if the compare-and-exchange operation is successful.

mask: is an optional mask value for the metadata. This value is used in cases like the forwarding pointer metadata, where some of the bits are reused by other metadata such as the forwarding bits.

success_order: is the atomic ordering used if the operation is successful.

failure_order: is the atomic ordering used if the operation fails.

Returns true if the operation is successful, and false otherwise.