Module mmtk::util::metadata::header_metadata

Structs

HeaderMetadataSpec

This module provides a default implementation of the access functions for in-header metadata. This struct stores the specification of a header metadata bit-set. It is used as an input to the (inline) functions provided by the header metadata module or ObjectModel metadata modules.

Constants

BITS_IN_U32

BIT_IN_U16

LOG_BITS_IN_U16

LOG_BITS_IN_U32

Functions

compare_exchange_metadata

This function provides a default implementation for the compare_exchange_metadata method from the ObjectModel trait.

fetch_add_metadata

This function provides a default implementation for the fetch_add_metadata method from the ObjectModel trait.

fetch_sub_metadata

This function provides a default implementation for the fetch_sub_metadata method from the ObjectModel trait.

load_metadata

This function provides a default implementation for the load_metadata method from the ObjectModel trait.

store_metadata

This function provides a default implementation for the store_metadata method from the ObjectModel trait.