mmtk::util::metadata
Module metadata_val_traits

Macros

impl_bits_trait 🔒

impl_bitwise_ops_trait 🔒

impl_metadata_value_trait 🔒

Traits

Bits

Describes bits and log2 bits for the numbers. If num_traits has this, we do not need our own implementation: https://github.com/rust-num/num-traits/issues/247

BitwiseOps

Describes bitwise operations. If num_traits has this, we do not need our own implementation: https://github.com/rust-num/num-traits/issues/232

MetadataValue

The number type for accessing metadata. It requires a few traits from num-traits and a few traits we defined above. The methods in this trait are mostly about atomically accessing such types.