mmtk::util::metadata

Module side_metadata

This module provides an implementation of side table metadata.

Modules

global 🔒

helpers 🔒

layout 🔒

ranges 🔒

Data types for visiting metadata ranges at different granularities.

sanity 🔒

side_metadata_tests 🔒

spec_defs 🔒

Structs

MetadataByteArrayRef

A byte array in side-metadata

SideMetadataSpec

This struct stores the specification of a side metadata bit-set. It is used as an input to the (inline) functions provided by the side metadata module.

Constants

GLOBAL_SIDE_METADATA_VM_BASE_OFFSET

The base offset for the global side metadata available to VM bindings.

LOCAL_SIDE_METADATA_VM_BASE_OFFSET

The base address for the local side metadata space available to VM bindings, to be used for the per-object metadata. VM bindings must use this to avoid overlap with core internal local side metadata.

Functions

global_side_metadata_base_address

Get the runtime side metadata base address.

global_side_metadata_vm_base_address

The base address for the global side metadata space available to VM bindings, to be used for the per-object metadata. VM bindings must use this to avoid overlap with core internal global side metadata.

initialize_side_metadata

Initialize side metadata runtime state and reserve the side metadata address range.

side_metadata_offset_after

Calculate the offset of the next side metadata spec after the given spec. This is used to calculate the offset field in crate::util::metadata::side_metadata::SideMetadataSpec.

vo_bit_side_metadata_addr

Base address of VO bit, public to VM bindings which may need to use this.