Module mmtk::util::metadata::side_metadata::global

Structs

MetadataByteArrayRef

A byte array in side-metadata

SideMetadataContext

This struct stores all the side metadata specs for a policy. Generally a policy needs to know its own side metadata spec as well as the plan’s specs.

SideMetadataSpec

This struct stores the specification of a side metadata bit-set. It is used as an input to the (inline) functions provided by the side metadata module.

Functions

bzero_metadata

Bulk-zero a specific metadata for a chunk.

compare_exchange_atomic

ensure_metadata_is_mapped

fetch_add_atomic

fetch_sub_atomic

load⚠

Non-atomic load of metadata.

load_atomic

store⚠

Non-atomic store of metadata.

store_atomic

Unions

SideMetadataOffset

A union of Address or relative offset (usize) used to store offset for a side metadata spec. If a spec is contiguous side metadata, it uses address. Othrewise it uses usize.