Module mmtk::util::metadata::side_metadata::helpers

Functions

addr_rshift

address_to_contiguous_meta_address

Performs address translation in contiguous metadata spaces (e.g. global and policy-specific in 64-bits, and global in 32-bits)

address_to_meta_address

Performs the translation of data address (data_addr) to metadata address for the specified metadata (metadata_spec).

meta_byte_lshift

meta_byte_mask

metadata_address_range_size

try_mmap_contiguous_metadata_space

Tries to mmap the metadata space (spec) for the specified data address range (start and size). Setting no_reserve to true means the function will only map address range, without reserving swap-space/physical memory. Returns the size in bytes that gets mmapped in the function if success.