mmtk::util::metadata::side_metadata

Module ranges

Data types for visiting metadata ranges at different granularities.

Currently, the break_bit_range function can break a bit range into sub-ranges of whole bytes and in-byte bits.

TODO:

Add a function to break a byte range into sub-ranges of whole words and in-word bytes.

And use it for searching side metadata for non-zero bits.

Add a function to break a byte range at chunk boundaries.

And use it for visiting discontiguous side metadata in bulk.

Enums

BitByteRange

A range of bytes or bits within a byte. It is the unit of visiting a contiguous bit range of a side metadata.

Functions

break_bit_range

Break a bit range into sub-ranges of whole bytes and in-byte bits.

Type Aliases

BitOffset

The type for bit offset in a byte.