Function mmtk::util::metadata::side_metadata::sanity::verify_store

pub fn verify_store(
    metadata_spec: SideMetadataSpec, 
    data_addr: Address, 
    metadata: usize
)

Commits a side metadata store operation. Panics if: 1 - the loaded side metadata content is not equal to the correct content.

Arguments:

metadata_spec: the metadata spec to commit the store operation for

data_addr: the address of the source data

metadata: the metadata content to store