Struct mmtk::util::metadata::side_metadata::sanity::CONTENT_SANITY_MAP

struct CONTENT_SANITY_MAP {
    __private_field: (),
}

This is a two-level hashmap to store the metadata content for verification purposes. It keeps a map from side metadata specifications to a second hashmap which maps data addresses to their current metadata content.

Fields
__private_field: ()