mmtk::util::metadata::side_metadata::sanity
Struct SideMetadataSanity

pub struct SideMetadataSanity {
    specs_sanity_map: HashMap<&'static str, Vec<SideMetadataSpec>>,
}

This struct includes a hashmap to store the metadata specs information for policy-specific and global metadata for each plan. It uses policy name (or GLOBAL_META_NAME for globals) as the key and keeps a vector of specs as the value. Each plan needs its exclusive instance of this struct to use side metadata specification and content sanity checker.

NOTE: Content sanity check is expensive and is only activated with the extreme_assertions feature.

FIXME: This struct should be pub(crate) visible, but changing its scope will need changing other scopes, such as the Space trait’s. For now, I will not do that.

Fields
specs_sanity_map: HashMap<&'static str, Vec<SideMetadataSpec>>
Implementations

impl SideMetadataSanity

pub fn new() -> SideMetadataSanity

Creates a new SideMetadataSanity instance.

fn get_all_specs(&self, global: bool) -> Vec<SideMetadataSpec>

Returns all global or policy-specific specs based-on the input argument.

Returns a vector of globals if global is true and a vector of locals otherwise.

Arguments:

global: a boolean to show whether global (true) or policy-specific (false) specs are required.

fn verify_local_specs(&self) -> Result<()>

Verifies that all local side metadata specs: 1 - are not too big, 2 - do not overlap.

Returns Ok(()) if no issue is detected, or Err otherwise.

pub(crate) fn verify_metadata_context( &mut self, policy_name: &'static str, metadata_context: &SideMetadataContext, )

An internal method to ensure that a metadata context does not have any issues.

Arguments:

policy_name: name of the policy of the calling space

metadata_context: the metadata context to examine

NOTE: Any unit test using metadata directly or indirectly may need to make sure: 1 - it uses util::test_util::serial_test to prevent metadata sanity conflicts, 2 - uses exclusive SideMetadata instances (v.s. static instances), and 3 - uses util::test_util::with_cleanup to call sanity::reset to cleanup the metadata sanity states to prevent future conflicts.