mmtk::util

Module object_enum

Helper types for object enumeration

Structs

ClosureObjectEnumerator

An implementation of ObjectEnumerator that wraps a callback.

Traits

BlockMayHaveObjects 🔒

Allow querying if a block may have objects. MarkSweepSpace and ImmixSpace use different Block types, and they have different block states. This trait lets both Block types provide the same may_have_objects method.

ObjectEnumerator

A trait for enumerating objects in spaces, used by Space::enumerate_objects.

Functions

enumerate_blocks_from_chunk_map 🔒

enumerate_blocks_from_monotonic_page_resource 🔒