Module mmtk::util::object_forwarding

Forwarding word in object copying.

Constants

BEING_FORWARDED

FORWARDED

FORWARDING_BITS

FORWARDING_MASK

FORWARDING_NOT_TRIGGERED_YET

FORWARDING_POINTER_MASK

Functions

attempt_to_forward
Attempt to become the worker thread who will forward the object. The successful worker will set the object forwarding bits to BEING_FORWARDED, preventing other workers from forwarding the same object.

clear_forwarding_bits
Zero the forwarding bits of an object. This function is used on new objects.

forward_object
Copy the object, mark it as forwarded and return a reference to the new object (copy).

forwarding_bits_offset_in_forwarding_pointer
(This function is only used internal to the util module)

is_forwarded

is_forwarded_or_being_forwarded

read_forwarding_pointer
Read the forwarding pointer of an object. This function is called on forwarded/being_forwarded objects.

spin_and_get_forwarded_object
Spin-wait for the object’s forwarding to become complete and then read the forwarding pointer to the new object.

state_is_being_forwarded

state_is_forwarded_or_being_forwarded

write_forwarding_pointer
Write the forwarding pointer of an object. This function is called on being_forwarded objects.