mmtk::util::object_forwarding
Function spin_and_get_forwarded_object

pub fn spin_and_get_forwarded_object<VM: VMBinding>(
    object: ObjectReference,
    forwarding_bits: u8,
) -> ObjectReference

Spin-wait for the object’s forwarding to become complete and then read the forwarding pointer to the new object.

Arguments:

object: the forwarded/being_forwarded object.

forwarding_bits: the last state of the forwarding bits before calling this function.

Returns a reference to the new object.