Struct mmtk::util::opaque_pointer::OpaquePointer

#[repr(transparent)]
pub struct OpaquePointer(*mut c_void);

OpaquePointer represents pointers that MMTk needs to know about but will not deferefence it. For example, a pointer to the thread or the thread local storage is an opaque pointer for MMTK. The type does not provide any method for dereferencing.

Tuple Fields
0: *mut c_void
Implementations

impl OpaquePointer

pub const UNINITIALIZED: Self = _

Represents an uninitialized value for OpaquePointer.

pub fn from_address(addr: Address) -> Self

Cast an Address type to an OpaquePointer.

pub fn to_address(self) -> Address

Cast the opaque pointer to an Address type.

pub fn is_null(self) -> bool

Is this opaque pointer null?