Struct mmtk::util::options::Options

pub struct Options {

    pub plan: MMTKOption<PlanSelector>,
    pub threads: MMTKOption<usize>,
    pub use_short_stack_scans: MMTKOption<bool>,
    pub use_return_barrier: MMTKOption<bool>,
    pub eager_complete_sweep: MMTKOption<bool>,
    pub ignore_system_gc: MMTKOption<bool>,
    pub nursery: MMTKOption<NurserySize>,
    pub full_heap_system_gc: MMTKOption<bool>,
    pub variable_size_heap: MMTKOption<bool>,
    pub no_finalizer: MMTKOption<bool>,
    pub no_reference_types: MMTKOption<bool>,
    pub nursery_zeroing: MMTKOption<NurseryZeroingOptions>,
    pub stress_factor: MMTKOption<usize>,
    pub analysis_factor: MMTKOption<usize>,
    pub precise_stress: MMTKOption<bool>,
    pub vm_space_start: MMTKOption<Address>,
    pub vm_space_size: MMTKOption<usize>,
    pub work_perf_events: MMTKOption<PerfEventOptions>,
    pub phase_perf_events: MMTKOption<PerfEventOptions>,
    pub thread_affinity: MMTKOption<AffinityKind>,
    pub gc_trigger: MMTKOption<GCTriggerSelector>,
    pub transparent_hugepages: MMTKOption<bool>,

}

Fields
plan: MMTKOption<PlanSelector>threads: MMTKOption<usize>use_short_stack_scans: MMTKOption<bool>use_return_barrier: MMTKOption<bool>eager_complete_sweep: MMTKOption<bool>ignore_system_gc: MMTKOption<bool>nursery: MMTKOption<NurserySize>full_heap_system_gc: MMTKOption<bool>variable_size_heap: MMTKOption<bool>no_finalizer: MMTKOption<bool>no_reference_types: MMTKOption<bool>nursery_zeroing: MMTKOption<NurseryZeroingOptions>stress_factor: MMTKOption<usize>analysis_factor: MMTKOption<usize>precise_stress: MMTKOption<bool>vm_space_start: MMTKOption<Address>vm_space_size: MMTKOption<usize>work_perf_events: MMTKOption<PerfEventOptions>phase_perf_events: MMTKOption<PerfEventOptions>thread_affinity: MMTKOption<AffinityKind>gc_trigger: MMTKOption<GCTriggerSelector>transparent_hugepages: MMTKOption<bool>
Implementations

impl Options

pub fn get_max_nursery_bytes(&self) -> usize

Return upper bound of the nursery size (in number of bytes)

pub fn get_max_nursery_pages(&self) -> usize

Return upper bound of the nursery size (in number of pages)

pub fn get_min_nursery_bytes(&self) -> usize

Return lower bound of the nursery size (in number of bytes)

pub fn get_min_nursery_pages(&self) -> usize

Return lower bound of the nursery size (in number of pages)

pub fn is_stress_test_gc_enabled(&self) -> bool

Check if the options are set for stress GC. If either stress_factor or analysis_factor is set, we should do stress GC.

impl Options

pub fn set_from_env_var(&mut self, s: &str, val: &str) -> bool

Set an option from env var

pub fn set_from_command_line(&mut self, s: &str, val: &str) -> bool

Set an option from command line

pub fn set_bulk_from_command_line(&mut self, options: &str) -> bool

Bulk process options. Returns true if all the options are processed successfully. This method returns false if the option string is invalid, or if it includes any invalid option.

Arguments:

options: a string that is key value pairs separated by white spaces or commas, e.g. threads=1 stress_factor=4096, or threads=1,stress_factor=4096

fn set_inner(&mut self, s: &str, val: &str) -> bool

Set an option and run its validator for its value.

fn new() -> Self

Create an Options instance with built-in default settings.

pub fn read_env_var_settings(&mut self)

Read options from environment variables, and apply those settings to self.

If we have environment variables that start with MMTK_ and match any option (such as MMTK_STRESS_FACTOR), we set the option to its value (if it is a valid value).