mmtk::util::os::memory

Trait OSMemory

pub trait OSMemory {
    // Required methods
    fn dzmmap(
        start: Address,
        size: usize,
        strategy: MmapStrategy,
        annotation: &MmapAnnotation<'_>,
    ) -> Result<Address>;
    fn is_mmap_oom(os_errno: i32) -> bool;
    fn munmap(start: Address, size: usize) -> Result<()>;
    fn set_memory_access(
        start: Address,
        size: usize,
        prot: MmapProtection,
    ) -> Result<()>;
    fn panic_if_unmapped(start: Address, size: usize);

    // Provided methods
    fn handle_mmap_error<VM: VMBinding>(
        error: Error,
        tls: VMThread,
        addr: Address,
        bytes: usize,
    ) { ... }
    fn get_system_total_memory() -> Result<u64> { ... }
}

Abstraction for OS memory operations.

Required Methods

fn dzmmap( start: Address, size: usize, strategy: MmapStrategy, annotation: &MmapAnnotation<'_>, ) -> Result<Address>

Perform a demand-zero mmap.

Fallback: annotation is only used for debugging. For platforms that do not support mmap annotations, this parameter can be ignored. Fallback: see crate::util::os::MmapStrategy for fallbacks for strategy.

fn is_mmap_oom(os_errno: i32) -> bool

Check whether the given OS error number indicates an out-of-memory condition.

fn munmap(start: Address, size: usize) -> Result<()>

Unmap a memory region.

fn set_memory_access( start: Address, size: usize, prot: MmapProtection, ) -> Result<()>

Change the protection of a memory region to the specified protection.

fn panic_if_unmapped(start: Address, size: usize)

Checks if the memory has already been mapped. If not, we panic.

Note that the checking may have a side effect that it will map the memory if it was unmapped. So we panic if it was unmapped. Be very careful about using this function.

Fallback: As the function is only used for assertions, it can be a no-op, and MMTk will still run and never panics in this function.

Provided Methods

fn handle_mmap_error<VM: VMBinding>( error: Error, tls: VMThread, addr: Address, bytes: usize, )

Handle mmap errors, possibly by signaling the VM about an out-of-memory condition.

fn get_system_total_memory() -> Result<u64>

Get the total memory of the system in bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.