mmtk::util
Module reference_processor

Reference processing implementation.

Structs

PhantomRefProcessing 🔒

RefEnqueue 🔒

RefForwarding 🔒

ReferenceProcessor

We create a reference processor for each semantics. Generally we expect these to happen for each processor:

ReferenceProcessorSync 🔒

ReferenceProcessors

Holds all reference processors for each weak reference Semantics. Currently this is based on Java’s weak reference semantics (soft/weak/phantom). We should make changes to make this general rather than Java specific.

RescanReferences 🔒

SoftRefProcessing 🔒

WeakRefProcessing 🔒

Enums

Semantics

Constants

INITIAL_SIZE 🔒