Struct mmtk::util::rust_util::InitializeOnce

pub struct InitializeOnce<T: 'static> {
    v: UnsafeCell<MaybeUninit<T>>,
    init_fn: &'static dyn Fn() -> T,
    once: Once,
}

InitializeOnce creates an uninitialized value that needs to be manually initialized later. InitializeOnce guarantees the value is only initialized once. This type is used to allow more efficient reads. Unlike the lazy_static! which checks whether the static is initialized in every read, InitializeOnce has no extra check for reads.

Fields
v: UnsafeCell<MaybeUninit<T>>init_fn: &'static dyn Fn() -> T

The function that is used to create the initialization value. This will be only called once.

once: Once

This is used to guarantee init_fn is only called once.

Implementations

impl<T> InitializeOnce<T>

pub const fn new(init_fn: &'static dyn Fn() -> T) -> Self

pub fn initialize_once(&self)

Initialize the value. This should be called before ever using the struct. If this method is called by multiple threads, the first thread will initialize the value, and the other threads will be blocked until the initialization is done (Once returns).

pub fn get_ref(&self) -> &T

Get the value. This should only be used after initialize_once()