mmtk::util

Module scanning_helper

This module provides some convenient functions for scanning objects.

Functions

visit_children

Visit and potentially update the children of object using Scanning::scan_object or Scanning::scan_object_and_trace_edges depending on the result of Scanning::support_slot_enqueuing.

visit_children_moving

Convenient wrapper of moving visit_children.

visit_children_non_moving

Convenient wrapper of non-moving visit_children.