mmtk::util
Module slot_logger

Logging slots to check duplicated edges in GC. This is a simple module to log slogs and check for duplicate slots.

It uses a hash-set to keep track of slots, and is so very expensive. We currently only use this as part of the extreme_assertions feature.

Structs

SlotLogger

Functions

should_check_duplicate_slots

Whether we should check duplicate slots. This depends on the actual plan.