mmtk::util::statistics

Module counter

Re-exports

pub use self::event_counter::EventCounter;

pub use self::long_counter::LongCounter;

pub use self::long_counter::Timer;

pub use self::size_counter::SizeCounter;

Modules

event_counter 🔒

long_counter 🔒

size_counter 🔒

Structs

MonotoneNanoTime

Traits

Counter

An abstraction over how a specific Diffable value is counted

Diffable

An abstraction over some changing values that we want to measure.