Trait mmtk::vm::collection::Collection

pub trait Collection<VM: VMBinding> {
    const COORDINATOR_ONLY_STW: bool;

    fn stop_all_mutators<E: ProcessEdgesWork<VM = VM>>(tls: VMWorkerThread);

    fn resume_mutators(tls: VMWorkerThread);

    fn block_for_gc(tls: VMMutatorThread);

    fn spawn_worker_thread(tls: VMThread, ctx: Option<&GCWorker<VM>>);

    fn prepare_mutator<T: MutatorContext<VM>>(
        tls_worker: VMWorkerThread, 
        tls_mutator: VMMutatorThread, 
        m: &T
    );

    fn out_of_memory(_tls: VMThread) { ... }

    fn schedule_finalization(_tls: VMWorkerThread) { ... }

    fn vm_release() { ... }

    fn process_weak_refs<E: ProcessEdgesWork<VM = VM>>(
        _worker: &mut GCWorker<VM>
    ) { ... }
}

VM-specific methods for garbage collection.

Associated Constants

const COORDINATOR_ONLY_STW: bool

If true, only the coordinator thread can call stop_all_mutators and the resume_mutators methods. If false, any GC thread can call these methods.

This constant exists because some VMs require the thread that resumes a thread to be the same thread that stopped it. The MMTk Core will use the appropriate thread to stop or start the world according to the value of this constant. If a VM does not have such a requirement, the VM binding shall set this to false to reduce an unnecessary context switch.

Required methods

fn stop_all_mutators<E: ProcessEdgesWork<VM = VM>>(tls: VMWorkerThread)

Stop all the mutator threads. MMTk calls this method when it requires all the mutator to yield for a GC. This method is called by a single thread in MMTk (the GC controller). This method should not return until all the threads are yielded. The actual thread synchronization mechanism is up to the VM, and MMTk does not make assumptions on that.

Arguments:

tls: The thread pointer for the GC controller/coordinator.

fn resume_mutators(tls: VMWorkerThread)

Resume all the mutator threads, the opposite of the above. When a GC is finished, MMTk calls this method.

Arguments:

tls: The thread pointer for the GC controller/coordinator.

fn block_for_gc(tls: VMMutatorThread)

Block the current thread for GC. This is called when an allocation request cannot be fulfilled and a GC is needed. MMTk calls this method to inform the VM that the current thread needs to be blocked as a GC is going to happen. Then MMTk starts a GC. For a stop-the-world GC, MMTk will then call stop_all_mutators() before the GC, and call resume_mutators() after the GC.

Arguments:

tls: The current thread pointer that should be blocked. The VM can optionally check if the current thread matches tls.

fn spawn_worker_thread(tls: VMThread, ctx: Option<&GCWorker<VM>>)

Ask the VM to spawn a GC thread for MMTk. A GC thread may later call into the VM through these VM traits. Some VMs have assumptions that those calls needs to be within VM internal threads. As a result, MMTk does not spawn GC threads itself to avoid breaking this kind of assumptions. MMTk calls this method to spawn GC threads during initialize_collection().

Arguments:

tls: The thread pointer for the parent thread that we spawn new threads from. This is the same tls when the VM calls initialize_collection() and passes as an argument.

ctx: The GC worker context for the GC thread. If None is passed, it means spawning a GC thread for the GC controller, which does not have a worker context.

fn prepare_mutator<T: MutatorContext<VM>>(
tls_worker: VMWorkerThread,
tls_mutator: VMMutatorThread,
m: &T
)

Allow VM-specific behaviors for a mutator after all the mutators are stopped and before any actual GC work starts.

Arguments:

tls_worker: The thread pointer for the worker thread performing this call.

tls_mutator: The thread pointer for the target mutator thread.

m: The mutator context for the thread.

Provided methods

fn out_of_memory(_tls: VMThread)

Inform the VM for an out-of-memory error. The VM can implement its own error routine for OOM. Note the VM needs to fail in this call. We do not expect the VM to resume in any way.

Arguments:

tls: The thread pointer for the mutator which failed the allocation and triggered the OOM.

fn schedule_finalization(_tls: VMWorkerThread)

Inform the VM to schedule finalization threads.

Arguments:

tls: The thread pointer for the current GC thread.

fn vm_release()

Inform the VM to do its VM-specific release work at the end of a GC.

fn process_weak_refs<E: ProcessEdgesWork<VM = VM>>(_worker: &mut GCWorker<VM>)

Delegate to the VM binding for reference processing.