Module mmtk::vm::edge_shape

Structs

AddressRangeIterator

Iterate edges within Range<Address>.

SimpleEdge

A simple edge implementation that represents a word-sized slot where an ObjectReference value is stored as is. It is the default edge type, and should be suitable for most VMs.

UnimplementedMemorySlice

Memory slice type with empty implementations. For VMs that do not use the memory slice type.

UnimplementedMemorySliceEdgeIterator

Edge iterator for UnimplementedMemorySlice.

Traits

Edge

An abstract edge. An edge holds an object reference. When we load from it, we get an ObjectReference; we can also store an ObjectReference into it.

MemorySlice

A abstract memory slice represents a piece of heap memory.