mmtk::vm::object_model::specs
Struct VMLocalForwardingPointerSpec

pub struct VMLocalForwardingPointerSpec(MetadataSpec);

1-word local metadata for spaces that may copy objects. This metadata has to be stored in the header. This metadata can be defined at a position within the object payload. As a forwarding pointer is only stored in dead objects which is not accessible by the language, it is okay that store a forwarding pointer overwrites object payload

Tuple Fields
0: MetadataSpec
Implementations

impl VMLocalForwardingPointerSpec

pub const LOG_NUM_BITS: usize = 6usize

The number of bits (in log2) that are needed for the spec.

pub const IS_GLOBAL: bool = false

Whether this spec is global or local. For side metadata, the binding needs to make sure global specs are laid out after another global spec, and local specs are laid out after another local spec. Otherwise, there will be an assertion failure.

pub const fn in_header(bit_offset: isize) -> Self

Declare that the VM uses in-header metadata for this metadata type. For the specification of the bit_offset argument, please refer to the document of [crate::util::metadata::header_metadata::HeaderMetadataSpec.bit_offset]. The binding needs to make sure that the bits used for a spec in the header do not conflict with the bits of another spec (unless it is specified that some bits may be reused).

pub const fn side_first() -> Self

Declare that the VM uses side metadata for this metadata type, and the side metadata is the first of its kind (global or local). The first global or local side metadata should be declared with side_first(), and the rest side metadata should be declared with side_after() after a defined side metadata of the same kind (global or local). Logically, all the declarations create two list of side metadata, one for global, and one for local.

pub const fn side_after(spec: &MetadataSpec) -> Self

Declare that the VM uses side metadata for this metadata type, and the side metadata should be laid out after the given side metadata spec. The first global or local side metadata should be declared with side_first(), and the rest side metadata should be declared with side_after() after a defined side metadata of the same kind (global or local). Logically, all the declarations create two list of side metadata, one for global, and one for local.

pub const fn as_spec(&self) -> &MetadataSpec

Return the inner [crate::util::metadata::MetadataSpec] for the metadata type.

pub const fn num_bits(&self) -> usize

Return the number of bits for the metadata type.