Trait mmtk::vm::reference_glue::ReferenceGlue

pub trait ReferenceGlue<VM: VMBinding> {
    fn get_referent(object: ObjectReference) -> ObjectReference;

    fn set_referent(reff: ObjectReference, referent: ObjectReference);

    fn process_reference<T: TraceLocal>(
        trace: &mut T, 
        reference: ObjectReference, 
        tls: VMWorkerThread
    ) -> ObjectReference;

    fn clear_referent(new_reference: ObjectReference) { ... }
}

VM-specific methods for reference processing.

Required methods

fn get_referent(object: ObjectReference) -> ObjectReference

Get the referent from a reference.

Arguments:

object: The object reference.

fn set_referent(reff: ObjectReference, referent: ObjectReference)

Set the referent in a reference.

Arguments:

reff: The object reference for the reference.

referent: The referent object reference.

fn process_reference<T: TraceLocal>(
trace: &mut T,
reference: ObjectReference,
tls: VMWorkerThread
) -> ObjectReference

Process a reference with the current semantics and return an updated reference (e.g. with a new address) if the reference is still alive, otherwise return a null object reference.

Arguments:

trace: A reference to a TraceLocal object for this reference.

reference: The address of the reference. This may or may not be the address of a heap object, depending on the VM.

tls: The GC thread that is processing this reference.

Loading content...
Provided methods

fn clear_referent(new_reference: ObjectReference)

Weak and soft references always clear the referent before enqueueing.

Arguments:

new_reference: The reference whose referent is to be cleared.

Loading content...