Trait mmtk::vm::ActivePlan

pub trait ActivePlan<VM: VMBinding> {
    fn global() -> &'static SelectedPlan<VM>;

    fn worker(tls: OpaquePointer) -> &'static mut GCWorker<VM>;

    unsafe fn is_mutator(tls: OpaquePointer) -> bool;

    unsafe fn mutator(
        tls: OpaquePointer
    ) -> &'static mut <SelectedPlan<VM> as Plan>::Mutator;

    fn collector_count() -> usize;

    fn reset_mutator_iterator();

    fn get_next_mutator(
    ) -> Option<&'static mut <SelectedPlan<VM> as Plan>::Mutator>;

    fn number_of_mutators() -> usize;

    fn mutators<'a>() -> SynchronizedMutatorIterator<'a, VM> { ... }
}

VM-specific methods for the current plan.

Required methods

fn global() -> &'static SelectedPlan<VM>

Return a reference to the current plan.

fn worker(tls: OpaquePointer) -> &'static mut GCWorker<VM>

Return a GCWorker reference for the thread.

Arguments:

tls: The thread to query.

unsafe fn is_mutator(tls: OpaquePointer) -> bool

Return whether there is a mutator created and associated with the thread.

Arguments:

tls: The thread to query.

Safety

TODO: I am not sure why this is unsafe.

unsafe fn mutator(
tls: OpaquePointer
) -> &'static mut <SelectedPlan<VM> as Plan>::Mutator

Return a Mutator reference for the thread.

Arguments:

tls: The thread to query.

Safety

TODO: I am not sure why this is unsafe.

fn collector_count() -> usize

👎 Deprecated

Return the number of GC collectors. This is unused by MMTk now, and will be removed.

fn reset_mutator_iterator()

Reset the mutator iterator so that get_next_mutator() returns the first mutator.

fn get_next_mutator(
) -> Option<&'static mut <SelectedPlan<VM> as Plan>::Mutator>

Return the next mutator if there is any. This method assumes that the VM implements stateful type to remember which mutator is returned and guarantees to return the next when called again. This does not need to be thread safe.

fn number_of_mutators() -> usize

Return the total count of mutators.

Loading content...
Provided methods

fn mutators<'a>() -> SynchronizedMutatorIterator<'a, VM>

A utility method to provide a thread-safe mutator iterator from reset_mutator_iterator() and get_next_mutator().

Loading content...