Trait mmtk::vm::VMBinding

pub trait VMBinding where
    Self: Sized + 'static + Send + Sync + Default, {
    type VMObjectModel: ObjectModel<Self>;
    type VMScanning: Scanning<Self>;
    type VMCollection: Collection<Self>;
    type VMActivePlan: ActivePlan<Self>;
    type VMReferenceGlue: ReferenceGlue<Self>;

    const ALIGNMENT_VALUE: usize;
    const LOG_MIN_ALIGNMENT: usize;
    const MIN_ALIGNMENT: usize;
    const MAX_ALIGNMENT_SHIFT: usize;
    const MAX_ALIGNMENT: usize;
    const ALLOC_END_ALIGNMENT: usize;
}

The VMBinding trait associates with each trait, and provides VM-specific constants.

Associated Types

type VMObjectModel: ObjectModel<Self>

type VMScanning: Scanning<Self>

type VMCollection: Collection<Self>

type VMActivePlan: ActivePlan<Self>

type VMReferenceGlue: ReferenceGlue<Self>

Associated Constants

const ALIGNMENT_VALUE: usize

A value to fill in alignment gaps. This value can be used for debugging.

const LOG_MIN_ALIGNMENT: usize

Allowed minimal alignment.

const MIN_ALIGNMENT: usize

Allowed minimal alignment in bytes.

const MAX_ALIGNMENT_SHIFT: usize

Allowed maximum alignment as shift by min alignment.

const MAX_ALIGNMENT: usize

Allowed maximum alignment in bytes.

const ALLOC_END_ALIGNMENT: usize

This value is used to assert if the cursor is reasonable after allocations. At the end of an allocation, the allocation cursor should be aligned to this value. Note that MMTk does not attempt to do anything to align the cursor to this value, but it merely asserts with this constant.