Crate mmtk

Memory Management ToolKit (MMTk) is a portable and high performance memory manager that includes various garbage collection algorithms and provides clean and efficient interfaces to cooperate with language implementations. MMTk features highly modular and highly reusable designs. It includes components such as allocators, spaces and work packets that GC implementers can choose from to compose their own GC plan easily.

Logically, this crate includes these major parts:

GC components:

Allocators: handlers of allocation requests which allocate objects to the bound space.

Policies: definitions of semantics and behaviors for memory regions. Each space is an instance of a policy, and takes up a unique proportion of the heap.

Work packets: units of GC works scheduled by the MMTk's scheduler.

GC plans: GC algorithms composed from components. Note that currently the choice of plans is made through Rust features, which is a build-time config, so only one plan is present in the generated binary and in this documentation. We plan to make this a run-time config so that users can choose GC plans at boot time.

Heap implementations: the underlying implementations of memory resources that support spaces.

Scheduler: the MMTk scheduler to allow flexible and parallel execution of GC works.

Interfaces: bi-directional interfaces between MMTk and language implementations i.e. the memory manager API that allows a language's memory manager to use MMTk and the VMBinding trait that allows MMTk to call the language implementation.

Re-exports

pub use crate::plan::AllocationSemantics;

pub use crate::plan::CopyContext;

pub use crate::plan::Mutator;

pub use crate::plan::MutatorContext;

pub use crate::plan::Plan;

pub use crate::plan::TraceLocal;

pub use crate::plan::TransitiveClosure;

Modules

memory_manager
VM-to-MMTk interface: safe Rust APIs.

plan

policy
Memory policies that can be used for spaces.

scheduler
A general scheduler implementation. MMTk uses it to schedule GC-related works.

util
Utilities used by other modules, including allocators, heap implementation, etc.

vm
MMTk-to-VM interface: the VMBinding trait.

Structs

MMTK
An MMTk instance. MMTk allows mutiple instances to run independently, and each instance gives users a separate heap. Note that multi-instances is not fully supported yet