mmtk
Module build_info

Some information for the current MMTk build.

Modules

raw 🔒

Structs

MMTK_FULL_BUILD_INFO

Full build info, including MMTk’s name, version, git, and features in the build, such as MMTk 0.14.0 (43e0ce8-dirty, DEFAULT, EXTREME_ASSERTIONS)

MMTK_FULL_BUILD_INFO_STRING 🔒

MMTK_GIT_VERSION

Git version as short commit hash, such as a96e8f9, or a96e8f9-dirty, or unknown-git-version if MMTk is not built from a git repo.

MMTK_GIT_VERSION_STRING 🔒

Statics

MMTK_FEATURES

Comma separated features enabled for this build The features as a comma-separated string.

MMTK_PKG_VERSION

MMTk crate version such as 0.14.0 The full version.