mmtk::plan
Module barriers

Read/Write barrier implementations.

Structs

NoBarrier

Empty barrier implementation. For GCs that do not need any barriers

ObjectBarrier

Generic object barrier with a type argument defining it’s slow-path behaviour.

Enums

BarrierSelector

BarrierSelector describes which barrier to use.

Traits

Barrier

A barrier is a combination of fast-path behaviour + slow-path semantics. This trait exposes generic barrier interfaces. The implementations will define their own fast-path code and slow-path semantics.

BarrierSemantics

A barrier semantics defines the barrier slow-path behaviour. For example, how an object barrier processes it’s modbufs. Specifically, it defines the slow-path call interfaces and a call to flush buffers.