Struct mmtk::plan::global::BasePlan

pub struct BasePlan<VM: VMBinding> {

    pub initialized: AtomicBool,
    pub gc_status: Mutex<GcStatus>,
    pub last_stress_pages: AtomicUsize,
    pub stacks_prepared: AtomicBool,
    pub emergency_collection: AtomicBool,
    pub user_triggered_collection: AtomicBool,
    pub allocation_success: AtomicBool,
    pub max_collection_attempts: AtomicUsize,
    pub cur_collection_attempts: AtomicUsize,
    pub control_collector_context: ControllerCollectorContext<VM>,
    pub stats: Stats,
    mmapper: &'static FragmentedMapper,
    pub vm_map: &'static Map64,
    pub options: Arc<UnsafeOptionsWrapper>,
    pub heap: HeapMeta,
    pub inside_sanity: AtomicBool,
    pub scanned_stacks: AtomicUsize,
    pub mutator_iterator_lock: Mutex<()>,
    pub allocation_bytes: AtomicUsize,
    pub analysis_manager: AnalysisManager<VM>,
    pub code_space: ImmortalSpace<VM>,
    pub code_lo_space: ImmortalSpace<VM>,
    pub ro_space: ImmortalSpace<VM>,
    pub vm_space: ImmortalSpace<VM>,

}

BasePlan should contain all plan-related state and functions that are fundamental to all plans. These include VM-specific (but not plan-specific) features such as a code space or vm space, which are fundamental to all plans for a given VM. Features that are common to many (but not intrinsically all) plans should instead be included in CommonPlan.

Fields
initialized: AtomicBoolgc_status: Mutex<GcStatus>last_stress_pages: AtomicUsizestacks_prepared: AtomicBoolemergency_collection: AtomicBooluser_triggered_collection: AtomicBoolallocation_success: AtomicBoolmax_collection_attempts: AtomicUsizecur_collection_attempts: AtomicUsizecontrol_collector_context: ControllerCollectorContext<VM>stats: Statsmmapper: &'static FragmentedMappervm_map: &'static Map64options: Arc<UnsafeOptionsWrapper>heap: HeapMetainside_sanity: AtomicBoolscanned_stacks: AtomicUsizemutator_iterator_lock: Mutex<()>allocation_bytes: AtomicUsizeanalysis_manager: AnalysisManager<VM>code_space: ImmortalSpace<VM>code_lo_space: ImmortalSpace<VM>ro_space: ImmortalSpace<VM>vm_space: ImmortalSpace<VM>
Implementations

impl<VM: VMBinding> BasePlan<VM>

pub fn new(
vm_map: &'static Map64,
mmapper: &'static FragmentedMapper,
options: Arc<UnsafeOptionsWrapper>,
heap: HeapMeta,
constraints: &'static PlanConstraints,
global_side_metadata_specs: Vec<SideMetadataSpec>
) -> BasePlan<VM>

pub fn gc_init(
&mut self,
heap_size: usize,
vm_map: &'static Map64,
scheduler: &Arc<Scheduler<MMTK<VM>>>
)

pub fn get_pages_used(&self) -> usize

pub fn trace_object<T: TransitiveClosure, C: CopyContext>(
&self,
_trace: &mut T,
_object: ObjectReference
) -> ObjectReference

pub fn prepare(&mut self, _tls: VMWorkerThread, _primary: bool)

pub fn release(&mut self, _tls: VMWorkerThread, _primary: bool)

pub fn set_collection_kind(&self)

pub fn set_gc_status(&self, s: GcStatus)

pub fn stacks_prepared(&self) -> bool

pub fn gc_in_progress(&self) -> bool

pub fn gc_in_progress_proper(&self) -> bool

pub fn is_user_triggered_collection(&self) -> bool

fn determine_collection_attempts(&self) -> usize

fn is_internal_triggered_collection(&self) -> bool

fn last_collection_was_exhaustive(&self) -> bool

fn force_full_heap_collection(&self)

pub fn increase_allocation_bytes_by(&self, size: usize)

pub(super) fn stress_test_gc_required(&self) -> bool

pub(super) fn collection_required<P: Plan>(
&self,
plan: &P,
space_full: bool,
_space: &dyn Space<VM>
) -> bool

pub(crate) fn verify_side_metadata_sanity(
&self,
side_metadata_sanity_checker: &mut SideMetadataSanity
)