Struct mmtk::plan::global::BasePlan

pub struct BasePlan<VM: VMBinding> {
    pub initialized: AtomicBool,
    pub gc_status: Mutex<GcStatus>,
    pub last_stress_pages: AtomicUsize,
    pub stacks_prepared: AtomicBool,
    pub emergency_collection: AtomicBool,
    pub user_triggered_collection: AtomicBool,
    pub allocation_success: AtomicBool,
    pub max_collection_attempts: AtomicUsize,
    pub cur_collection_attempts: AtomicUsize,
    pub oom_lock: Mutex<()>,
    pub control_collector_context: ControllerCollectorContext<VM>,
    pub stats: Stats,
    pub vm_map: &'static VMMap,
    pub options: Arc<UnsafeOptionsWrapper>,
    pub heap: HeapMeta,
    pub unsync: UnsafeCell<BaseUnsync<VM>>,
    pub inside_sanity: AtomicBool,
    pub scanned_stacks: AtomicUsize,
    pub mutator_iterator_lock: Mutex<()>,
    pub allocation_bytes: AtomicUsize,
    pub analysis_manager: AnalysisManager<VM>,
    // some fields omitted
}

BasePlan should contain all plan-related state and functions that are fundamental to all plans. These include VM-specific (but not plan-specific) features such as a code space or vm space, which are fundamental to all plans for a given VM. Features that are common to many (but not intrinsically all) plans should instead be included in CommonPlan.

Fields
initialized: AtomicBoolgc_status: Mutex<GcStatus>last_stress_pages: AtomicUsizestacks_prepared: AtomicBoolemergency_collection: AtomicBooluser_triggered_collection: AtomicBoolallocation_success: AtomicBoolmax_collection_attempts: AtomicUsizecur_collection_attempts: AtomicUsizeoom_lock: Mutex<()>control_collector_context: ControllerCollectorContext<VM>stats: Statsvm_map: &'static VMMapoptions: Arc<UnsafeOptionsWrapper>heap: HeapMetaunsync: UnsafeCell<BaseUnsync<VM>>inside_sanity: AtomicBoolscanned_stacks: AtomicUsizemutator_iterator_lock: Mutex<()>allocation_bytes: AtomicUsizeanalysis_manager: AnalysisManager<VM>
Implementations

impl<VM: VMBinding> BasePlan<VM>

pub fn new(
vm_map: &'static VMMap,
mmapper: &'static Mmapper,
options: Arc<UnsafeOptionsWrapper>,
heap: HeapMeta,
constraints: &'static PlanConstraints
) -> BasePlan<VM>

pub fn gc_init(
&mut self,
heap_size: usize,
vm_map: &'static VMMap,
scheduler: &Arc<MMTkScheduler<VM>>
)

pub fn get_pages_used(&self) -> usize

pub fn trace_object<T: TransitiveClosure, C: CopyContext>(
&self,
_trace: &mut T,
_object: ObjectReference
) -> ObjectReference

pub fn prepare(&self, _tls: OpaquePointer, _primary: bool)

pub fn release(&self, _tls: OpaquePointer, _primary: bool)

pub fn set_collection_kind(&self)

pub fn set_gc_status(&self, s: GcStatus)

pub fn stacks_prepared(&self) -> bool

pub fn gc_in_progress(&self) -> bool

pub fn gc_in_progress_proper(&self) -> bool

pub fn increase_allocation_bytes_by(&self, size: usize)