mmtk::plan::global
Struct BasePlan

pub struct BasePlan<VM: VMBinding> {
    pub(crate) global_state: Arc<GlobalState>,
    pub options: Arc<Options>,
    pub gc_trigger: Arc<GCTrigger<VM>>,
    pub scheduler: Arc<GCWorkScheduler<VM>>,
    pub code_space: ImmortalSpace<VM>,
    pub code_lo_space: ImmortalSpace<VM>,
    pub ro_space: ImmortalSpace<VM>,
    pub vm_space: VMSpace<VM>,
}

BasePlan should contain all plan-related state and functions that are fundamental to all plans. These include VM-specific (but not plan-specific) features such as a code space or vm space, which are fundamental to all plans for a given VM. Features that are common to many (but not intrinsically all) plans should instead be included in CommonPlan.

Fields
global_state: Arc<GlobalState>options: Arc<Options>gc_trigger: Arc<GCTrigger<VM>>scheduler: Arc<GCWorkScheduler<VM>>code_space: ImmortalSpace<VM>code_lo_space: ImmortalSpace<VM>ro_space: ImmortalSpace<VM>vm_space: VMSpace<VM>

A VM space is a space allocated and populated by the VM. Currently it is used by JikesRVM for boot image.

If VM space is present, it has some special interaction with the memory_manager::is_mmtk_object and the memory_manager::is_in_mmtk_spaces functions.

The functions is_mmtk_object and find_object_from_internal_pointer require the valid object (VO) bit side metadata to identify objects. If the binding maintains the VO bit for objects in VM spaces, those functions will work accordingly. Otherwise, calling them is undefined behavior.

The is_in_mmtk_spaces currently returns true if the given object reference is in the VM space.

Implementations

impl<VM: VMBinding> BasePlan<VM>

pub fn new(args: CreateSpecificPlanArgs<'_, VM>) -> BasePlan<VM>

pub fn get_used_pages(&self) -> usize

pub fn prepare(&mut self, _tls: VMWorkerThread, _full_heap: bool)

pub fn release(&mut self, _tls: VMWorkerThread, _full_heap: bool)

pub fn clear_side_log_bits(&self)

pub fn set_side_log_bits(&self)

pub fn end_of_gc(&mut self, _tls: VMWorkerThread)

pub(crate) fn collection_required<P: Plan>( &self, plan: &P, space_full: bool, ) -> bool