Struct mmtk::plan::global::CommonPlan

pub struct CommonPlan<VM: VMBinding> {
    pub immortal: ImmortalSpace<VM>,
    pub los: LargeObjectSpace<VM>,
    pub nonmoving: ImmortalSpace<VM>,
    pub base: BasePlan<VM>,
}

CommonPlan is for representing state and features used by many plans, but that are not fundamental to all plans. Examples include the Large Object Space and an Immortal space. Features that are fundamental to all plans must be included in BasePlan.

Fields
immortal: ImmortalSpace<VM>los: LargeObjectSpace<VM>nonmoving: ImmortalSpace<VM>base: BasePlan<VM>
Implementations

impl<VM: VMBinding> CommonPlan<VM>

pub fn new(args: CreateSpecificPlanArgs<VM>) -> CommonPlan<VM>

pub fn get_spaces(&self) -> Vec<&dyn Space<VM>>

pub fn get_used_pages(&self) -> usize

pub fn trace_object<Q: ObjectQueue>( &self, queue: &mut Q, object: ObjectReference, worker: &mut GCWorker<VM> ) -> ObjectReference

pub fn prepare(&mut self, tls: VMWorkerThread, full_heap: bool)

pub fn release(&mut self, tls: VMWorkerThread, full_heap: bool)

pub fn stacks_prepared(&self) -> bool

pub fn get_immortal(&self) -> &ImmortalSpace<VM>

pub fn get_los(&self) -> &LargeObjectSpace<VM>

pub fn get_nonmoving(&self) -> &ImmortalSpace<VM>

pub(crate) fn verify_side_metadata_sanity( &self, side_metadata_sanity_checker: &mut SideMetadataSanity )