Trait mmtk::plan::global::Plan

pub trait Plan: 'static + Sync + Send + Downcast {
    type VM: VMBinding;
    pub fn constraints(&self) -> &'static PlanConstraints;

    pub fn create_worker_local(
        &self, 
        tls: OpaquePointer, 
        mmtk: &'static MMTK<Self::VM>
    ) -> GCWorkerLocalPtr;

    pub fn base(&self) -> &BasePlan<Self::VM>;

    pub fn schedule_collection(
        &'static self, 
        _scheduler: &MMTkScheduler<Self::VM>
    );

    pub fn gc_init(
        &mut self, 
        heap_size: usize, 
        vm_map: &'static VMMap, 
        scheduler: &Arc<MMTkScheduler<Self::VM>>
    );

    pub fn get_allocator_mapping(
        &self
    ) -> &'static EnumMap<AllocationSemantics, AllocatorSelector>;

    pub fn prepare(&self, tls: OpaquePointer);

    pub fn release(&self, tls: OpaquePointer);

    pub fn get_pages_used(&self) -> usize;

    pub fn common(&self) -> &CommonPlan<Self::VM> { ... }

    pub fn mmapper(&self) -> &'static Mmapper { ... }

    pub fn options(&self) -> &Options { ... }

    pub fn in_nursery(&self) -> bool { ... }

    pub fn enter_sanity(&self) { ... }

    pub fn leave_sanity(&self) { ... }

    pub fn is_in_sanity(&self) -> bool { ... }

    pub fn is_initialized(&self) -> bool { ... }

    pub fn poll(&self, space_full: bool, space: &dyn Space<Self::VM>) -> bool { ... }

    pub fn log_poll(&self, space: &dyn Space<Self::VM>, message: &'static str) { ... }

    pub fn collection_required(
        &self, 
        space_full: bool, 
        _space: &dyn Space<Self::VM>
    ) -> bool { ... }

    pub fn get_pages_reserved(&self) -> usize { ... }

    pub fn get_total_pages(&self) -> usize { ... }

    pub fn get_pages_avail(&self) -> usize { ... }

    pub fn get_collection_reserve(&self) -> usize { ... }

    pub fn is_emergency_collection(&self) -> bool { ... }

    pub fn get_free_pages(&self) -> usize { ... }

    pub fn stress_test_gc_required(&self) -> bool { ... }

    pub fn handle_user_collection_request(
        &self, 
        tls: OpaquePointer, 
        force: bool
    ) { ... }

    pub fn reset_collection_trigger(&self) { ... }

    pub fn modify_check(&self, object: ObjectReference) { ... }

    pub fn global_side_metadata_per_chunk(&self) -> usize { ... }
}

A plan describes the global core functionality for all memory management schemes. All global MMTk plans should implement this trait.

The global instance defines and manages static resources (such as memory and virtual memory resources).

Associated Types

type VM: VMBinding

Loading content...
Required methods

pub fn constraints(&self) -> &'static PlanConstraints

pub fn create_worker_local(
&self,
tls: OpaquePointer,
mmtk: &'static MMTK<Self::VM>
) -> GCWorkerLocalPtr

pub fn base(&self) -> &BasePlan<Self::VM>

pub fn schedule_collection(&'static self, _scheduler: &MMTkScheduler<Self::VM>)

pub fn gc_init(
&mut self,
heap_size: usize,
vm_map: &'static VMMap,
scheduler: &Arc<MMTkScheduler<Self::VM>>
)

pub fn get_allocator_mapping(
&self
) -> &'static EnumMap<AllocationSemantics, AllocatorSelector>

pub fn prepare(&self, tls: OpaquePointer)

pub fn release(&self, tls: OpaquePointer)

pub fn get_pages_used(&self) -> usize

Loading content...
Provided methods

pub fn common(&self) -> &CommonPlan<Self::VM>

pub fn mmapper(&self) -> &'static Mmapper

pub fn options(&self) -> &Options

pub fn in_nursery(&self) -> bool

pub fn enter_sanity(&self)

pub fn leave_sanity(&self)

pub fn is_in_sanity(&self) -> bool

pub fn is_initialized(&self) -> bool

pub fn poll(&self, space_full: bool, space: &dyn Space<Self::VM>) -> bool

pub fn log_poll(&self, space: &dyn Space<Self::VM>, message: &'static str)

pub fn collection_required(
&self,
space_full: bool,
_space: &dyn Space<Self::VM>
) -> bool

This method controls the triggering of a GC. It is called periodically during allocation. Returns true to trigger a collection.

@param spaceFull Space request failed, must recover pages within 'space'. @param space TODO @return true if a collection is requested by the plan.

pub fn get_pages_reserved(&self) -> usize

pub fn get_total_pages(&self) -> usize

pub fn get_pages_avail(&self) -> usize

pub fn get_collection_reserve(&self) -> usize

pub fn is_emergency_collection(&self) -> bool

pub fn get_free_pages(&self) -> usize

pub fn stress_test_gc_required(&self) -> bool

pub fn handle_user_collection_request(&self, tls: OpaquePointer, force: bool)

pub fn reset_collection_trigger(&self)

pub fn modify_check(&self, object: ObjectReference)

pub fn global_side_metadata_per_chunk(&self) -> usize

Loading content...
Implementations

impl<VM> dyn Plan<VM = VM> where
VM: Any + 'static,

pub fn is<__T: Plan<VM = VM>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Plan<VM = VM>>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: Plan<VM = VM>>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: Plan<VM = VM>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: Plan<VM = VM>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.