mmtk::plan

Module tracing

This module contains code useful for tracing, i.e. visiting the reachable objects by traversing all or part of an object graph.

Structs

ObjectsClosure

A transitive closure visitor to collect the slots from objects. It maintains a buffer for the slots, and flushes slots to a new work packet if the buffer is full or if the type gets dropped.

SlotIterator 🔒

For iterating over the slots of an object.

VectorQueue

An implementation of ObjectQueue using a Vec.

Traits

ObjectQueue

This trait represents an object queue to enqueue objects during tracing.

Type Aliases

VectorObjectQueue

A vector queue for object references.