Trait mmtk::plan::Plan

pub trait Plan: 'static + Sync + Downcast {
    type VM: VMBinding;

    fn constraints(&self) -> &'static PlanConstraints;

    fn create_worker_local(
        &self, 
        tls: VMWorkerThread, 
        mmtk: &'static MMTK<Self::VM>
    ) -> WorkerLocalPtr;

    fn base(&self) -> &BasePlan<Self::VM>;

    fn schedule_collection(
        &'static self, 
        _scheduler: &Scheduler<MMTK<Self::VM>>
    );

    fn gc_init(
        &mut self, 
        heap_size: usize, 
        vm_map: &'static Map64, 
        scheduler: &Arc<Scheduler<MMTK<Self::VM>>>
    );

    fn get_allocator_mapping(
        &self
    ) -> &'static EnumMap<AllocationSemantics, AllocatorSelector>;

    fn prepare(&mut self, tls: VMWorkerThread);

    fn release(&mut self, tls: VMWorkerThread);

    fn collection_required(
        &self, 
        space_full: bool, 
        _space: &dyn Space<Self::VM>
    ) -> bool;

    fn get_pages_used(&self) -> usize;

    fn common(&self) -> &CommonPlan<Self::VM> { ... }

    fn mmapper(&self) -> &'static FragmentedMapper { ... }

    fn options(&self) -> &Options { ... }

    fn is_current_gc_nursery(&self) -> bool { ... }

    fn enter_sanity(&self) { ... }

    fn leave_sanity(&self) { ... }

    fn is_in_sanity(&self) -> bool { ... }

    fn is_initialized(&self) -> bool { ... }

    fn poll(&self, space_full: bool, space: &dyn Space<Self::VM>) -> bool { ... }

    fn log_poll(&self, space: &dyn Space<Self::VM>, message: &'static str) { ... }

    fn get_pages_reserved(&self) -> usize { ... }

    fn get_total_pages(&self) -> usize { ... }

    fn get_pages_avail(&self) -> usize { ... }

    fn get_collection_reserve(&self) -> usize { ... }

    fn is_emergency_collection(&self) -> bool { ... }

    fn get_free_pages(&self) -> usize { ... }

    fn handle_user_collection_request(&self, tls: VMMutatorThread, force: bool) { ... }

    fn reset_collection_trigger(&self) { ... }

    fn modify_check(&self, object: ObjectReference) { ... }

}

A plan describes the global core functionality for all memory management schemes. All global MMTk plans should implement this trait.

The global instance defines and manages static resources (such as memory and virtual memory resources).

Associated Types

type VM: VMBinding

Loading content...
Required methods

fn constraints(&self) -> &'static PlanConstraints

fn create_worker_local(
&self,
tls: VMWorkerThread,
mmtk: &'static MMTK<Self::VM>
) -> WorkerLocalPtr

fn base(&self) -> &BasePlan<Self::VM>

fn schedule_collection(&'static self, _scheduler: &Scheduler<MMTK<Self::VM>>)

fn gc_init(
&mut self,
heap_size: usize,
vm_map: &'static Map64,
scheduler: &Arc<Scheduler<MMTK<Self::VM>>>
)

fn get_allocator_mapping(
&self
) -> &'static EnumMap<AllocationSemantics, AllocatorSelector>

fn prepare(&mut self, tls: VMWorkerThread)

fn release(&mut self, tls: VMWorkerThread)

fn collection_required(
&self,
space_full: bool,
_space: &dyn Space<Self::VM>
) -> bool

This method controls the triggering of a GC. It is called periodically during allocation. Returns true to trigger a collection.

@param spaceFull Space request failed, must recover pages within ‘space’. @param space TODO @return true if a collection is requested by the plan.

fn get_pages_used(&self) -> usize

Loading content...
Provided methods

fn common(&self) -> &CommonPlan<Self::VM>

fn mmapper(&self) -> &'static FragmentedMapper

fn options(&self) -> &Options

fn is_current_gc_nursery(&self) -> bool

Is current GC only collecting objects allocated since last GC?

fn enter_sanity(&self)

fn leave_sanity(&self)

fn is_in_sanity(&self) -> bool

fn is_initialized(&self) -> bool

fn poll(&self, space_full: bool, space: &dyn Space<Self::VM>) -> bool

fn log_poll(&self, space: &dyn Space<Self::VM>, message: &'static str)

fn get_pages_reserved(&self) -> usize

fn get_total_pages(&self) -> usize

fn get_pages_avail(&self) -> usize

fn get_collection_reserve(&self) -> usize

fn is_emergency_collection(&self) -> bool

fn get_free_pages(&self) -> usize

fn handle_user_collection_request(&self, tls: VMMutatorThread, force: bool)

fn reset_collection_trigger(&self)

fn modify_check(&self, object: ObjectReference)

Loading content...
Implementations

impl<VM> dyn Plan<VM = VM> where
VM: Any + 'static,

pub fn is<__T: Plan<VM = VM>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Plan<VM = VM>>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Plan<VM = VM>>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Plan<VM = VM>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Plan<VM = VM>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.