mmtk

Module policy

Memory policies that can be used for spaces.

Modules

compressor

copy_context

Copy context defines the thread local copy allocator for copying policies.

copyspace

gc_work

Policy specific GC work

immix

immortalspace

largeobjectspace

lockfreeimmortalspace

markcompactspace

marksweepspace

Mark sweep space. MMTk provides two implementations of mark sweep:

sft

sft_map

space

This class defines and manages spaces. Each policy is an instance of a space. A space is a region of virtual memory (contiguous or discontigous) which is subject to the same memory management regime. Multiple spaces (instances of this class or its descendants) may have the same policy (eg there could be numerous instances of CopySpace, each with different roles). Spaces are defined in terms of a unique region of virtual memory, so no two space instances ever share any virtual memory.

vmspace