mmtk::policy::immix::defrag
Struct Defrag

pub struct Defrag {
    in_defrag_collection: AtomicBool,
    defrag_space_exhausted: AtomicBool,
    pub mark_histograms: Mutex<Vec<[usize; 17]>>,
    pub defrag_spill_threshold: AtomicUsize,
    available_clean_pages_for_defrag: AtomicUsize,
}

Fields
in_defrag_collection: AtomicBool

Is current GC a defrag GC?

defrag_space_exhausted: AtomicBool

Is defrag space exhausted?

mark_histograms: Mutex<Vec<[usize; 17]>>

A list of completed mark histograms reported by workers

defrag_spill_threshold: AtomicUsize

A block with number of holes greater than this threshold will be defragmented.

available_clean_pages_for_defrag: AtomicUsize

The number of remaining clean pages in defrag space.

Implementations

impl Defrag

const NUM_BINS: usize = 17usize

const DEFRAG_LINE_REUSE_RATIO: f32 = 0.990000009f32

const MIN_SPILL_THRESHOLD: usize = 2usize

pub const fn new_histogram(&self) -> [usize; 17]

Allocate a new local histogram.

pub fn add_completed_mark_histogram(&self, histogram: [usize; 17])

Report back a completed mark histogram

pub fn in_defrag(&self) -> bool

Check if the current GC is a defrag GC.

pub fn decide_whether_to_defrag( &self, defrag_enabled: bool, emergency_collection: bool, collect_whole_heap: bool, collection_attempts: usize, user_triggered: bool, exhausted_reusable_space: bool, full_heap_system_gc: bool, stress_defrag: bool, )

Determine whether the current GC should do defragmentation.

pub fn defrag_headroom_pages<VM: VMBinding>( &self, space: &ImmixSpace<VM>, ) -> usize

Get the number of defrag headroom pages.

pub fn space_exhausted(&self) -> bool

Check if the defrag space is exhausted.

pub fn notify_new_clean_block(&self, copy: bool)

Update available_clean_pages_for_defrag counter when a clean block is allocated.

pub fn prepare<VM: VMBinding>( &self, space: &ImmixSpace<VM>, plan_stats: StatsForDefrag, )

Prepare work. Should be called in ImmixSpace::prepare.

fn get_available_lines<VM: VMBinding>( &self, space: &ImmixSpace<VM>, spill_avail_histograms: &mut [usize; 17], ) -> usize

Get the numebr of all the recyclable lines in all the reusable blocks.

fn establish_defrag_spill_threshold<VM: VMBinding>( &self, space: &ImmixSpace<VM>, )

Calculate the defrag threshold.

pub fn reset_in_defrag(&self)

Reset the in-defrag state.