Struct mmtk::policy::largeobjectspace::LargeObjectSpace

pub struct LargeObjectSpace<VM: VMBinding> {
    common: CommonSpace<VM>,
    pr: FreeListPageResource<VM>,
    mark_state: u8,
    in_nursery_gc: bool,
    treadmill: TreadMill,
}

This type implements a policy for large objects. Each instance corresponds to one Treadmill space.

Fields
common: CommonSpace<VM>pr: FreeListPageResource<VM>mark_state: u8in_nursery_gc: booltreadmill: TreadMill
Implementations

impl<VM: VMBinding> LargeObjectSpace<VM>

pub fn new( args: PlanCreateSpaceArgs<'_, VM>, protect_memory_on_release: bool ) -> Self

pub fn prepare(&mut self, full_heap: bool)

pub fn release(&mut self, full_heap: bool)

pub fn trace_object<Q: ObjectQueue>( &self, queue: &mut Q, object: ObjectReference ) -> ObjectReference

fn sweep_large_pages(&mut self, sweep_nursery: bool)

pub fn allocate_pages(&self, tls: VMThread, pages: usize) -> Address

Allocate an object

fn test_and_mark(&self, object: ObjectReference, value: u8) -> bool

Test if the object’s mark bit is the same as the given value. If it is not the same, the method will attemp to mark the object and clear its nursery bit. If the attempt succeeds, the method will return true, meaning the object is marked by this invocation. Otherwise, it returns false.

fn test_mark_bit(&self, object: ObjectReference, value: u8) -> bool

fn is_in_nursery(&self, object: ObjectReference) -> bool

Check if a given object is in nursery