mmtk::policy::markcompactspace
Struct MarkCompactSpace

pub struct MarkCompactSpace<VM: VMBinding> {
    common: CommonSpace<VM>,
    pr: MonotonePageResource<VM>,
}

Fields
common: CommonSpace<VM>pr: MonotonePageResource<VM>
Implementations

impl<VM: VMBinding> MarkCompactSpace<VM>

pub const HEADER_RESERVED_IN_BYTES: usize = _

We need one extra header word for each object. Considering the alignment requirement, this is the actual bytes we need to reserve for each allocation.

fn header_forwarding_pointer_address(object: ObjectReference) -> Address

Get the address for header forwarding pointer

fn get_header_forwarding_pointer( object: ObjectReference, ) -> Option<ObjectReference>

Get header forwarding pointer for an object

fn store_header_forwarding_pointer( object: ObjectReference, forwarding_pointer: ObjectReference, )

Store header forwarding pointer for an object

fn clear_header_forwarding_pointer(object: ObjectReference)

pub fn new(args: PlanCreateSpaceArgs<'_, VM>) -> Self

pub fn prepare(&self)

pub fn release(&self)

pub fn trace_mark_object<Q: ObjectQueue>( &self, queue: &mut Q, object: ObjectReference, ) -> ObjectReference

pub fn trace_forward_object<Q: ObjectQueue>( &self, queue: &mut Q, object: ObjectReference, ) -> ObjectReference

pub fn test_and_mark(object: ObjectReference) -> bool

pub fn test_and_clear_mark(object: ObjectReference) -> bool

pub fn is_marked(object: ObjectReference) -> bool

fn to_be_compacted(object: &ObjectReference) -> bool

fn linear_scan_objects( &self, range: Range<Address>, ) -> impl Iterator<Item = ObjectReference>

Linear scan all the live objects in the given memory region

pub fn calculate_forwarding_pointer(&self)

pub fn compact(&self)