Struct mmtk::policy::marksweepspace::native_ms::global::MarkSweepSpace

pub struct MarkSweepSpace<VM: VMBinding> {
    pub common: CommonSpace<VM>,
    pr: BlockPageResource<VM, Block>,
    chunk_map: ChunkMap,
    scheduler: Arc<GCWorkScheduler<VM>>,
    abandoned: Mutex<AbandonedBlockLists>,
    abandoned_in_gc: Mutex<AbandonedBlockLists>,
    pending_release_packets: AtomicUsize,
}

A mark sweep space.

The space and each free list allocator own some block lists. A block that is in use belongs to exactly one of the block lists. In this case, whoever owns a block list has exclusive access on the blocks in the list. There should be no data race to access blocks. A thread should NOT access a block list if it does not own the block list.

The table below roughly describes what we do in each phase.

Phase	Allocator local block lists	Global abandoned block lists	Chunk map

Allocation	Alloc from local	Move blocks from global to local block lists	-

Lazy: sweep local blocks

GC - Prepare	-	-	Find used chunks, reset block mark, bzero mark bit

GC - Trace	Trace object and mark blocks.	Trace object and mark blocks.	-

No block list access.	No block list access.

GC - Release	Lazy: Move blocks to local unswept list	Lazy: Move blocks to global unswept list	_

Eager: Sweep local blocks	Eager: Sweep global blocks

Both: Return local blocks to a temp global list

GC - End of GC	-	Merge the temp global lists	-

Fields
common: CommonSpace<VM>pr: BlockPageResource<VM, Block>chunk_map: ChunkMap

Allocation status for all chunks in MS space

scheduler: Arc<GCWorkScheduler<VM>>

Work packet scheduler

abandoned: Mutex<AbandonedBlockLists>

Abandoned blocks. If a mutator dies, all its blocks go to this abandoned block lists. We reuse blocks in these lists in the mutator phase. The space needs to do the release work for these block lists.

abandoned_in_gc: Mutex<AbandonedBlockLists>

Abandoned blocks during a GC. Each allocator finishes doing release work, and returns their local blocks to the global lists. Thus we do not need to do release work for these block lists in the space. These lists are only filled in the release phase, and will be moved to the abandoned lists above at the end of a GC.

pending_release_packets: AtomicUsize

Count the number of pending ReleaseMarkSweepSpace and ReleaseMutator work packets during the Release stage.

Implementations

impl<VM: VMBinding> MarkSweepSpace<VM>

pub fn extend_global_side_metadata_specs(_specs: &mut Vec<SideMetadataSpec>)

pub fn new(args: PlanCreateSpaceArgs<'_, VM>) -> MarkSweepSpace<VM>

fn trace_object<Q: ObjectQueue>( &self, queue: &mut Q, object: ObjectReference ) -> ObjectReference

pub fn record_new_block(&self, block: Block)

pub fn prepare(&mut self)

pub fn release(&mut self)

pub fn release_block(&self, block: Block)

Release a block.

pub fn block_clear_metadata(&self, block: Block)

pub fn acquire_block( &self, tls: VMThread, size: usize, align: usize ) -> BlockAcquireResult

pub fn get_abandoned_block_lists(&self) -> &Mutex<AbandonedBlockLists>

pub fn get_abandoned_block_lists_in_gc(&self) -> &Mutex<AbandonedBlockLists>

pub fn release_packet_done(&self)

fn generate_sweep_tasks(&self) -> Vec<Box<dyn GCWork<VM>>>

fn recycle_blocks(&self)