mmtk::policy

Module space

This class defines and manages spaces. Each policy is an instance of a space. A space is a region of virtual memory (contiguous or discontigous) which is subject to the same memory management regime. Multiple spaces (instances of this class or its descendants) may have the same policy (eg there could be numerous instances of CopySpace, each with different roles). Spaces are defined in terms of a unique region of virtual memory, so no two space instances ever share any virtual memory.

In addition to tracking virtual memory use and the mapping to policy, spaces also manage memory consumption (used virtual memory).

Structs

CommonSpace

PlanCreateSpaceArgs

Arguments passed from a plan to create a space.

PolicyCreateSpaceArgs

Arguments passed from a policy to create a space. This includes policy specific args.

Traits

Space

Functions

get_frac_available 🔒

print_vm_map 🔒

Print the VM map for a space. Space needs to be object-safe, so it cannot have methods that use extra generic type paramters. So this method is placed outside the Space trait. This method can be invoked on a &dyn Space (space.as_space() will return &dyn Space).

required_chunks