Trait mmtk::policy::space::SFT

pub trait SFT {
    fn name(&self) -> &str;

    fn is_live(&self, object: ObjectReference) -> bool;

    fn is_movable(&self) -> bool;

    fn is_sane(&self) -> bool;

    fn initialize_object_metadata(&self, object: ObjectReference, alloc: bool);

    fn get_forwarded_object(
        &self, 
        _object: ObjectReference
    ) -> Option<ObjectReference> { ... }

    fn is_reachable(&self, object: ObjectReference) -> bool { ... }
}

Space Function Table (SFT).

This trait captures functions that reflect space-specific per-object semantics. These functions are implemented for each object via a special space-based dynamic dispatch mechanism where the semantics are not determined by the object’s type, but rather, are determined by the space that the object is in.

The underlying mechanism exploits the fact that spaces use the address space at an MMTk chunk granularity with the consequence that each chunk maps to exactluy one space, so knowing the chunk for an object reveals its space. The dispatch then works by performing simple address arithmetic on the object reference to find a chunk index which is used to index a table which returns the space. The relevant function is then dispatched against that space object.

We use the SFT trait to simplify typing for Rust, so our table is a table of SFT rather than Space.

Required methods

fn name(&self) -> &str

The space name

fn is_live(&self, object: ObjectReference) -> bool

Is the object live, determined by the policy?

fn is_movable(&self) -> bool

Is the object movable, determined by the policy? E.g. the policy is non-moving, or the object is pinned.

fn is_sane(&self) -> bool

Is the object sane? A policy should return false if there is any abnormality about object - the sanity checker will fail if an object is not sane.

fn initialize_object_metadata(&self, object: ObjectReference, alloc: bool)

Initialize object metadata (in the header, or in the side metadata).

Loading content...
Provided methods

fn get_forwarded_object(
&self,
_object: ObjectReference
) -> Option<ObjectReference>

Get forwarding pointer if the object is forwarded.

fn is_reachable(&self, object: ObjectReference) -> bool

Is the object reachable, determined by the policy? Note: Objects in ImmortalSpace may have is_live = true but are actually unreachable.

Loading content...