Trait mmtk::scheduler::gc_work::ScanObjectsWork

pub trait ScanObjectsWork<VM: VMBinding>: GCWork<VM> + Sized {
    type E: ProcessEdgesWork<VM = VM>;

    // Required methods
    fn post_scan_object(&self, object: ObjectReference);
    fn make_another(&self, buffer: Vec<ObjectReference>) -> Self;
    fn get_bucket(&self) -> WorkBucketStage;

    // Provided method
    fn do_work_common(
        &self,
        buffer: &[ObjectReference],
        worker: &mut GCWorker<<Self::E as ProcessEdgesWork>::VM>,
        _mmtk: &'static MMTK<<Self::E as ProcessEdgesWork>::VM>
    ) { ... }
}

Trait for a work packet that scans objects

Required Associated Types

type E: ProcessEdgesWork<VM = VM>

The associated ProcessEdgesWork for processing the edges of the objects in this packet.

Required Methods

fn post_scan_object(&self, object: ObjectReference)

Called after each object is scanned.

fn make_another(&self, buffer: Vec<ObjectReference>) -> Self

Create another object-scanning work packet of the same kind, to scan adjacent objects of the objects in this packet.

fn get_bucket(&self) -> WorkBucketStage

Provided Methods

fn do_work_common( &self, buffer: &[ObjectReference], worker: &mut GCWorker<<Self::E as ProcessEdgesWork>::VM>, _mmtk: &'static MMTK<<Self::E as ProcessEdgesWork>::VM> )

The common code for ScanObjects and PlanScanObjects.

Object Safety

This trait is not object safe.