Trait mmtk::scheduler::work::GCWorkContext

pub trait GCWorkContext: Send + 'static {
    type VM: VMBinding;
    type PlanType: Plan<VM = Self::VM>;
    type ProcessEdgesWorkType: ProcessEdgesWork<VM = Self::VM>;
    type TPProcessEdges: ProcessEdgesWork<VM = Self::VM>;
}

This trait provides a group of associated types that are needed to create GC work packets for a certain plan. For example, GCWorkScheduler.schedule_common_work() needs this trait to schedule different work packets. For certain plans, they may need to provide several types that implement this trait, e.g. one for nursery GC, one for mature GC.

Note: Because GCWorkContext is often used as parameters of implementations of GCWork, we let GCWorkContext require Send + 'static. Since GCWorkContext is just a group of associated types, its implementations should not have any actual fields other than PhantomData, and will automatically have Send + 'static.

Required Associated Types

type VM: VMBinding

type PlanType: Plan<VM = Self::VM>

type ProcessEdgesWorkType: ProcessEdgesWork<VM = Self::VM>

type TPProcessEdges: ProcessEdgesWork<VM = Self::VM>