mmtk::scheduler::work_bucket
Struct WorkBucket

pub struct WorkBucket<VM: VMBinding> {
    active: AtomicBool,
    queue: BucketQueue<VM>,
    prioritized_queue: Option<BucketQueue<VM>>,
    monitor: Arc<WorkerMonitor>,
    can_open: Option<Box<dyn Fn(&GCWorkScheduler<VM>) -> bool + Send>>,
    sentinel: Mutex<Option<Box<dyn GCWork<VM>>>>,
}

Fields
active: AtomicBoolqueue: BucketQueue<VM>prioritized_queue: Option<BucketQueue<VM>>monitor: Arc<WorkerMonitor>can_open: Option<Box<dyn Fn(&GCWorkScheduler<VM>) -> bool + Send>>sentinel: Mutex<Option<Box<dyn GCWork<VM>>>>

After this bucket is activated and all pending work packets (including the packets in this bucket) are drained, this work packet, if exists, will be added to this bucket. When this happens, it will prevent opening subsequent work packets.

The sentinel work packet may set another work packet as the new sentinel which will be added to this bucket again after all pending work packets are drained. This may happend again and again, causing the GC to stay at the same stage and drain work packets in a loop.

This is useful for handling weak references that may expand the transitive closure recursively, such as ephemerons and Java-style SoftReference and finalizers. Sentinels can be used repeatedly to discover and process more such objects.

Implementations

impl<VM: VMBinding> WorkBucket<VM>

pub(crate) fn new(active: bool, monitor: Arc<WorkerMonitor>) -> Self

fn notify_one_worker(&self)

pub fn notify_all_workers(&self)

pub fn is_activated(&self) -> bool

pub fn activate(&self)

Enable the bucket

pub fn is_empty(&self) -> bool

Test if the bucket is drained

pub fn is_drained(&self) -> bool

pub fn deactivate(&self)

Disable the bucket

pub fn add_prioritized(&self, work: Box<dyn GCWork<VM>>)

Add a work packet to this bucket Panic if this bucket cannot receive prioritized packets.

pub fn add<W: GCWork<VM>>(&self, work: W)

Add a work packet to this bucket

pub fn add_boxed(&self, work: Box<dyn GCWork<VM>>)

Add a work packet to this bucket

pub(crate) fn add_no_notify<W: GCWork<VM>>(&self, work: W)

Add a work packet to this bucket, but do not notify any workers. This is useful when the current thread is holding the mutex of WorkerMonitor which is used for notifying workers. This usually happens if the current thread is the last worker parked.

pub(crate) fn add_boxed_no_notify(&self, work: Box<dyn GCWork<VM>>)

Like WorkBucket::add_no_notify, but the work is boxed.

pub fn bulk_add_prioritized(&self, work_vec: Vec<Box<dyn GCWork<VM>>>)

Add multiple packets with a higher priority. Panic if this bucket cannot receive prioritized packets.

pub fn bulk_add(&self, work_vec: Vec<Box<dyn GCWork<VM>>>)

Add multiple packets

pub fn poll( &self, worker: &Worker<Box<dyn GCWork<VM>>>, ) -> Steal<Box<dyn GCWork<VM>>>

Get a work packet from this bucket

pub fn set_open_condition( &mut self, pred: impl Fn(&GCWorkScheduler<VM>) -> bool + Send + 'static, )

pub fn set_sentinel(&self, new_sentinel: Box<dyn GCWork<VM>>)

pub fn has_sentinel(&self) -> bool

pub fn update(&self, scheduler: &GCWorkScheduler<VM>) -> bool

pub fn maybe_schedule_sentinel(&self) -> bool