Struct mmtk::util::address::Address

#[repr(transparent)]
pub struct Address(usize);

Address represents an arbitrary address. This is designed to represent address and do address arithmetic mostly in a safe way, and to allow mark some operations as unsafe. This type needs to be zero overhead (memory wise and time wise). The idea is from the paper High-level Low-level Programming (VEE09) and JikesRVM.

Tuple Fields
0: usize
Implementations

impl Address

pub const ZERO: Self = _

The lowest possible address.

pub const MAX: Self = _

The highest possible address.

pub fn from_ptr<T>(ptr: *const T) -> Address

creates Address from a pointer

pub fn from_ref<T>(r: &T) -> Address

creates Address from a Rust reference

pub fn from_mut_ptr<T>(ptr: *mut T) -> Address

creates Address from a mutable pointer

pub const unsafe fn zero() -> Address

creates a null Address (0)

Safety

It is unsafe and the user needs to be aware that they are creating an invalid address. The zero address should only be used as unininitialized or sentinel values in performance critical code (where you dont want to use Option<Address>).

pub unsafe fn max() -> Address

creates an Address of (usize::MAX)

Safety

It is unsafe and the user needs to be aware that they are creating an invalid address. The max address should only be used as unininitialized or sentinel values in performance critical code (where you dont want to use Option<Address>).

pub const unsafe fn from_usize(raw: usize) -> Address

creates an arbitrary Address

Safety

It is unsafe and the user needs to be aware that they may create an invalid address. This creates arbitrary addresses which may not be valid. This should only be used for hard-coded addresses. Any other uses of this function could be replaced with more proper alternatives.

pub fn shift<T>(self, offset: isize) -> Self

shifts the address by N T-typed objects (returns addr + N * size_of(T))

pub const fn get_extent(self, other: Address) -> ByteSize

Get the number of bytes between two addresses. The current address needs to be higher than the other address.

pub const fn get_offset(self, other: Address) -> ByteOffset

Get the offset from other to self. The result is negative is self is lower than other.

pub const fn add(self, size: usize) -> Address

Add an offset to the address.

pub const fn sub(self, size: usize) -> Address

Subtract an offset from the address.

pub const fn and(self, mask: usize) -> usize

Bitwise ‘and’ with a mask.

pub const fn saturating_sub(self, size: usize) -> Address

Perform a saturating subtract on the Address

pub unsafe fn load<T: Copy>(self) -> T

loads a value of type T from the address

Safety

This could throw a segment fault if the address is invalid

pub unsafe fn store<T>(self, value: T)

stores a value of type T to the address

Safety

This could throw a segment fault if the address is invalid

pub unsafe fn atomic_load<T: Atomic>(self, order: Ordering) -> T::Type

atomic operation: load

Safety

This could throw a segment fault if the address is invalid

pub unsafe fn atomic_store<T: Atomic>(self, val: T::Type, order: Ordering)

atomic operation: store

Safety

This could throw a segment fault if the address is invalid

pub unsafe fn compare_exchange<T: Atomic>( self, old: T::Type, new: T::Type, success: Ordering, failure: Ordering ) -> Result<T::Type, T::Type>

atomic operation: compare and exchange usize

Safety

This could throw a segment fault if the address is invalid

pub fn is_zero(self) -> bool

is this address zero?

pub const fn align_up(self, align: ByteSize) -> Address

aligns up the address to the given alignment

pub const fn align_down(self, align: ByteSize) -> Address

aligns down the address to the given alignment

pub const fn is_aligned_to(self, align: usize) -> bool

is this address aligned to the given alignment

pub fn to_ptr<T>(self) -> *const T

converts the Address to a pointer

pub fn to_mut_ptr<T>(self) -> *mut T

converts the Address to a mutable pointer

pub unsafe fn as_ref<'a, T>(self) -> &'a T

converts the Address to a Rust reference

Safety

The caller must guarantee the address actually points to a Rust object.

pub unsafe fn as_mut_ref<'a, T>(self) -> &'a mut T

converts the Address to a mutable Rust reference

Safety

The caller must guarantee the address actually points to a Rust object.

pub const fn as_usize(self) -> usize

converts the Address to a pointer-sized integer

pub fn chunk_index(self) -> usize

returns the chunk index for this address

pub fn is_mapped(self) -> bool

return true if the referenced memory is mapped

pub fn range_intersection( r1: &Range<Address>, r2: &Range<Address> ) -> Range<Address>

Returns the intersection of the two address ranges. The returned range could be empty if there is no intersection between the ranges.