Module mmtk::util::alloc

Allocators Various allocators implementation.

Re-exports

pub use self::immix_allocator::ImmixAllocator;

pub use free_list_allocator::FreeListAllocator;

Modules

allocator 🔒

The allocator trait and allocation-related functions.

allocators 🔒

A list of all the allocators, embedded in Mutator

bumpallocator 🔒

Bump pointer allocator

embedded_meta_data 🔒

Embedded metadata pages

free_list_allocator

immix_allocator

Immix allocator

large_object_allocator 🔒

malloc_allocator 🔒

An alloactor backed by malloc

markcompact_allocator 🔒

Mark compact allocator (actually a bump pointer allocator with an extra heade word)

Structs

BumpAllocator

BumpPointer

A common fast-path bump-pointer allocator shared across different allocator implementations that use bump-pointer allocation. A BumpPointer is always initialized with cursor = 0, limit = 0, so the first allocation always fails the check of cursor + size < limit and goes to the slowpath. A binding can also take advantage of this design to zero-initialize the a bump pointer.

LargeObjectAllocator

MallocAllocator

MarkCompactAllocator

A thin wrapper(specific implementation) of bump allocator reserve extra bytes when allocating

Enums

AllocationError

A list of errors that MMTk can encounter during allocation.

AllocatorInfo

This type describes allocator information. It is used to generate fast paths for the GC. All offset fields are relative to Mutator.

AllocatorSelector

Traits

Allocator

A trait which implements allocation routines. Every allocator needs to implements this trait.

Functions

fill_alignment_gap