mmtk::util::heap::blockpageresource

Struct BlockPool

pub struct BlockPool<B: Region> {
    head_global_freed_blocks: RwLock<Option<BlockQueue<B>>>,
    global_freed_blocks: RwLock<Vec<BlockQueue<B>>>,
    worker_local_freed_blocks: Vec<BlockQueue<B>>,
    count: AtomicUsize,
}

A block queue which contains a global pool and a set of thread-local queues.

Mutator or collector threads always allocate blocks by poping from the global pool。

Collector threads free blocks to their thread-local queues, and then flush to the global pools before GC ends.

Fields
head_global_freed_blocks: RwLock<Option<BlockQueue<B>>>

First global BlockArray for fast allocation

global_freed_blocks: RwLock<Vec<BlockQueue<B>>>

A list of BlockArray that is flushed to the global pool

worker_local_freed_blocks: Vec<BlockQueue<B>>

Thread-local block queues

count: AtomicUsize

Total number of blocks in the whole BlockQueue

Implementations

impl<B: Region> BlockPool<B>

pub fn new(num_workers: usize) -> Self

Create a BlockQueue

fn add_global_array(&self, array: BlockQueue<B>)

Add a BlockArray to the global pool

pub fn push(&self, block: B)

Push a block to the thread-local queue

pub fn pop(&self) -> Option<B>

Pop a block from the global pool

fn flush(&self, id: usize)

Flush a given thread-local queue to the global pool

pub fn flush_all(&self)

Flush all thread-local queues to the global pool

pub fn len(&self) -> usize

Get total number of blocks in the whole BlockQueue

pub fn iterate_blocks(&self, f: &mut impl FnMut(B))

Iterate all the blocks in the BlockQueue