mmtk::util::heap::chunk_map
Struct ChunkMap

pub struct ChunkMap {
    space_index: usize,
    chunk_range: Mutex<Range<Chunk>>,
}

A byte-map to record all the allocated chunks. A plan can use this to maintain records for the chunks that they used, and the states of the chunks. Any plan that uses the chunk map should include the ALLOC_TABLE spec in their global sidemetadata specs.

A chunk map is created for a space (identified by the space index), and will only update or list chunks for that space.

Fields
space_index: usize

The space that uses this chunk map.

chunk_range: Mutex<Range<Chunk>>

The range of chunks that are used by the space. The range only records the lowest chunk and the highest chunk. All the chunks that are used for the space are within the range, but not necessarily that all the chunks in the range are used for the space. Spaces may be discontiguous, thus the range may include chunks that do not belong to the space. We need to use the space index in the chunk map and the space index encoded with the chunk state to know if the chunk belongs to the current space.

Implementations

impl ChunkMap

pub const ALLOC_TABLE: SideMetadataSpec = crate::util::metadata::side_metadata::spec_defs::CHUNK_MARK

Chunk alloc table

pub fn new(space_index: usize) -> Self

pub fn set_allocated(&self, chunk: Chunk, allocated: bool)

Set a chunk as allocated, or as free.

pub fn get(&self, chunk: Chunk) -> Option<ChunkState>

Get chunk state. Return None if the chunk does not belong to the space.

fn get_internal(&self, chunk: Chunk) -> ChunkState

Get chunk state, regardless of the space. This should always be private.

pub fn all_chunks(&self) -> impl Iterator<Item = Chunk> + '_

A range of all allocated chunks by this space in the heap.

pub fn generate_tasks<VM: VMBinding>( &self, func: impl Fn(Chunk) -> Box<dyn GCWork<VM>>, ) -> Vec<Box<dyn GCWork<VM>>>

Helper function to create per-chunk processing work packets for each allocated chunks.