mmtk::util::heap::gc_trigger
Struct GCTrigger

pub struct GCTrigger<VM: VMBinding> {
    plan: MaybeUninit<&'static dyn Plan<VM = VM>>,
    pub policy: Box<dyn GCTriggerPolicy<VM>>,
    request_flag: AtomicBool,
    scheduler: Arc<GCWorkScheduler<VM>>,
    options: Arc<Options>,
    state: Arc<GlobalState>,
}

GCTrigger is responsible for triggering GCs based on the given policy. All the decisions about heap limit and GC triggering should be resolved here. Depending on the actual policy, we may either forward the calls either to the plan or to the binding/runtime.

Fields
plan: MaybeUninit<&'static dyn Plan<VM = VM>>

The current plan. This is uninitialized when we create it, and later initialized once we have a fixed address for the plan.

policy: Box<dyn GCTriggerPolicy<VM>>

The triggering policy.

request_flag: AtomicBool

Set by mutators to trigger GC. It is atomic so that mutators can check if GC has already been requested efficiently in poll without acquiring any mutex.

scheduler: Arc<GCWorkScheduler<VM>>options: Arc<Options>state: Arc<GlobalState>
Implementations

impl<VM: VMBinding> GCTrigger<VM>

pub fn new( options: Arc<Options>, scheduler: Arc<GCWorkScheduler<VM>>, state: Arc<GlobalState>, ) -> Self

pub fn set_plan(&mut self, plan: &'static dyn Plan<VM = VM>)

Set the plan. This is called in create_plan() after we created a boxed plan.

fn plan(&self) -> &dyn Plan<VM = VM>

fn request(&self)

Request a GC. Called by mutators when polling (during allocation) and when handling user GC requests (e.g. System.gc(); in Java).

pub fn clear_request(&self)

Clear the “GC requested” flag so that mutators can trigger the next GC. Called by a GC worker when all mutators have come to a stop.

pub fn poll(&self, space_full: bool, space: Option<&dyn Space<VM>>) -> bool

This method is called periodically by the allocation subsystem (by default, each time a page is consumed), and provides the collector with an opportunity to collect.

Arguments:

space_full: Space request failed, must recover pages within ‘space’.

space: The space that triggered the poll. This could None if the poll is not triggered by a space.

pub fn handle_user_collection_request( &self, force: bool, exhaustive: bool, ) -> bool

This method is called when the user manually requests a collection, such as System.gc() in Java. Returns true if a collection is actually requested.

Arguments

force: If true, we force a collection regardless of the settings. If false, we only trigger a collection if the settings allow it.

exhaustive: If true, we try to make the collection exhaustive (e.g. full heap collection). If false, the collection kind is determined internally.

pub fn trigger_internal_collection_request(&self)

MMTK has requested stop-the-world activity (e.g., stw within a concurrent gc).

pub fn should_do_stress_gc(&self) -> bool

pub(crate) fn should_do_stress_gc_inner( state: &GlobalState, options: &Options, ) -> bool

Check if we should do a stress GC now. If GC is initialized and the allocation bytes exceeds the stress factor, we should do a stress GC.

pub fn is_heap_full(&self) -> bool

Check if the heap is full

pub fn get_max_nursery_bytes(&self) -> usize

Return upper bound of the nursery size (in number of bytes)

pub fn get_min_nursery_bytes(&self) -> usize

Return lower bound of the nursery size (in number of bytes)

pub fn get_max_nursery_pages(&self) -> usize

Return upper bound of the nursery size (in number of pages)

pub fn get_min_nursery_pages(&self) -> usize

Return lower bound of the nursery size (in number of pages)