Struct mmtk::util::heap::gc_trigger::MemBalancerTrigger

pub struct MemBalancerTrigger {
    min_heap_pages: usize,
    max_heap_pages: usize,
    current_heap_pages: AtomicUsize,
    pending_pages: AtomicUsize,
    stats: AtomicRefCell<MemBalancerStats>,
}

An implementation of MemBalancer (Optimal heap limits for reducing browser memory use, https://dl.acm.org/doi/10.1145/3563323) We use MemBalancer to decide a heap limit between the min heap and the max heap. The current implementation is a simplified version of mem balancer and it does not take collection/allocation speed into account, and uses a fixed constant instead.

Fields
min_heap_pages: usize

The min heap size

max_heap_pages: usize

The max heap size

current_heap_pages: AtomicUsize

The current heap size

pending_pages: AtomicUsize

The number of pending allocation pages. The allocation requests for them have failed, and a GC is triggered. We will need to take them into consideration so that the new heap size can accomodate those allocations.

stats: AtomicRefCell<MemBalancerStats>

Statistics

Implementations

impl MemBalancerTrigger

fn new(min_heap_pages: usize, max_heap_pages: usize) -> Self

fn access_stats<F>(&self, f: F)
where F: FnMut(&mut MemBalancerStats),

fn compute_new_heap_limit( &self, live: usize, extra_reserve: usize, stats: &mut MemBalancerStats )