mmtk::util::heap

Module layout

Re-exports

pub use self::mmapper::Mmapper;

pub use self::map::VMMap;

Modules

heap_parameters

map 🔒

map32 🔒

map64 🔒

mmapper 🔒

vm_layout

The module defines virutal memory layout parameters.

Functions

available_range

The avialable heap range between AVAILABLE_START and AVAILABLE_END. Available range is what MMTk may allocate into.

create_mmapper

create_vm_map

heap_range

The heap range between HEAP_START and HEAP_END Heap range include the availble range, but may include some address ranges that we count as part of the heap but we do not allocate into, such as VM spaces. However, currently, heap range is the same as available range.