Struct mmtk::util::heap::layout::map32::Map32

pub struct Map32 {
    sync: Mutex<()>,
    inner: UnsafeCell<Map32Inner>,
}

Fields
sync: Mutex<()>inner: UnsafeCell<Map32Inner>
Implementations

impl Map32

pub fn new() -> Self

impl Map32

unsafe fn mut_self(&self) -> &mut Map32Inner

Safety

The caller needs to guarantee there is no race condition. Either only one single thread is using this method, or multiple threads are accessing mutally exclusive data (e.g. different indices in arrays). In other cases, use mut_self_with_sync().

fn mut_self_with_sync(&self) -> (MutexGuard<'_, ()>, &mut Map32Inner)

fn free_contiguous_chunks_no_lock(&self, chunk: i32) -> usize

fn get_discontig_freelist_pr_ordinal(&self) -> usize