mmtk::util::heap::layout::map64
Struct Map64

pub struct Map64 {
    inner: UnsafeCell<Map64Inner>,
}

Fields
inner: UnsafeCell<Map64Inner>
Implementations

impl Map64

pub fn new() -> Self

impl Map64

unsafe fn mut_self(&self) -> &mut Map64Inner

Safety

The caller needs to guarantee there is no race condition. Either only one single thread is using this method, or multiple threads are accessing mutally exclusive data (e.g. different indices in arrays). In other cases, use mut_self_with_sync().

fn inner(&self) -> &Map64Inner

fn space_index(addr: Address) -> Option<usize>

fn is_space_start(base: Address) -> bool