Trait mmtk::util::heap::pageresource::PageResource

pub trait PageResource<VM: VMBinding>: 'static {

    fn alloc_pages(
        &self, 
        space_descriptor: SpaceDescriptor, 
        reserved_pages: usize, 
        required_pages: usize, 
        tls: VMThread
    ) -> Result<PRAllocResult, PRAllocFail>;

    fn adjust_for_metadata(&self, pages: usize) -> usize;

    fn common(&self) -> &CommonPageResource;

    fn common_mut(&mut self) -> &mut CommonPageResource;

    fn get_new_pages(
        &self, 
        space_descriptor: SpaceDescriptor, 
        reserved_pages: usize, 
        required_pages: usize, 
        tls: VMThread
    ) -> Result<PRAllocResult, PRAllocFail> { ... }

    fn reserve_pages(&self, pages: usize) -> usize { ... }

    fn clear_request(&self, reserved_pages: usize) { ... }

    fn update_zeroing_approach(&self, _nontemporal: bool, concurrent: bool) { ... }

    fn skip_concurrent_zeroing(&self) { ... }

    fn trigger_concurrent_zeroing(&self) { ... }

    fn concurrent_zeroing(&self) { ... }

    fn commit_pages(
        &self, 
        reserved_pages: usize, 
        actual_pages: usize, 
        tls: VMThread
    ) { ... }

    fn reserved_pages(&self) -> usize { ... }

    fn committed_pages(&self) -> usize { ... }

    fn vm_map(&self) -> &'static Map64 { ... }

}

Required methods

fn alloc_pages(
&self,
space_descriptor: SpaceDescriptor,
reserved_pages: usize,
required_pages: usize,
tls: VMThread
) -> Result<PRAllocResult, PRAllocFail>

fn adjust_for_metadata(&self, pages: usize) -> usize

fn common(&self) -> &CommonPageResource

fn common_mut(&mut self) -> &mut CommonPageResource

Provided methods

fn get_new_pages(
&self,
space_descriptor: SpaceDescriptor,
reserved_pages: usize,
required_pages: usize,
tls: VMThread
) -> Result<PRAllocResult, PRAllocFail>

Allocate pages from this resource. Simply bump the cursor, and fail if we hit the sentinel. Return The start of the first page if successful, zero on failure.

fn reserve_pages(&self, pages: usize) -> usize

fn clear_request(&self, reserved_pages: usize)

fn update_zeroing_approach(&self, _nontemporal: bool, concurrent: bool)

fn skip_concurrent_zeroing(&self)

fn trigger_concurrent_zeroing(&self)

fn concurrent_zeroing(&self)

fn commit_pages(
&self,
reserved_pages: usize,
actual_pages: usize,
tls: VMThread
)

Commit pages to the page budget. This is called after successfully determining that the request can be satisfied by both the page budget and virtual memory. This simply accounts for the discrepancy between committed and reserved while the request was pending.

This MUST be called by each PageResource during the allocPages, and the caller must hold the lock.

fn reserved_pages(&self) -> usize

fn committed_pages(&self) -> usize

fn vm_map(&self) -> &'static Map64