mmtk::util::heap
Module vm_layout

The module defines virutal memory layout parameters.

Structs

VMLayout

Runtime-initialized virtual memory constants

Constants

BYTES_IN_CHUNK

Coarsest unit of address space allocation.

CHUNK_MASK

Mask for chunk size.

LOG_BYTES_IN_CHUNK

log_2 of the coarsest unit of address space allocation.

PAGES_IN_CHUNK

Coarsest unit of address space allocation, in pages

Functions

vm_layout

Get the current virtual memory layout in use. If the binding would like to set a custom virtual memory layout (crate::mmtk::MMTKBuilder::set_vm_layout), they should not call this function before they set a custom layout.