Trait mmtk::util::linear_scan::Region

pub trait Region: Copy + PartialEq + PartialOrd {
    const LOG_BYTES: usize;
    const BYTES: usize = _;

    // Required methods
    fn from_aligned_address(address: Address) -> Self;
    fn start(&self) -> Address;

    // Provided methods
    fn from_unaligned_address(address: Address) -> Self { ... }
    fn align(address: Address) -> Address { ... }
    fn is_aligned(address: Address) -> bool { ... }
    fn end(&self) -> Address { ... }
    fn next(&self) -> Self { ... }
    fn next_nth(&self, n: usize) -> Self { ... }
    fn containing(object: ObjectReference) -> Self { ... }
    fn includes_address(&self, addr: Address) -> bool { ... }
}

Region represents a memory region with a properly aligned address as its start and a fixed size for the region. Region provides a set of utility methods, along with a RegionIterator that linearly scans at the step of a region.

Required Associated Constants

const LOG_BYTES: usize

log2 of the size in bytes for the region.

Provided Associated Constants

const BYTES: usize = _

The size in bytes for the region.

Required Methods

fn from_aligned_address(address: Address) -> Self

Create a region from an address that is aligned to the region boundary. The method should panic if the address is not properly aligned to the region. For performance, this method should always be inlined.

fn start(&self) -> Address

Return the start address of the region. For performance, this method should always be inlined.

Provided Methods

fn from_unaligned_address(address: Address) -> Self

Create a region from an arbitrary address.

fn align(address: Address) -> Address

Align the address to the region.

fn is_aligned(address: Address) -> bool

Check if an address is aligned to the region.

fn end(&self) -> Address

Return the end address of the region. Note that the end address is not in the region.

fn next(&self) -> Self

Return the next region after this one.

fn next_nth(&self, n: usize) -> Self

Return the next nth region after this one.

fn containing(object: ObjectReference) -> Self

Return the region that contains the object.

fn includes_address(&self, addr: Address) -> bool

Check if the given address is in the region.

Object Safety

This trait is not object safe.