Function mmtk::util::metadata::store_metadata

pub fn store_metadata<VM: VMBinding>(
    metadata_spec: &MetadataSpec, 
    object: ObjectReference, 
    val: usize, 
    mask: Option<usize>, 
    atomic_ordering: Option<Ordering>
)

A function to store a value to the specified metadata.

Arguments:

metadata_spec: is one of the const MetadataSpec instances from the ObjectModel trait, for the target metadata. Whether the metadata is in-header or on-side is a VM-specific choice.

object: is a reference to the target object.

val: is the new metadata value to be stored.

mask: is an optional mask value for the metadata. This value is used in cases like the forwarding pointer metadata, where some of the bits are reused by other metadata such as the forwarding bits.

atomic_ordering: is an optional atomic ordering for the store operation. An input value of None means the store operation is not atomic, and an input value of Some(Ordering::X) means the atomic store operation will use the Ordering::X.