mmtk::util::metadata
Module header_metadata

This module provides a default implementation of the access functions for in-header metadata.

Structs

HeaderMetadataSpec

This struct stores the specification of a header metadata bit-set. It supports either bits metadata of 1-7 bits in the same byte, or u8/u16/u32/u64 at an offset of their natural alignment.

Constants

BITS_IN_U16 🔒

BITS_IN_U32 🔒

BITS_IN_U64 🔒

LOG_BITS_IN_U16 🔒

LOG_BITS_IN_U32 🔒

LOG_BITS_IN_U64 🔒