Module mmtk::util::metadata::side_metadata::constants

Constants

GLOBAL_SIDE_METADATA_BASE_ADDRESS

Global side metadata start address

GLOBAL_SIDE_METADATA_BASE_OFFSET 🔒

GLOBAL_SIDE_METADATA_VM_BASE_ADDRESS

The base address for the global side metadata space available to VM bindings, to be used for the per-object metadata. VM bindings must use this to avoid overlap with core internal global side metadata.

GLOBAL_SIDE_METADATA_VM_BASE_OFFSET

The base offset for the global side metadata available to VM bindings.

LOCAL_SIDE_METADATA_BASE_ADDRESS 🔒

LOCAL_SIDE_METADATA_BASE_OFFSET 🔒

LOCAL_SIDE_METADATA_VM_BASE_OFFSET

The base address for the local side metadata space available to VM bindings, to be used for the per-object metadata. VM bindings must use this to avoid overlap with core internal local side metadata.

LOG_GLOBAL_SIDE_METADATA_WORST_CASE_RATIO 🔒

LOG_LOCAL_SIDE_METADATA_WORST_CASE_RATIO 🔒

LOG_MAX_GLOBAL_SIDE_METADATA_SIZE 🔒

The max bytes (in log2) that may be used for global side metadata.

VO_BIT_SIDE_METADATA_ADDR

Base address of VO bit, public to VM bindings which may need to use this.