mmtk::util
Module options

MMTk command line options.

Macros

options 🔒

Structs

MMTKOption

An MMTk option of a given type. This type allows us to store some metadata for the option. To get the value of an option, you can simply dereference it (for example, *options.threads).

Options

Options for an MMTk instance. It affects many aspects of the behavior of the MMTk instance, including the number of GC worker threads, the GC plan to use, etc.

PerfEventOptions

MMTk option for perf events

Enums

AffinityKind

AffinityKind describes how to set the affinity of GC threads. Note that we currently assume that each GC thread is equivalent to an OS or hardware thread.

GCTriggerSelector

Select a GC trigger for MMTk.

NurserySize

An option that provides a min/max interface to MMTk and a Bounded/Fixed interface to the user/VM.

NurseryZeroingOptions

The zeroing approach to use for new object allocations. Affects each plan differently.

PlanSelector

Select a GC plan for MMTk.

SetOptionByStringError 🔒

Error when setting an option by option name and option value as strings.

Constants

DEFAULT_MAX_NURSERY

The default max nursery size. This does not affect the actual space we create as nursery. It is only used in the GC trigger check.

DEFAULT_MAX_NURSERY_32

The default max nursery size for 32 bits.

DEFAULT_MIN_NURSERY

The default min nursery size. This does not affect the actual space we create as nursery. It is only used in the GC trigger check.

DEFAULT_PROPORTIONAL_MAX_NURSERY

The default max nursery size proportional to the current heap size

DEFAULT_PROPORTIONAL_MIN_NURSERY

The default min nursery size proportional to the current heap size

DEFAULT_STRESS_FACTOR

The default stress factor. This is set to the max usize, which means we will never trigger a stress GC for the default value.

Functions

always_valid 🔒