Struct mmtk::util::side_metadata::global::SideMetadata

pub struct SideMetadata {
    context: SideMetadataContext,
    accounting: PageAccounting,
}

Fields
context: SideMetadataContextaccounting: PageAccounting
Implementations

impl SideMetadata

pub fn new(context: SideMetadataContext) -> SideMetadata

pub fn reserved_pages(&self) -> usize

pub fn try_map_metadata_space(&self, start: Address, size: usize) -> Result<()>

Tries to map the required metadata space and returns true is successful. This can be called at page granularity.

pub fn try_map_metadata_address_range(
&self,
start: Address,
size: usize
) -> Result<()>

Tries to map the required metadata address range, without reserving swap-space/physical memory for it. This will make sure the address range is exclusive to the caller. This should be called at chunk granularity.

NOTE: Accessing addresses in this range will produce a segmentation fault if swap-space is not mapped using the try_map_metadata_space function.

pub(in util::side_metadata::global) fn map_metadata_internal(
&self,
start: Address,
size: usize,
no_reserve: bool
) -> Result<()>

The internal function to mmap metadata

Arguments

start - The starting address of the source data.

size - The size of the source data (in bytes).

no_reserve - whether to invoke mmap with a noreserve flag (we use this flag to quanrantine address range)

pub fn ensure_unmap_metadata_space(&self, start: Address, size: usize)

Unmap the corresponding metadata space or panic.

Note-1: This function is only used for test and debug right now.

Note-2: This function uses munmap() which works at page granularity. If the corresponding metadata space's size is not a multiple of page size, the actual unmapped space will be bigger than what you specify.