Trait mmtk::vm::active_plan::ActivePlan

pub trait ActivePlan<VM: VMBinding> {
    fn global() -> &'static dyn Plan<VM = VM>;

    fn is_mutator(tls: VMThread) -> bool;

    fn mutator(tls: VMMutatorThread) -> &'static mut Mutator<VM>;

    fn reset_mutator_iterator();

    fn get_next_mutator() -> Option<&'static mut Mutator<VM>>;

    fn number_of_mutators() -> usize;

    fn mutators<'a>() -> SynchronizedMutatorIterator<'a, VM> { ... }
}

VM-specific methods for the current plan.

Required methods

fn global() -> &'static dyn Plan<VM = VM>

Return a reference to the current plan.

fn is_mutator(tls: VMThread) -> bool

Return whether there is a mutator created and associated with the thread.

Arguments:

tls: The thread to query.

Safety

The caller needs to make sure that the thread is valid (a value passed in by the VM binding through API).

fn mutator(tls: VMMutatorThread) -> &'static mut Mutator<VM>

Return a Mutator reference for the thread.

Arguments:

tls: The thread to query.

Safety

The caller needs to make sure that the thread is a mutator thread.

fn reset_mutator_iterator()

Reset the mutator iterator so that get_next_mutator() returns the first mutator.

fn get_next_mutator() -> Option<&'static mut Mutator<VM>>

Return the next mutator if there is any. This method assumes that the VM implements stateful type to remember which mutator is returned and guarantees to return the next when called again. This does not need to be thread safe.

fn number_of_mutators() -> usize

Return the total count of mutators.

Provided methods

fn mutators<'a>() -> SynchronizedMutatorIterator<'a, VM>

A utility method to provide a thread-safe mutator iterator from reset_mutator_iterator() and get_next_mutator().