Module mmtk::vm::edge_shape

Allows MMTk to access edges in a VM-defined way.

Structs

AddressRangeIterator

Iterate edges within Range<Address>.

SimpleEdge

A simple edge implementation that represents a word-sized slot which holds the raw address of an ObjectReference, or 0 if it is holding a null reference.

UnimplementedMemorySlice

Memory slice type with empty implementations. For VMs that do not use the memory slice type.

UnimplementedMemorySliceEdgeIterator

Edge iterator for UnimplementedMemorySlice.

Traits

Edge

An Edge represents a slot in an object (a.k.a. a field), on the stack (i.e. a local variable) or any other places (such as global variables). A slot may hold an object reference. We can load the object reference from it, and we can store an ObjectReference into it. For some VMs, a slot may sometimes not hold an object reference. For example, it can hold a special NULL pointer which does not point to any object, or it can hold a tagged non-reference value, such as small integers and special values such as true, false, null (a.k.a. “none”, “nil”, etc. for other VMs), undefined, etc.

MemorySlice

A abstract memory slice represents a piece of heap memory.