Trait mmtk::vm::ActivePlan

pub trait ActivePlan<VM: VMBinding> {
    pub fn global() -> &'static dyn Plan<VM = VM>;

    pub unsafe fn worker(tls: OpaquePointer) -> &'static mut GCWorker<VM>;

    pub unsafe fn is_mutator(tls: OpaquePointer) -> bool;

    pub unsafe fn mutator(tls: OpaquePointer) -> &'static mut Mutator<VM>;

    pub fn reset_mutator_iterator();

    pub fn get_next_mutator() -> Option<&'static mut Mutator<VM>>;

    pub fn number_of_mutators() -> usize;

    pub fn mutators<'a>() -> SynchronizedMutatorIterator<'a, VM> { ... }
}

VM-specific methods for the current plan.

Required methods

pub fn global() -> &'static dyn Plan<VM = VM>

Return a reference to the current plan.

pub unsafe fn worker(tls: OpaquePointer) -> &'static mut GCWorker<VM>

Return a GCWorker reference for the thread.

Arguments:

tls: The thread to query.

Safety

The caller needs to make sure that the thread is a GC worker thread.

pub unsafe fn is_mutator(tls: OpaquePointer) -> bool

Return whether there is a mutator created and associated with the thread.

Arguments:

tls: The thread to query.

Safety

The caller needs to make sure that the thread is valid (a value passed in by the VM binding through API).

pub unsafe fn mutator(tls: OpaquePointer) -> &'static mut Mutator<VM>

Return a Mutator reference for the thread.

Arguments:

tls: The thread to query.

Safety

The caller needs to make sure that the thread is a mutator thread.

pub fn reset_mutator_iterator()

Reset the mutator iterator so that get_next_mutator() returns the first mutator.

pub fn get_next_mutator() -> Option<&'static mut Mutator<VM>>

Return the next mutator if there is any. This method assumes that the VM implements stateful type to remember which mutator is returned and guarantees to return the next when called again. This does not need to be thread safe.

pub fn number_of_mutators() -> usize

Return the total count of mutators.

Loading content...
Provided methods

pub fn mutators<'a>() -> SynchronizedMutatorIterator<'a, VM>

A utility method to provide a thread-safe mutator iterator from reset_mutator_iterator() and get_next_mutator().

Loading content...