Trait mmtk::vm::Scanning

pub trait Scanning<VM: VMBinding> {
    pub const SCAN_MUTATORS_IN_SAFEPOINT: bool;
    pub const SINGLE_THREAD_MUTATOR_SCANNING: bool;

    pub fn scan_object<T: TransitiveClosure>(
        trace: &mut T, 
        object: ObjectReference, 
        tls: VMWorkerThread
    );

    pub fn notify_initial_thread_scan_complete(
        partial_scan: bool, 
        tls: VMWorkerThread
    );

    pub fn scan_objects<W: ProcessEdgesWork<VM = VM>>(
        objects: &[ObjectReference], 
        worker: &mut GCWorker<VM>
    );

    pub fn scan_thread_roots<W: ProcessEdgesWork<VM = VM>>();

    pub fn scan_thread_root<W: ProcessEdgesWork<VM = VM>>(
        mutator: &'static mut Mutator<VM>, 
        tls: VMWorkerThread
    );

    pub fn scan_vm_specific_roots<W: ProcessEdgesWork<VM = VM>>();

    pub fn supports_return_barrier() -> bool;
}

VM-specific methods for scanning roots/objects.

Associated Constants

pub const SCAN_MUTATORS_IN_SAFEPOINT: bool

Scan stack roots after all mutators are paused.

pub const SINGLE_THREAD_MUTATOR_SCANNING: bool

Scan all the mutators within a single work packet.

SCAN_MUTATORS_IN_SAFEPOINT should also be enabled

Loading content...
Required methods

pub fn scan_object<T: TransitiveClosure>(
trace: &mut T,
object: ObjectReference,
tls: VMWorkerThread
)

Delegated scanning of a object, processing each pointer field encountered. This method probably will be removed in the future, in favor of bulk scanning scan_objects.

Arguments:

trace: The TransitiveClosure to use for scanning.

object: The object to be scanned.

tls: The GC worker thread that is doing this tracing.

pub fn notify_initial_thread_scan_complete(
partial_scan: bool,
tls: VMWorkerThread
)

MMTk calls this method at the first time during a collection that thread's stacks have been scanned. This can be used (for example) to clean up obsolete compiled methods that are no longer being executed.

Arguments:

partial_scan: Whether the scan was partial or full-heap.

tls: The GC thread that is performing the thread scan.

pub fn scan_objects<W: ProcessEdgesWork<VM = VM>>(
objects: &[ObjectReference],
worker: &mut GCWorker<VM>
)

Bulk scanning of objects, processing each pointer field for each object.

Arguments:

objects: The slice of object references to be scanned.

pub fn scan_thread_roots<W: ProcessEdgesWork<VM = VM>>()

Scan all the mutators for roots.

pub fn scan_thread_root<W: ProcessEdgesWork<VM = VM>>(
mutator: &'static mut Mutator<VM>,
tls: VMWorkerThread
)

Scan one mutator for roots.

Arguments:

mutator: The reference to the mutator whose roots will be scanned.

tls: The GC thread that is performing this scanning.

pub fn scan_vm_specific_roots<W: ProcessEdgesWork<VM = VM>>()

Scan VM-specific roots. The creation of all root scan tasks (except thread scanning) goes here.

pub fn supports_return_barrier() -> bool

Return whether the VM supports return barriers. This is unused at the moment.

Loading content...