Trait mmtk::vm::VMBinding

pub trait VMBinding where
    Self: Sized + 'static + Send + Sync + Default, {
    type VMObjectModel: ObjectModel<Self>;
    type VMScanning: Scanning<Self>;
    type VMCollection: Collection<Self>;
    type VMActivePlan: ActivePlan<Self>;
    type VMReferenceGlue: ReferenceGlue<Self>;

    pub const ALIGNMENT_VALUE: usize;
    pub const LOG_MIN_ALIGNMENT: usize;
    pub const MIN_ALIGNMENT: usize;
    pub const MAX_ALIGNMENT_SHIFT: usize;
    pub const MAX_ALIGNMENT: usize;
    pub const ALLOC_END_ALIGNMENT: usize;
}

The VMBinding trait associates with each trait, and provides VM-specific constants.

Associated Types

type VMObjectModel: ObjectModel<Self>

type VMScanning: Scanning<Self>

type VMCollection: Collection<Self>

type VMActivePlan: ActivePlan<Self>

type VMReferenceGlue: ReferenceGlue<Self>

Loading content...
Associated Constants

pub const ALIGNMENT_VALUE: usize

A value to fill in alignment gaps. This value can be used for debugging.

pub const LOG_MIN_ALIGNMENT: usize

Allowed minimal alignment.

pub const MIN_ALIGNMENT: usize

Allowed minimal alignment in bytes.

pub const MAX_ALIGNMENT_SHIFT: usize

Allowed maximum alignment as shift by min alignment.

pub const MAX_ALIGNMENT: usize

Allowed maximum alignment in bytes.

pub const ALLOC_END_ALIGNMENT: usize

This value is used to assert if the cursor is reasonable after allocations. At the end of an allocation, the allocation cursor should be aligned to this value. Note that MMTk does not attempt to do anything to align the cursor to this value, but it merely asserts with this constant.

Loading content...